Network.AWS.ELB.AddTags

Description

Adds one or more tags for the specified load balancer. Each load balancer can have a maximum of 10 tags. Each tag consists of a key and an optional value.

Tag keys must be unique for each load balancer. If a tag with the same key is already associated with the load balancer, this action will update the value of the key.

For more information, see Tagging in the Elastic Load Balancing DeveloperGuide.

http://docs.aws.amazon.com/ElasticLoadBalancing/latest/APIReference/API_AddTags.html

Request

data AddTags

Request constructor

addTags

Arguments

:: NonEmpty Tag
atTags

-> AddTags

AddTags constructor.

The fields accessible through corresponding lenses are:

atLoadBalancerNames :: [Text]

atTags :: NonEmpty Tag

Request lenses

atLoadBalancerNames :: Lens' AddTags [Text]

The name of the load balancer to tag. You can specify a maximum of one load balancer name.

atTags :: Lens' AddTags (NonEmpty Tag)

A list of tags for each load balancer.

Response

data AddTagsResponse

Response constructor

addTagsResponse :: AddTagsResponse

AddTagsResponse constructor.