Network.AWS.ELB.ApplySecurityGroupsToLoadBalancer

Description

Associates one or more security groups with your load balancer in Amazon Virtual Private Cloud (Amazon VPC). The provided security group IDs will override any currently applied security groups. For more information, see Manage Security Groups in Amazon VPC in the Elastic Load Balancing Developer Guide. https:/elasticloadbalancing.amazonaws.com?SecurityGroups.member.1=sg-123456789 &LoadBalancerName=my-test-vpc-loadbalancer &Version=2012-06-01 &Action=ApplySecurityGroupsToLoadBalancer &AUTHPARAMS sg-123456789 06b5decc-102a-11e3-9ad6-bf3e4EXAMPLE.

Request

data ApplySecurityGroupsToLoadBalancer

The input for the ApplySecurityGroupsToLoadBalancer action.

Request constructor

applySecurityGroupsToLoadBalancer

Arguments

:: Text
asgtlbLoadBalancerName

-> [Text]
asgtlbSecurityGroups

-> ApplySecurityGroupsToLoadBalancer

Smart constructor for the minimum required parameters to construct a valid ApplySecurityGroupsToLoadBalancer request.

The fields accessible through corresponding lenses are:

LoadBalancerName :: Text

SecurityGroups :: [Text]

Request lenses

asgtlbLoadBalancerName :: Lens' ApplySecurityGroupsToLoadBalancer Text

The name associated with the load balancer. The name must be unique within the set of load balancers associated with your AWS account.

asgtlbSecurityGroups :: Lens' ApplySecurityGroupsToLoadBalancer [Text]

A list of security group IDs to associate with your load balancer in VPC. The security group IDs must be provided as the ID and not the security group name (For example, sg-1234).

Response

data ApplySecurityGroupsToLoadBalancerResponse

The out for the ApplySecurityGroupsToLoadBalancer action.

Response constructor

applySecurityGroupsToLoadBalancerResponse :: ApplySecurityGroupsToLoadBalancerResponse

Smart constructor for the minimum required parameters to construct a valid ApplySecurityGroupsToLoadBalancerResponse response.

This constructor is provided for convenience and testing purposes.

The fields accessible through corresponding lenses are:

SecurityGroups :: [Text]

Response lenses

asgtlbrSecurityGroups :: Lens' ApplySecurityGroupsToLoadBalancerResponse [Text]

A list of security group IDs associated with your load balancer.