Network.AWS.ELB.AttachLoadBalancerToSubnets

Description

Adds one or more subnets to the set of configured subnets in the Amazon Virtual Private Cloud (Amazon VPC) for the load balancer. The load balancers evenly distribute requests across all of the registered subnets. For more information, see Deploy Elastic Load Balancing in Amazon VPC in the Elastic Load Balancing Developer Guide. https:/elasticloadbalancing.amazonaws.com?Subnets.member.1=subnet-3561b05e &LoadBalancerName=my-test-vpc-loadbalancer &Version=2012-06-01 &Action=AttachLoadBalancerToSubnets &AUTHPARAMS subnet-119f0078 subnet-3561b05e 07b1ecbc-1100-11e3-acaf-dd7edEXAMPLE.

Request

data AttachLoadBalancerToSubnets

The input for the AttachLoadBalancerToSubnets action.

Request constructor

mkAttachLoadBalancerToSubnets

Arguments

:: Text
albtsLoadBalancerName

-> [Text]
albtsSubnets

-> AttachLoadBalancerToSubnets

Smart constructor for the minimum required parameters to construct a valid AttachLoadBalancerToSubnets request.

The fields accessible through corresponding lenses are:

LoadBalancerName :: Text

Subnets :: [Text]

Request lenses

albtsLoadBalancerName :: Lens' AttachLoadBalancerToSubnets Text

The name associated with the load balancer. The name must be unique within the set of load balancers associated with your AWS account.

albtsSubnets :: Lens' AttachLoadBalancerToSubnets [Text]

A list of subnet IDs to add for the load balancer. You can add only one subnet per Availability Zone.

Response

data AttachLoadBalancerToSubnetsResponse

The output for the AttachLoadBalancerToSubnets action.

Response constructor

mkAttachLoadBalancerToSubnetsResponse :: AttachLoadBalancerToSubnetsResponse

Smart constructor for the minimum required parameters to construct a valid AttachLoadBalancerToSubnetsResponse response.

This constructor is provided for convenience and testing purposes.

The fields accessible through corresponding lenses are:

Subnets :: [Text]

Response lenses

albtsrSubnets :: Lens' AttachLoadBalancerToSubnetsResponse [Text]

A list of subnet IDs attached to the load balancer.