Network.AWS.ELB.ConfigureHealthCheck

Description

Specifies the health check settings to use for evaluating the health state of your back-end instances.

For more information, see Health Check in the Elastic Load BalancingDeveloper Guide.

http://docs.aws.amazon.com/ElasticLoadBalancing/latest/APIReference/API_ConfigureHealthCheck.html

Request

data ConfigureHealthCheck

Request constructor

configureHealthCheck

Arguments

:: Text
chcLoadBalancerName

-> HealthCheck
chcHealthCheck

-> ConfigureHealthCheck

ConfigureHealthCheck constructor.

The fields accessible through corresponding lenses are:

chcHealthCheck :: HealthCheck

chcLoadBalancerName :: Text

Request lenses

chcHealthCheck :: Lens' ConfigureHealthCheck HealthCheck

A structure containing the configuration information for the new healthcheck.

chcLoadBalancerName :: Lens' ConfigureHealthCheck Text

The mnemonic name associated with the load balancer. The name must be unique within the set of load balancers associated with your AWS account.

Response

data ConfigureHealthCheckResponse

Response constructor

configureHealthCheckResponse :: ConfigureHealthCheckResponse

ConfigureHealthCheckResponse constructor.

The fields accessible through corresponding lenses are:

chcrHealthCheck :: Maybe HealthCheck

Response lenses

chcrHealthCheck :: Lens' ConfigureHealthCheckResponse (Maybe HealthCheck)

The updated healthcheck for the instances.