Network.AWS.ELB.CreateLBCookieStickinessPolicy

Description

Generates a stickiness policy with sticky session lifetimes controlled by the lifetime of the browser (user-agent) or a specified expiration period. This policy can be associated only with HTTP/HTTPS listeners.

When a load balancer implements this policy, the load balancer uses a special cookie to track the backend server instance for each request. When the load balancer receives a request, it first checks to see if this cookie is present in the request. If so, the load balancer sends the request to the application server specified in the cookie. If not, the load balancer sends the request to a server that is chosen based on the existing load balancing algorithm.

A cookie is inserted into the response for binding subsequent requests from the same user to that server. The validity of the cookie is based on the cookie expiration time, which is specified in the policy configuration.

For more information, see Enabling Duration-Based Session Stickiness in the Elastic Load Balancing Developer Guide.

http://docs.aws.amazon.com/ElasticLoadBalancing/latest/APIReference/API_CreateLBCookieStickinessPolicy.html

Request

data CreateLBCookieStickinessPolicy

Request constructor

createLBCookieStickinessPolicy

Arguments

:: Text
clbcspLoadBalancerName

-> Text
clbcspPolicyName

-> CreateLBCookieStickinessPolicy

CreateLBCookieStickinessPolicy constructor.

The fields accessible through corresponding lenses are:

clbcspCookieExpirationPeriod :: Maybe Integer

clbcspLoadBalancerName :: Text

clbcspPolicyName :: Text

Request lenses

clbcspCookieExpirationPeriod :: Lens' CreateLBCookieStickinessPolicy (Maybe Integer)

The time period in seconds after which the cookie should be considered stale. Not specifying this parameter indicates that the sticky session will last for the duration of the browser session.

clbcspLoadBalancerName :: Lens' CreateLBCookieStickinessPolicy Text

The name associated with the load balancer.

clbcspPolicyName :: Lens' CreateLBCookieStickinessPolicy Text

The name of the policy being created. The name must be unique within the set of policies for this load balancer.

Response

data CreateLBCookieStickinessPolicyResponse

Response constructor

createLBCookieStickinessPolicyResponse :: CreateLBCookieStickinessPolicyResponse

CreateLBCookieStickinessPolicyResponse constructor.