Network.AWS.ELB.CreateLoadBalancerListeners

Description

Creates one or more listeners on a load balancer for the specified port. If a listener with the given port does not already exist, it will be created; otherwise, the properties of the new listener must match the properties of the existing listener. For more information, see Add a Listener to Your Load Balancer in the Elastic Load Balancing Developer Guide. Create an HTTPS Load Balancer listener in EC2-Classic https:/elasticloadbalancing.amazonaws.com?Listeners.member.1.Protocol=https &Listeners.member.1.LoadBalancerPort=443 &Listeners.member.1.InstancePort=443 &Listeners.member.1.InstanceProtocol=https &Listeners.member.1.SSLCertificateId=arn:aws:iam::123456789012:server-certificate/servercert &LoadBalancerName=MyHTTPSLoadBalancer &Version=2012-06-01 &Action=CreateLoadBalancerListeners &AUTHPARAMS 1549581b-12b7-11e3-895e-1334aEXAMPLE.

Request

data CreateLoadBalancerListeners

The input for the CreateLoadBalancerListeners action.

Request constructor

createLoadBalancerListeners

Arguments

:: Text
clblLoadBalancerName

-> [Listener]
clblListeners

-> CreateLoadBalancerListeners

Smart constructor for the minimum required parameters to construct a valid CreateLoadBalancerListeners request.

The fields accessible through corresponding lenses are:

LoadBalancerName :: Text

Listeners :: [Listener]

Request lenses

clblLoadBalancerName :: Lens' CreateLoadBalancerListeners Text

The name of the load balancer.

clblListeners :: Lens' CreateLoadBalancerListeners [Listener]

A list of LoadBalancerPort, InstancePort, Protocol, InstanceProtocol, and SSLCertificateId items.

Response

data CreateLoadBalancerListenersResponse

The output for the CreateLoadBalancerListeners action.

Response constructor

createLoadBalancerListenersResponse :: CreateLoadBalancerListenersResponse

Smart constructor for the minimum required parameters to construct a valid CreateLoadBalancerListenersResponse response.

This constructor is provided for convenience and testing purposes.