Network.AWS.ELB.CreateLoadBalancerPolicy

Description

Creates a new policy that contains the necessary attributes depending on the policy type. Policies are settings that are saved for your load balancer and that can be applied to the front-end listener, or the back-end application server, depending on your policy type.

http://docs.aws.amazon.com/ElasticLoadBalancing/latest/APIReference/API_CreateLoadBalancerPolicy.html

Request

data CreateLoadBalancerPolicy

Request constructor

createLoadBalancerPolicy

Arguments

:: Text
clbpLoadBalancerName

-> Text
clbpPolicyName

-> Text
clbpPolicyTypeName

-> CreateLoadBalancerPolicy

CreateLoadBalancerPolicy constructor.

The fields accessible through corresponding lenses are:

clbpLoadBalancerName :: Text

clbpPolicyAttributes :: [PolicyAttribute]

clbpPolicyName :: Text

clbpPolicyTypeName :: Text

Request lenses

clbpLoadBalancerName :: Lens' CreateLoadBalancerPolicy Text

The name associated with the LoadBalancer for which the policy is being created.

clbpPolicyAttributes :: Lens' CreateLoadBalancerPolicy [PolicyAttribute]

A list of attributes associated with the policy being created.

clbpPolicyName :: Lens' CreateLoadBalancerPolicy Text

The name of the load balancer policy being created. The name must be unique within the set of policies for this load balancer.

clbpPolicyTypeName :: Lens' CreateLoadBalancerPolicy Text

The name of the base policy type being used to create this policy. To get the list of policy types, use the DescribeLoadBalancerPolicyTypes action.

Response

data CreateLoadBalancerPolicyResponse

Response constructor

createLoadBalancerPolicyResponse :: CreateLoadBalancerPolicyResponse

CreateLoadBalancerPolicyResponse constructor.