Network.AWS.ELB.DeleteLoadBalancer

Description

Deletes the specified load balancer.

If you are attempting to recreate a load balancer, you must reconfigure all settings. The DNS name associated with a deleted load balancer are no longer usable. The name and associated DNS record of the deleted load balancer no longer exist and traffic sent to any of its IP addresses is no longer delivered to your instances.

If the load balancer does not exist or has already been deleted, the call to DeleteLoadBalancer still succeeds.

Creating a Request

deleteLoadBalancer

Arguments

:: Text
dlbLoadBalancerName

-> DeleteLoadBalancer

Creates a value of DeleteLoadBalancer with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dlbLoadBalancerName - The name of the load balancer.

data DeleteLoadBalancer

Contains the parameters for DeleteLoadBalancer.

See: deleteLoadBalancer smart constructor.

Request Lenses

dlbLoadBalancerName :: Lens' DeleteLoadBalancer Text

The name of the load balancer.

Destructuring the Response

deleteLoadBalancerResponse

Arguments

:: Int
drsResponseStatus

-> DeleteLoadBalancerResponse

Creates a value of DeleteLoadBalancerResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

drsResponseStatus - -- | The response status code.

data DeleteLoadBalancerResponse

Contains the output of DeleteLoadBalancer.

See: deleteLoadBalancerResponse smart constructor.

Response Lenses

drsResponseStatus :: Lens' DeleteLoadBalancerResponse Int

- | The response status code.