Network.AWS.ELB.DeleteLoadBalancerListeners

Description

Deletes the specified listeners from the specified load balancer.

Creating a Request

deleteLoadBalancerListeners

Arguments

:: Text
dlblLoadBalancerName

-> DeleteLoadBalancerListeners

Creates a value of DeleteLoadBalancerListeners with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dlblLoadBalancerName - The name of the load balancer.

dlblLoadBalancerPorts - The client port numbers of the listeners.

data DeleteLoadBalancerListeners

Contains the parameters for DeleteLoadBalancerListeners.

See: deleteLoadBalancerListeners smart constructor.

Request Lenses

dlblLoadBalancerName :: Lens' DeleteLoadBalancerListeners Text

The name of the load balancer.

dlblLoadBalancerPorts :: Lens' DeleteLoadBalancerListeners [Int]

The client port numbers of the listeners.

Destructuring the Response

deleteLoadBalancerListenersResponse

Arguments

:: Int
dlblrsResponseStatus

-> DeleteLoadBalancerListenersResponse

Creates a value of DeleteLoadBalancerListenersResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dlblrsResponseStatus - -- | The response status code.

data DeleteLoadBalancerListenersResponse

Contains the output of DeleteLoadBalancerListeners.

See: deleteLoadBalancerListenersResponse smart constructor.

Response Lenses

dlblrsResponseStatus :: Lens' DeleteLoadBalancerListenersResponse Int

- | The response status code.