Network.AWS.ELB.DeleteLoadBalancerPolicy

Description

Deletes the specified policy from the specified load balancer. This policy must not be enabled for any listeners.

Creating a Request

deleteLoadBalancerPolicy

Arguments

:: Text
dLoadBalancerName

-> Text
dPolicyName

-> DeleteLoadBalancerPolicy

Creates a value of DeleteLoadBalancerPolicy with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dLoadBalancerName - The name of the load balancer.

dPolicyName - The name of the policy.

data DeleteLoadBalancerPolicy

Contains the parameters for DeleteLoadBalancerPolicy.

See: deleteLoadBalancerPolicy smart constructor.

Request Lenses

dLoadBalancerName :: Lens' DeleteLoadBalancerPolicy Text

The name of the load balancer.

dPolicyName :: Lens' DeleteLoadBalancerPolicy Text

The name of the policy.

Destructuring the Response

deleteLoadBalancerPolicyResponse

Arguments

:: Int
delrsResponseStatus

-> DeleteLoadBalancerPolicyResponse

Creates a value of DeleteLoadBalancerPolicyResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

delrsResponseStatus - -- | The response status code.

data DeleteLoadBalancerPolicyResponse

Contains the output of DeleteLoadBalancerPolicy.

See: deleteLoadBalancerPolicyResponse smart constructor.

Response Lenses

delrsResponseStatus :: Lens' DeleteLoadBalancerPolicyResponse Int

- | The response status code.