Network.AWS.ELB.DeregisterInstancesFromLoadBalancer

Description

Deregisters instances from the load balancer. Once the instance is deregistered, it will stop receiving traffic from the load balancer.

In order to successfully call this API, the same account credentials as those used to create the load balancer must be provided.

For more information, see De-register and Register Amazon EC2 Instances in the Elastic Load Balancing Developer Guide.

You can use DescribeLoadBalancers to verify if the instance is deregistered from the load balancer.

http://docs.aws.amazon.com/ElasticLoadBalancing/latest/APIReference/API_DeregisterInstancesFromLoadBalancer.html

Request

data DeregisterInstancesFromLoadBalancer

Request constructor

deregisterInstancesFromLoadBalancer

Arguments

:: Text
diflbLoadBalancerName

-> DeregisterInstancesFromLoadBalancer

DeregisterInstancesFromLoadBalancer constructor.

The fields accessible through corresponding lenses are:

diflbInstances :: [Instance]

diflbLoadBalancerName :: Text

Request lenses

diflbInstances :: Lens' DeregisterInstancesFromLoadBalancer [Instance]

A list of EC2 instance IDs consisting of all instances to be deregistered.

diflbLoadBalancerName :: Lens' DeregisterInstancesFromLoadBalancer Text

The name associated with the load balancer.

Response

data DeregisterInstancesFromLoadBalancerResponse

Response constructor

deregisterInstancesFromLoadBalancerResponse :: DeregisterInstancesFromLoadBalancerResponse

DeregisterInstancesFromLoadBalancerResponse constructor.

The fields accessible through corresponding lenses are:

diflbrInstances :: [Instance]

Response lenses

diflbrInstances :: Lens' DeregisterInstancesFromLoadBalancerResponse [Instance]

An updated list of remaining instances registered with the load balancer.