Network.AWS.ELB.DescribeAccountLimits

Description

Describes the current Elastic Load Balancing resource limits for your AWS account.

For more information, see Limits for Your Classic Load Balancer in the Classic Load Balancer Guide .

Creating a Request

describeAccountLimits :: DescribeAccountLimits

Creates a value of DescribeAccountLimits with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dalMarker - The marker for the next set of results. (You received this marker from a previous call.)

dalPageSize - The maximum number of results to return with this call.

data DescribeAccountLimits

See: describeAccountLimits smart constructor.

Request Lenses

dalMarker :: Lens' DescribeAccountLimits (Maybe Text)

The marker for the next set of results. (You received this marker from a previous call.)

dalPageSize :: Lens' DescribeAccountLimits (Maybe Natural)

The maximum number of results to return with this call.

Destructuring the Response

describeAccountLimitsResponse

Arguments

:: Int
dalrsResponseStatus

-> DescribeAccountLimitsResponse

Creates a value of DescribeAccountLimitsResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dalrsLimits - Information about the limits.

dalrsNextMarker - The marker to use when requesting the next set of results. If there are no additional results, the string is empty.

dalrsResponseStatus - -- | The response status code.

data DescribeAccountLimitsResponse

See: describeAccountLimitsResponse smart constructor.

Response Lenses

dalrsLimits :: Lens' DescribeAccountLimitsResponse [Limit]

Information about the limits.

dalrsNextMarker :: Lens' DescribeAccountLimitsResponse (Maybe Text)

The marker to use when requesting the next set of results. If there are no additional results, the string is empty.

dalrsResponseStatus :: Lens' DescribeAccountLimitsResponse Int

- | The response status code.