Network.AWS.ELB.DescribeInstanceHealth

Description

Returns the current state of the specified instances registered with the specified load balancer. If no instances are specified, the state of all the instances registered with the load balancer is returned. You must provide the same account credentials as those that were used to create the load balancer. Description of a healthy (InService) instance https:/elasticloadbalancing.amazonaws.com?LoadBalancerName=my-test-loadbalancer &Version=2012-06-01 &Action=DescribeInstanceHealth &AUTHPARAMS N/A i-90d8c2a5 InService N/A 1549581b-12b7-11e3-895e-1334aEXAMPLE Description of an instance with registration in progress https:/elasticloadbalancing.amazonaws.com?LoadBalancerName=my-test-loadbalancer &Version=2012-06-01 &Action=DescribeInstanceHealth &AUTHPARAMS Instance registration is still in progress. i-315b7e51 OutOfService ELB 1549581b-12b7-11e3-895e-1334aEXAMPLE Description of an unhealthy (OutOfService) instance https:/elasticloadbalancing.amazonaws.com?LoadBalancerName=my-test-loadbalancer &Version=2012-06-01 &Action=DescribeInstanceHealth &AUTHPARAMS Instance has failed at least the UnhealthyThreshold number of health checks consecutively. i-fda142c9 OutOfService Instance 83c88b9d-12b7-11e3-8b82-87b12EXAMPLE Description of an instance in an unknown state https:/elasticloadbalancing.amazonaws.com?LoadBalancerName=my-test-loadbalancer &Version=2012-06-01 &Action=DescribeInstanceHealth &AUTHPARAMS A transient error occurred. Please try again later. i-7f12e649 Unknown ELB 83c88b9d-12b7-11e3-8b82-87b12EXAMPLE.

Request

data DescribeInstanceHealth

The input for the DescribeEndPointState action.

Request constructor

mkDescribeInstanceHealth

Arguments

:: Text
dihLoadBalancerName

-> DescribeInstanceHealth

Smart constructor for the minimum required parameters to construct a valid DescribeInstanceHealth request.

The fields accessible through corresponding lenses are:

LoadBalancerName :: Text

Instances :: [Instance]

Request lenses

dihLoadBalancerName :: Lens' DescribeInstanceHealth Text

The name of the load balancer.

dihInstances :: Lens' DescribeInstanceHealth [Instance]

A list of instance IDs whose states are being queried.

Response

data DescribeInstanceHealthResponse

The output for the DescribeInstanceHealth action.

Response constructor

mkDescribeInstanceHealthResponse :: DescribeInstanceHealthResponse

Smart constructor for the minimum required parameters to construct a valid DescribeInstanceHealthResponse response.

This constructor is provided for convenience and testing purposes.

The fields accessible through corresponding lenses are:

InstanceStates :: [InstanceState]

Response lenses

dihrInstanceStates :: Lens' DescribeInstanceHealthResponse [InstanceState]

A list containing health information for the specified instances.