Network.AWS.ELB.DescribeLoadBalancerAttributes

Description

Describes the attributes for the specified load balancer.

Creating a Request

describeLoadBalancerAttributes

Arguments

:: Text
dlbaLoadBalancerName

-> DescribeLoadBalancerAttributes

Creates a value of DescribeLoadBalancerAttributes with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dlbaLoadBalancerName - The name of the load balancer.

data DescribeLoadBalancerAttributes

Contains the parameters for DescribeLoadBalancerAttributes.

See: describeLoadBalancerAttributes smart constructor.

Request Lenses

dlbaLoadBalancerName :: Lens' DescribeLoadBalancerAttributes Text

The name of the load balancer.

Destructuring the Response

describeLoadBalancerAttributesResponse

Arguments

:: Int
dlbarsResponseStatus

-> DescribeLoadBalancerAttributesResponse

Creates a value of DescribeLoadBalancerAttributesResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dlbarsLoadBalancerAttributes - Information about the load balancer attributes.

dlbarsResponseStatus - -- | The response status code.

data DescribeLoadBalancerAttributesResponse

Contains the output of DescribeLoadBalancerAttributes.

See: describeLoadBalancerAttributesResponse smart constructor.

Response Lenses

dlbarsLoadBalancerAttributes :: Lens' DescribeLoadBalancerAttributesResponse (Maybe LoadBalancerAttributes)

Information about the load balancer attributes.

dlbarsResponseStatus :: Lens' DescribeLoadBalancerAttributesResponse Int

- | The response status code.