Network.AWS.ELB.DescribeLoadBalancerPolicies

Description

Returns detailed descriptions of the policies. If you specify a load balancer name, the action returns the descriptions of all the policies created for the load balancer. If you specify a policy name associated with your load balancer, the action returns the description of that policy. If you don't specify a load balancer name, the action returns descriptions of the specified sample policies, or descriptions of all the sample policies. The names of the sample policies have the 'ELBSample-' prefix.

http://docs.aws.amazon.com/ElasticLoadBalancing/latest/APIReference/API_DescribeLoadBalancerPolicies.html

Request

data DescribeLoadBalancerPolicies

Request constructor

describeLoadBalancerPolicies :: DescribeLoadBalancerPolicies

DescribeLoadBalancerPolicies constructor.

The fields accessible through corresponding lenses are:

dlbpLoadBalancerName :: Maybe Text

dlbpPolicyNames :: [Text]

Request lenses

dlbpLoadBalancerName :: Lens' DescribeLoadBalancerPolicies (Maybe Text)

The mnemonic name associated with the load balancer. If no name is specified, the operation returns the attributes of either all the sample policies pre-defined by Elastic Load Balancing or the specified sample polices.

dlbpPolicyNames :: Lens' DescribeLoadBalancerPolicies [Text]

The names of load balancer policies you've created or Elastic Load Balancing sample policy names.

Response

data DescribeLoadBalancerPoliciesResponse

Response constructor

describeLoadBalancerPoliciesResponse :: DescribeLoadBalancerPoliciesResponse

DescribeLoadBalancerPoliciesResponse constructor.

The fields accessible through corresponding lenses are:

dlbprPolicyDescriptions :: [PolicyDescription]

Response lenses

dlbprPolicyDescriptions :: Lens' DescribeLoadBalancerPoliciesResponse [PolicyDescription]

A list of policy description structures.