Network.AWS.ELB.DescribeLoadBalancerPolicyTypes

Description

Returns meta-information on the specified load balancer policies defined by the Elastic Load Balancing service. The policy types that are returned from this action can be used in a CreateLoadBalancerPolicy action to instantiate specific policy configurations that will be applied to a load balancer.

http://docs.aws.amazon.com/ElasticLoadBalancing/latest/APIReference/API_DescribeLoadBalancerPolicyTypes.html

Request

data DescribeLoadBalancerPolicyTypes

Request constructor

describeLoadBalancerPolicyTypes :: DescribeLoadBalancerPolicyTypes

DescribeLoadBalancerPolicyTypes constructor.

The fields accessible through corresponding lenses are:

dlbptPolicyTypeNames :: [Text]

Request lenses

dlbptPolicyTypeNames :: Lens' DescribeLoadBalancerPolicyTypes [Text]

Specifies the name of the policy types. If no names are specified, returns the description of all the policy types defined by Elastic Load Balancing service.

Response

data DescribeLoadBalancerPolicyTypesResponse

Response constructor

describeLoadBalancerPolicyTypesResponse :: DescribeLoadBalancerPolicyTypesResponse

DescribeLoadBalancerPolicyTypesResponse constructor.

The fields accessible through corresponding lenses are:

dlbptrPolicyTypeDescriptions :: [PolicyTypeDescription]

Response lenses

dlbptrPolicyTypeDescriptions :: Lens' DescribeLoadBalancerPolicyTypesResponse [PolicyTypeDescription]

List of policy type description structures of the specified policy type. If no policy type names are specified, returns the description of all the policy types defined by Elastic Load Balancing service.