Network.AWS.ELB.DescribeLoadBalancers

Description

Returns detailed configuration information for all the load balancers created for the account. If you specify load balancer names, the action returns configuration information of the specified load balancers. In order to retrieve this information, you must provide the same account credentials that was used to create the load balancer. Description of a specified load balancer https:/elasticloadbalancing.amazonaws.com?LoadBalancerNames.member.1=MyLoadBalancer &Version=2012-06-01 &Action=DescribeLoadBalancers &AUTHPARAMS MyLoadBalancer 2013-05-24T21:15:31.280Z 90 HTTP:80/ 2 60 10 HTTP 80 HTTP 80 i-e4cbe38d us-east-1a ZZZZZZZZZZZ123X MyLoadBalancer-123456789.us-east-1.elb.amazonaws.com internet-facing amazon-elb amazon-elb-sg MyLoadBalancer-123456789.us-east-1.elb.amazonaws.com 83c88b9d-12b7-11e3-8b82-87b12EXAMPLE.

Request

data DescribeLoadBalancers

The input for the DescribeLoadBalancers action.

Request constructor

describeLoadBalancers :: DescribeLoadBalancers

Smart constructor for the minimum required parameters to construct a valid DescribeLoadBalancers request.

The fields accessible through corresponding lenses are:

LoadBalancerNames :: [Text]

Marker :: Maybe Text

PageSize :: Maybe Integer

Request lenses

dlb1LoadBalancerNames :: Lens' DescribeLoadBalancers [Text]

A list of load balancer names associated with the account.

dlb1Marker :: Lens' DescribeLoadBalancers (Maybe Text)

An optional parameter used for pagination of results from this call. If specified, the response includes only records beyond the marker.

dlb1PageSize :: Lens' DescribeLoadBalancers (Maybe Integer)

The number of results returned in each page. The default is 400. You cannot specify a page size greater than 400 or less than 1.

Response

data DescribeLoadBalancersResponse

The output for the DescribeLoadBalancers action.

Response constructor

describeLoadBalancersResponse :: DescribeLoadBalancersResponse

Smart constructor for the minimum required parameters to construct a valid DescribeLoadBalancersResponse response.

This constructor is provided for convenience and testing purposes.

The fields accessible through corresponding lenses are:

LoadBalancerDescriptions :: [LoadBalancerDescription]

NextMarker :: Maybe Text

Response lenses

dlbrrLoadBalancerDescriptions :: Lens' DescribeLoadBalancersResponse [LoadBalancerDescription]

A list of load balancer description structures.

dlbrrNextMarker :: Lens' DescribeLoadBalancersResponse (Maybe Text)

Specifies the value of next marker if the request returned more than one page of results.