Network.AWS.ELB.DescribeTags

Description

Describes the tags associated with one or more load balancers. https:/elasticloadbalancing.amazonaws.com/?Action=DescribeTags &LoadBalancerNames.member.1=my-test-loadbalancer &Version=2012-06-01 &AUTHPARAMS my-test-project project test environment my-test-loadbalancer 07b1ecbc-1100-11e3-acaf-dd7edEXAMPLE.

Request

data DescribeTags

The input for the DescribeTags action.

Request constructor

mkDescribeTags

Arguments

:: List1 Text
dtLoadBalancerNames

-> DescribeTags

Smart constructor for the minimum required parameters to construct a valid DescribeTags request.

The fields accessible through corresponding lenses are:

LoadBalancerNames :: List1 Text

Request lenses

dtLoadBalancerNames :: Lens' DescribeTags (List1 Text)

The names of the load balancers.

Response

data DescribeTagsResponse

The output for the DescribeTags action.

Response constructor

mkDescribeTagsResponse :: DescribeTagsResponse

Smart constructor for the minimum required parameters to construct a valid DescribeTagsResponse response.

This constructor is provided for convenience and testing purposes.

The fields accessible through corresponding lenses are:

TagDescriptions :: [TagDescription]

Response lenses

dtrTagDescriptions :: Lens' DescribeTagsResponse [TagDescription]

A list of tag description structures.