Network.AWS.ELB.DetachLoadBalancerFromSubnets

Description

Removes subnets from the set of configured subnets in the Amazon Virtual Private Cloud (Amazon VPC) for the load balancer.

After a subnet is removed all of the EC2 instances registered with the load balancer that are in the removed subnet will go into the OutOfService state. When a subnet is removed, the load balancer will balance the traffic among the remaining routable subnets for the load balancer.

http://docs.aws.amazon.com/ElasticLoadBalancing/latest/APIReference/API_DetachLoadBalancerFromSubnets.html

Request

data DetachLoadBalancerFromSubnets

Request constructor

detachLoadBalancerFromSubnets

Arguments

:: Text
dlbfsLoadBalancerName

-> DetachLoadBalancerFromSubnets

DetachLoadBalancerFromSubnets constructor.

The fields accessible through corresponding lenses are:

dlbfsLoadBalancerName :: Text

dlbfsSubnets :: [Text]

Request lenses

dlbfsLoadBalancerName :: Lens' DetachLoadBalancerFromSubnets Text

The name associated with the load balancer to be detached.

dlbfsSubnets :: Lens' DetachLoadBalancerFromSubnets [Text]

A list of subnet IDs to remove from the set of configured subnets for the load balancer.

Response

data DetachLoadBalancerFromSubnetsResponse

Response constructor

detachLoadBalancerFromSubnetsResponse :: DetachLoadBalancerFromSubnetsResponse

DetachLoadBalancerFromSubnetsResponse constructor.

The fields accessible through corresponding lenses are:

dlbfsrSubnets :: [Text]

Response lenses

dlbfsrSubnets :: Lens' DetachLoadBalancerFromSubnetsResponse [Text]

A list of subnet IDs the load balancer is now attached to.