Network.AWS.ELB.DisableAvailabilityZonesForLoadBalancer

Description

Removes the specified EC2 Availability Zones from the set of configured Availability Zones for the load balancer. There must be at least one Availability Zone registered with a load balancer at all times. Once an Availability Zone is removed, all the instances registered with the load balancer that are in the removed Availability Zone go into the OutOfService state. Upon Availability Zone removal, the load balancer attempts to equally balance the traffic among its remaining usable Availability Zones. Trying to remove an Availability Zone that was not associated with the load balancer does nothing. For more information, see Disable an Availability Zone from a Load-Balanced Application in the Elastic Load Balancing Developer Guide. https:/elasticloadbalancing.amazonaws.com?AvailabilityZones.member.1=us-east-1a &LoadBalancerName=MyHTTPSLoadBalancer &Version=2012-06-01 &Action=DisableAvailabilityZonesForLoadBalancer &AUTHPARAMS us-east-1b ba6267d5-2566-11e3-9c6d-eb728EXAMPLE.

Request

data DisableAvailabilityZonesForLoadBalancer

The input for the DisableAvailabilityZonesForLoadBalancer action.

Request constructor

disableAvailabilityZonesForLoadBalancer

Arguments

:: Text
dazflbLoadBalancerName

-> [Text]
dazflbAvailabilityZones

-> DisableAvailabilityZonesForLoadBalancer

Smart constructor for the minimum required parameters to construct a valid DisableAvailabilityZonesForLoadBalancer request.

The fields accessible through corresponding lenses are:

LoadBalancerName :: Text

AvailabilityZones :: [Text]

Request lenses

dazflbLoadBalancerName :: Lens' DisableAvailabilityZonesForLoadBalancer Text

The name associated with the load balancer.

dazflbAvailabilityZones :: Lens' DisableAvailabilityZonesForLoadBalancer [Text]

A list of Availability Zones to be removed from the load balancer. There must be at least one Availability Zone registered with a load balancer at all times. Specified Availability Zones must be in the same region.

Response

data DisableAvailabilityZonesForLoadBalancerResponse

The output for the DisableAvailabilityZonesForLoadBalancer action.

Response constructor

disableAvailabilityZonesForLoadBalancerResponse :: DisableAvailabilityZonesForLoadBalancerResponse

Smart constructor for the minimum required parameters to construct a valid DisableAvailabilityZonesForLoadBalancerResponse response.

This constructor is provided for convenience and testing purposes.

The fields accessible through corresponding lenses are:

AvailabilityZones :: [Text]

Response lenses

dazflbrAvailabilityZones :: Lens' DisableAvailabilityZonesForLoadBalancerResponse [Text]

A list of updated Availability Zones for the load balancer.