Network.AWS.ELB.EnableAvailabilityZonesForLoadBalancer

Description

Adds one or more EC2 Availability Zones to the load balancer.

The load balancer evenly distributes requests across all its registered Availability Zones that contain instances.

The new EC2 Availability Zones to be added must be in the same EC2 Region as the Availability Zones for which the load balancer was created. For more information, see Expand a Load Balanced Application to an AdditionalAvailability Zone in the Elastic Load Balancing Developer Guide.

http://docs.aws.amazon.com/ElasticLoadBalancing/latest/APIReference/API_EnableAvailabilityZonesForLoadBalancer.html

Request

data EnableAvailabilityZonesForLoadBalancer

Request constructor

enableAvailabilityZonesForLoadBalancer

Arguments

:: Text
eazflbLoadBalancerName

-> EnableAvailabilityZonesForLoadBalancer

EnableAvailabilityZonesForLoadBalancer constructor.

The fields accessible through corresponding lenses are:

eazflbAvailabilityZones :: [Text]

eazflbLoadBalancerName :: Text

Request lenses

eazflbAvailabilityZones :: Lens' EnableAvailabilityZonesForLoadBalancer [Text]

A list of new Availability Zones for the load balancer. Each Availability Zone must be in the same region as the load balancer.

eazflbLoadBalancerName :: Lens' EnableAvailabilityZonesForLoadBalancer Text

The name associated with the load balancer.

Response

data EnableAvailabilityZonesForLoadBalancerResponse

Response constructor

enableAvailabilityZonesForLoadBalancerResponse :: EnableAvailabilityZonesForLoadBalancerResponse

EnableAvailabilityZonesForLoadBalancerResponse constructor.

The fields accessible through corresponding lenses are:

eazflbrAvailabilityZones :: [Text]

Response lenses

eazflbrAvailabilityZones :: Lens' EnableAvailabilityZonesForLoadBalancerResponse [Text]

An updated list of Availability Zones for the load balancer.