Network.AWS.ELB.ModifyLoadBalancerAttributes

Description

Modifies the attributes of the specified load balancer.

You can modify the load balancer attributes, such as AccessLogs , ConnectionDraining , and CrossZoneLoadBalancing by either enabling or disabling them. Or, you can modify the load balancer attribute ConnectionSettings by specifying an idle connection timeout value for your load balancer.

For more information, see the following in the Classic Load Balancer Guide :

Cross-Zone Load Balancing

Connection Draining

Access Logs

Idle Connection Timeout

Creating a Request

modifyLoadBalancerAttributes

Arguments

:: Text
mlbaLoadBalancerName

-> LoadBalancerAttributes
mlbaLoadBalancerAttributes

-> ModifyLoadBalancerAttributes

Creates a value of ModifyLoadBalancerAttributes with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

mlbaLoadBalancerName - The name of the load balancer.

mlbaLoadBalancerAttributes - The attributes for the load balancer.

data ModifyLoadBalancerAttributes

Contains the parameters for ModifyLoadBalancerAttributes.

See: modifyLoadBalancerAttributes smart constructor.

Request Lenses

mlbaLoadBalancerName :: Lens' ModifyLoadBalancerAttributes Text

The name of the load balancer.

mlbaLoadBalancerAttributes :: Lens' ModifyLoadBalancerAttributes LoadBalancerAttributes

The attributes for the load balancer.

Destructuring the Response

modifyLoadBalancerAttributesResponse

Arguments

:: Int
mlbarsResponseStatus

-> ModifyLoadBalancerAttributesResponse

Creates a value of ModifyLoadBalancerAttributesResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

mlbarsLoadBalancerName - The name of the load balancer.

mlbarsLoadBalancerAttributes - Information about the load balancer attributes.

mlbarsResponseStatus - -- | The response status code.

data ModifyLoadBalancerAttributesResponse

Contains the output of ModifyLoadBalancerAttributes.

See: modifyLoadBalancerAttributesResponse smart constructor.

Response Lenses

mlbarsLoadBalancerName :: Lens' ModifyLoadBalancerAttributesResponse (Maybe Text)

The name of the load balancer.

mlbarsLoadBalancerAttributes :: Lens' ModifyLoadBalancerAttributesResponse (Maybe LoadBalancerAttributes)

Information about the load balancer attributes.

mlbarsResponseStatus :: Lens' ModifyLoadBalancerAttributesResponse Int

- | The response status code.