Network.AWS.ELB.RegisterInstancesWithLoadBalancer

Description

Adds new instances to the load balancer.

Once the instance is registered, it starts receiving traffic and requests from the load balancer. Any instance that is not in any of the Availability Zones registered for the load balancer will be moved to the OutOfService state. It will move to the InService state when the Availability Zone is added to the load balancer.

When an instance registered with a load balancer is stopped and then restarted, the IP addresses associated with the instance changes. Elastic Load Balancing cannot recognize the new IP address, which prevents it from routing traffic to the instances. We recommend that you de-register your Amazon EC2 instances from your load balancer after you stop your instance, and then register the load balancer with your instance after you've restarted. To de-register your instances from load balancer, use DeregisterInstancesFromLoadBalancer action.

For more information, see De-register and Register Amazon EC2 Instances in the Elastic Load Balancing Developer Guide.

In order for this call to be successful, you must provide the same account credentials as those that were used to create the load balancer. Completion of this API does not guarantee that operation has completed. Rather, it means that the request has been registered and the changes will happen shortly. You can use DescribeLoadBalancers or DescribeInstanceHealth action to check the state of the newly registered instances.

http://docs.aws.amazon.com/ElasticLoadBalancing/latest/APIReference/API_RegisterInstancesWithLoadBalancer.html

Request

data RegisterInstancesWithLoadBalancer

Request constructor

registerInstancesWithLoadBalancer

Arguments

:: Text
riwlbLoadBalancerName

-> RegisterInstancesWithLoadBalancer

RegisterInstancesWithLoadBalancer constructor.

The fields accessible through corresponding lenses are:

riwlbInstances :: [Instance]

riwlbLoadBalancerName :: Text

Request lenses

riwlbInstances :: Lens' RegisterInstancesWithLoadBalancer [Instance]

A list of instance IDs that should be registered with the load balancer.

riwlbLoadBalancerName :: Lens' RegisterInstancesWithLoadBalancer Text

The name associated with the load balancer. The name must be unique within your set of load balancers.

Response

data RegisterInstancesWithLoadBalancerResponse

Response constructor

registerInstancesWithLoadBalancerResponse :: RegisterInstancesWithLoadBalancerResponse

RegisterInstancesWithLoadBalancerResponse constructor.

The fields accessible through corresponding lenses are:

riwlbrInstances :: [Instance]

Response lenses

riwlbrInstances :: Lens' RegisterInstancesWithLoadBalancerResponse [Instance]

An updated list of instances for the load balancer.