Network.AWS.ELB.RemoveTags

Description

Removes one or more tags from the specified load balancer. Remove Two Tag Keys from the Load Balancer https:/elasticloadbalancing.amazonaws.com?LoadBalancerName=my-test-loadbalancer &Tags.member.1.Key=owner &Tags.member.2.Key=project &Action=RemoveTags &Version=2012-06-01 &AUTHPARAMS 83c88b9d-12b7-11e3-8b82-87b12EXAMPLE.

Request

data RemoveTags

The input for the RemoveTags action.

Request constructor

removeTags

Arguments

:: [Text]
rtLoadBalancerNames

-> List1 TagKeyOnly
rtTags

-> RemoveTags

Smart constructor for the minimum required parameters to construct a valid RemoveTags request.

The fields accessible through corresponding lenses are:

LoadBalancerNames :: [Text]

Tags :: List1 TagKeyOnly

Request lenses

rtLoadBalancerNames :: Lens' RemoveTags [Text]

The name of the load balancer. You can specify a maximum of one load balancer name.

rtTags :: Lens' RemoveTags (List1 TagKeyOnly)

A list of tag keys to remove.

Response

data RemoveTagsResponse

The output for the RemoveTags action.

Response constructor

removeTagsResponse :: RemoveTagsResponse

Smart constructor for the minimum required parameters to construct a valid RemoveTagsResponse response.

This constructor is provided for convenience and testing purposes.