Network.AWS.ELB.SetLoadBalancerListenerSSLCertificate

Description

Sets the certificate that terminates the specified listener's SSL connections. The specified certificate replaces any prior certificate that was used on the same load balancer and port.

For more information about updating your SSL certificate, see Replace the SSL Certificate for Your Load Balancer in the Classic Load Balancer Guide .

Creating a Request

setLoadBalancerListenerSSLCertificate

Arguments

:: Text
slblscLoadBalancerName

-> Int
slblscLoadBalancerPort

-> Text
slblscSSLCertificateId

-> SetLoadBalancerListenerSSLCertificate

Creates a value of SetLoadBalancerListenerSSLCertificate with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

slblscLoadBalancerName - The name of the load balancer.

slblscLoadBalancerPort - The port that uses the specified SSL certificate.

slblscSSLCertificateId - The Amazon Resource Name (ARN) of the SSL certificate.

data SetLoadBalancerListenerSSLCertificate

Contains the parameters for SetLoadBalancerListenerSSLCertificate.

See: setLoadBalancerListenerSSLCertificate smart constructor.

Request Lenses

slblscLoadBalancerName :: Lens' SetLoadBalancerListenerSSLCertificate Text

The name of the load balancer.

slblscLoadBalancerPort :: Lens' SetLoadBalancerListenerSSLCertificate Int

The port that uses the specified SSL certificate.

slblscSSLCertificateId :: Lens' SetLoadBalancerListenerSSLCertificate Text

The Amazon Resource Name (ARN) of the SSL certificate.

Destructuring the Response

setLoadBalancerListenerSSLCertificateResponse

Arguments

:: Int
slblscrsResponseStatus

-> SetLoadBalancerListenerSSLCertificateResponse

Creates a value of SetLoadBalancerListenerSSLCertificateResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

slblscrsResponseStatus - -- | The response status code.

data SetLoadBalancerListenerSSLCertificateResponse

Contains the output of SetLoadBalancerListenerSSLCertificate.

See: setLoadBalancerListenerSSLCertificateResponse smart constructor.

Response Lenses

slblscrsResponseStatus :: Lens' SetLoadBalancerListenerSSLCertificateResponse Int

- | The response status code.