Network.AWS.ELB.SetLoadBalancerPoliciesOfListener

Description

Associates, updates, or disables a policy with a listener on the load balancer. You can associate multiple policies with a listener.

http://docs.aws.amazon.com/ElasticLoadBalancing/latest/APIReference/API_SetLoadBalancerPoliciesOfListener.html

Request

data SetLoadBalancerPoliciesOfListener

Request constructor

setLoadBalancerPoliciesOfListener

Arguments

:: Text
slbpolLoadBalancerName

-> Int
slbpolLoadBalancerPort

-> SetLoadBalancerPoliciesOfListener

SetLoadBalancerPoliciesOfListener constructor.

The fields accessible through corresponding lenses are:

slbpolLoadBalancerName :: Text

slbpolLoadBalancerPort :: Int

slbpolPolicyNames :: [Text]

Request lenses

slbpolLoadBalancerName :: Lens' SetLoadBalancerPoliciesOfListener Text

The name of the load balancer.

slbpolLoadBalancerPort :: Lens' SetLoadBalancerPoliciesOfListener Int

The external port of the load balancer to associate the policy.

slbpolPolicyNames :: Lens' SetLoadBalancerPoliciesOfListener [Text]

List of policies to be associated with the listener. If the list is empty, the current policy is removed from the listener.

Response

data SetLoadBalancerPoliciesOfListenerResponse

Response constructor

setLoadBalancerPoliciesOfListenerResponse :: SetLoadBalancerPoliciesOfListenerResponse

SetLoadBalancerPoliciesOfListenerResponse constructor.