Network.AWS.ELB.V2012_06_01.AddTags

Description

Adds one or more tags for the specified load balancer. Each load balancer can have a maximum of 10 tags. Each tag consists of a key and an optional value. Tag keys must be unique for each load balancer. If a tag with the same key is already associated with the load balancer, this action will update the value of the key. For more information, see Tagging in the Elastic Load Balancing Developer Guide. https:/elasticloadbalancing.amazonaws.com?LoadBalancerNames.member.1=my-test-loadbalancer &Action=AddTags &Tags.member.1.Key=project &Tags.member.1.Value=my-test-project &Version=2012-06-01 &AUTHPARAMS 360e81f7-1100-11e4-b6ed-0f30EXAMPLE.

Documentation

data AddTags

Constructors

AddTags

Fields

_atiLoadBalancerNames :: [Text]

The name of the load balancer to tag. You can specify a maximum of one load balancer name.

_atiTags :: [Tag]

A list of tags for each load balancer.

atiTags :: Lens' AddTags [Tag]

atiLoadBalancerNames :: Lens' AddTags [Text]

data AddTagsResponse

Constructors

AddTagsResponse