Network.AWS.ELB.V2012_06_01.ApplySecurityGroupsToLoadBalancer

Description

Associates one or more security groups with your load balancer in Amazon Virtual Private Cloud (Amazon VPC). The provided security group IDs will override any currently applied security groups. For more information, see Manage Security Groups in Amazon VPC in the Elastic Load Balancing Developer Guide. https:/elasticloadbalancing.amazonaws.com?SecurityGroups.member.1=sg-123456789 &LoadBalancerName=my-test-vpc-loadbalancer &Version=2012-06-01 &Action=ApplySecurityGroupsToLoadBalancer &AUTHPARAMS sg-123456789 06b5decc-102a-11e3-9ad6-bf3e4EXAMPLE.

Documentation

data ApplySecurityGroupsToLoadBalancer

Constructors

ApplySecurityGroupsToLoadBalancer

Fields

_asgtlbiLoadBalancerName :: Text

The name associated with the load balancer. The name must be unique within the set of load balancers associated with your AWS account.

_asgtlbiSecurityGroups :: [Text]

A list of security group IDs to associate with your load balancer in VPC. The security group IDs must be provided as the ID and not the security group name (For example, sg-1234).

asgtlbiSecurityGroups :: Lens' ApplySecurityGroupsToLoadBalancer [Text]

asgtlbiLoadBalancerName :: Lens' ApplySecurityGroupsToLoadBalancer Text

data ApplySecurityGroupsToLoadBalancerResponse

Constructors

ApplySecurityGroupsToLoadBalancerResponse

Fields

_asgtlboSecurityGroups :: [Text]

A list of security group IDs associated with your load balancer.

asgtlboSecurityGroups :: Iso' ApplySecurityGroupsToLoadBalancerResponse [Text]