Network.AWS.ELB.V2012_06_01.AttachLoadBalancerToSubnets

Description

Adds one or more subnets to the set of configured subnets in the Amazon Virtual Private Cloud (Amazon VPC) for the load balancer. The load balancers evenly distribute requests across all of the registered subnets. For more information, see Deploy Elastic Load Balancing in Amazon VPC in the Elastic Load Balancing Developer Guide. https:/elasticloadbalancing.amazonaws.com?Subnets.member.1=subnet-3561b05e &LoadBalancerName=my-test-vpc-loadbalancer &Version=2012-06-01 &Action=AttachLoadBalancerToSubnets &AUTHPARAMS subnet-119f0078 subnet-3561b05e 07b1ecbc-1100-11e3-acaf-dd7edEXAMPLE.

Documentation

data AttachLoadBalancerToSubnets

Constructors

AttachLoadBalancerToSubnets

Fields

_albtsiLoadBalancerName :: Text

The name associated with the load balancer. The name must be unique within the set of load balancers associated with your AWS account.

_albtsiSubnets :: [Text]

A list of subnet IDs to add for the load balancer. You can add only one subnet per Availability Zone.

albtsiSubnets :: Lens' AttachLoadBalancerToSubnets [Text]

albtsiLoadBalancerName :: Lens' AttachLoadBalancerToSubnets Text

data AttachLoadBalancerToSubnetsResponse

Constructors

AttachLoadBalancerToSubnetsResponse

Fields

_albtsoSubnets :: [Text]

A list of subnet IDs attached to the load balancer.

albtsoSubnets :: Iso' AttachLoadBalancerToSubnetsResponse [Text]