Network.AWS.ELB.V2012_06_01.ConfigureHealthCheck

Description

Specifies the health check settings to use for evaluating the health state of your back-end instances. For more information, see Health Check in the Elastic Load Balancing Developer Guide. https:/elasticloadbalancing.amazonaws.com?HealthCheck.HealthyThreshold=2 &HealthCheck.UnhealthyThreshold=2 &LoadBalancerName=MyLoadBalancer &HealthCheck.Target=HTTP:80/ping &HealthCheck.Interval=30 &HealthCheck.Timeout=3 &Version=2012-06-01 &Action=ConfigureHealthCheck &AUTHPARAMS 30 HTTP:80/ping 2 3 2 83c88b9d-12b7-11e3-8b82-87b12EXAMPLE.

Documentation

data ConfigureHealthCheck

Constructors

ConfigureHealthCheck

Fields

_chciLoadBalancerName :: Text

The mnemonic name associated with the load balancer. The name must be unique within the set of load balancers associated with your AWS account.

_chciHealthCheck :: HealthCheck

A structure containing the configuration information for the new healthcheck.

chciLoadBalancerName :: Lens' ConfigureHealthCheck Text

chciHealthCheck :: Lens' ConfigureHealthCheck HealthCheck

data ConfigureHealthCheckResponse

Constructors

ConfigureHealthCheckResponse

Fields

_chcoHealthCheck :: Maybe HealthCheck

The updated healthcheck for the instances.

chcoHealthCheck :: Iso' ConfigureHealthCheckResponse (Maybe HealthCheck)