Network.AWS.ELB.V2012_06_01.CreateLoadBalancerPolicy

Description

Creates a new policy that contains the necessary attributes depending on the policy type. Policies are settings that are saved for your load balancer and that can be applied to the front-end listener, or the back-end application server, depending on your policy type. https:/elasticloadbalancing.amazonaws.com?PolicyAttributes.member.1.AttributeName=ProxyProtocol &PolicyAttributes.member.1.AttributeValue=true &PolicyTypeName=ProxyProtocolPolicyType &LoadBalancerName=my-test-loadbalancer &PolicyName=EnableProxyProtocol &Version=2012-06-01 &Action=CreateLoadBalancerPolicy &AUTHPARAMS 83c88b9d-12b7-11e3-8b82-87b12EXAMPLE.

Documentation

createLoadBalancerPolicy

Arguments

:: Text
_clbpiLoadBalancerName

-> Text
_clbpiPolicyName

-> Text
_clbpiPolicyTypeName

-> CreateLoadBalancerPolicy

Minimum specification for a CreateLoadBalancerPolicy request.

data CreateLoadBalancerPolicy

Constructors

CreateLoadBalancerPolicy

Fields

_clbpiLoadBalancerName :: Text

The name associated with the LoadBalancer for which the policy is being created.

_clbpiPolicyName :: Text

The name of the load balancer policy being created. The name must be unique within the set of policies for this load balancer.

_clbpiPolicyTypeName :: Text

The name of the base policy type being used to create this policy. To get the list of policy types, use the DescribeLoadBalancerPolicyTypes action.

_clbpiPolicyAttributes :: [PolicyAttribute]

A list of attributes associated with the policy being created.

clbpiPolicyTypeName :: Lens' CreateLoadBalancerPolicy Text

clbpiPolicyName :: Lens' CreateLoadBalancerPolicy Text

clbpiPolicyAttributes :: Lens' CreateLoadBalancerPolicy [PolicyAttribute]

clbpiLoadBalancerName :: Lens' CreateLoadBalancerPolicy Text

data CreateLoadBalancerPolicyResponse

Constructors

CreateLoadBalancerPolicyResponse