Network.AWS.ELB.V2012_06_01.DeleteLoadBalancer

Description

Deletes the specified load balancer. If attempting to recreate the load balancer, you must reconfigure all the settings. The DNS name associated with a deleted load balancer will no longer be usable. Once deleted, the name and associated DNS record of the load balancer no longer exist and traffic sent to any of its IP addresses will no longer be delivered to back-end instances. To successfully call this API, you must provide the same account credentials as were used to create the load balancer. By design, if the load balancer does not exist or has already been deleted, a call to DeleteLoadBalancer action still succeeds.

Request

data DeleteLoadBalancer

The input for the DeleteLoadBalancer action.

Request constructor

mkDeleteLoadBalancer

Arguments

:: Text
dlbLoadBalancerName

-> DeleteLoadBalancer

Smart constructor for the minimum required parameters to construct a valid DeleteLoadBalancer request.

The fields accessible through corresponding lenses are:

LoadBalancerName :: Text

Request lenses

dlbLoadBalancerName :: Lens' DeleteLoadBalancer Text

The name associated with the load balancer.

Response

data DeleteLoadBalancerResponse

The output for the DeleteLoadBalancer action.

Response constructor

mkDeleteLoadBalancerResponse :: DeleteLoadBalancerResponse

Smart constructor for the minimum required parameters to construct a valid DeleteLoadBalancerResponse response.

This constructor is provided for convenience and testing purposes.