Network.AWS.ELB.V2012_06_01.DeleteLoadBalancerListeners

Description

Deletes listeners from the load balancer for the specified port.

Request

data DeleteLoadBalancerListeners

The input for the DeleteLoadBalancerListeners action.

Request constructor

mkDeleteLoadBalancerListeners

Arguments

:: Text
dlblLoadBalancerName

-> [Integer]
dlblLoadBalancerPorts

-> DeleteLoadBalancerListeners

Smart constructor for the minimum required parameters to construct a valid DeleteLoadBalancerListeners request.

Request lenses

dlblLoadBalancerName :: Lens' DeleteLoadBalancerListeners Text

The mnemonic name associated with the load balancer.

dlblLoadBalancerPorts :: Lens' DeleteLoadBalancerListeners [Integer]

The client port number(s) of the load balancer listener(s) to be removed.

Response

data DeleteLoadBalancerListenersResponse

The output for the DeleteLoadBalancerListeners action.

Response constructor

mkDeleteLoadBalancerListenersResponse :: DeleteLoadBalancerListenersResponse

Smart constructor for the minimum required parameters to construct a valid DeleteLoadBalancerListenersResponse response.

This constructor is provided for convenience and testing purposes.