Network.AWS.ELB.V2012_06_01.DeregisterInstancesFromLoadBalancer

Description

Deregisters instances from the load balancer. Once the instance is deregistered, it will stop receiving traffic from the load balancer. In order to successfully call this API, the same account credentials as those used to create the load balancer must be provided. For more information, see De-register and Register Amazon EC2 Instances in the Elastic Load Balancing Developer Guide. You can use DescribeLoadBalancers to verify if the instance is deregistered from the load balancer. Deregister instance i-e3677ad7 from MyHTTPSLoadBalancer load balancer. https:/elasticloadbalancing.amazonaws.com?Instances.member.1.InstanceId=i-e3677ad7 &LoadBalancerName=MyHTTPSLoadBalancer &Version=2012-06-01 &Action=DeregisterInstancesFromLoadBalancer &AUTHPARAMS i-6ec63d59 83c88b9d-12b7-11e3-8b82-87b12EXAMPLE.

Request

data DeregisterInstancesFromLoadBalancer

The input for the DeregisterInstancesFromLoadBalancer action.

Request constructor

mkDeregisterInstancesFromLoadBalancer

Arguments

:: Text
diflbLoadBalancerName

-> [Instance]
diflbInstances

-> DeregisterInstancesFromLoadBalancer

Smart constructor for the minimum required parameters to construct a valid DeregisterInstancesFromLoadBalancer request.

The fields accessible through corresponding lenses are:

LoadBalancerName :: Text

Instances :: [Instance]

Request lenses

diflbLoadBalancerName :: Lens' DeregisterInstancesFromLoadBalancer Text

The name associated with the load balancer.

diflbInstances :: Lens' DeregisterInstancesFromLoadBalancer [Instance]

A list of EC2 instance IDs consisting of all instances to be deregistered.

Response

data DeregisterInstancesFromLoadBalancerResponse

The output for the DeregisterInstancesFromLoadBalancer action.

Response constructor

mkDeregisterInstancesFromLoadBalancerResponse :: DeregisterInstancesFromLoadBalancerResponse

Smart constructor for the minimum required parameters to construct a valid DeregisterInstancesFromLoadBalancerResponse response.

This constructor is provided for convenience and testing purposes.

The fields accessible through corresponding lenses are:

Instances :: [Instance]

Response lenses

diflbrInstances :: Lens' DeregisterInstancesFromLoadBalancerResponse [Instance]

An updated list of remaining instances registered with the load balancer.