Network.AWS.ELB.V2012_06_01.DescribeLoadBalancerAttributes

Description

Returns detailed information about all of the attributes associated with the specified load balancer. https:/elasticloadbalancing.amazonaws.com?LoadBalancerName=my-test-loadbalancer &Version=2012-06-01 &Action=DescribeLoadBalancerAttributes &AUTHPARAMS true my-loadbalancer-logs testprefix 5 30 true true 60 83c88b9d-12b7-11e3-8b82-87b12EXAMPLE.

Request

data DescribeLoadBalancerAttributes

The input for the DescribeLoadBalancerAttributes action.

Request constructor

mkDescribeLoadBalancerAttributes

Arguments

:: Text
dlbaLoadBalancerName

-> DescribeLoadBalancerAttributes

Smart constructor for the minimum required parameters to construct a valid DescribeLoadBalancerAttributes request.

Request lenses

dlbaLoadBalancerName :: Lens' DescribeLoadBalancerAttributes Text

The name of the load balancer.

Response

data DescribeLoadBalancerAttributesResponse

The following element is returned in a structure named DescribeLoadBalancerAttributesResult.

Response lenses

dlbarsLoadBalancerAttributes :: Lens' DescribeLoadBalancerAttributesResponse (Maybe LoadBalancerAttributes)

The load balancer attributes structure.