Network.AWS.ELB.V2012_06_01.DescribeLoadBalancers

Description

Returns detailed configuration information for all the load balancers created for the account. If you specify load balancer names, the action returns configuration information of the specified load balancers. In order to retrieve this information, you must provide the same account credentials that was used to create the load balancer. Description of a specified load balancer https:/elasticloadbalancing.amazonaws.com?LoadBalancerNames.member.1=MyLoadBalancer &Version=2012-06-01 &Action=DescribeLoadBalancers &AUTHPARAMS MyLoadBalancer 2013-05-24T21:15:31.280Z 90 HTTP:80/ 2 60 10 HTTP 80 HTTP 80 i-e4cbe38d us-east-1a ZZZZZZZZZZZ123X MyLoadBalancer-123456789.us-east-1.elb.amazonaws.com internet-facing amazon-elb amazon-elb-sg MyLoadBalancer-123456789.us-east-1.elb.amazonaws.com 83c88b9d-12b7-11e3-8b82-87b12EXAMPLE.

Documentation

describeLoadBalancers :: DescribeLoadBalancers

Minimum specification for a DescribeLoadBalancers request.

data DescribeLoadBalancers

Constructors

DescribeLoadBalancers

Fields

_dapiLoadBalancerNames :: [Text]

A list of load balancer names associated with the account.

_dapiMarker :: Maybe Text

An optional parameter used for pagination of results from this call. If specified, the response includes only records beyond the marker.

_dapiPageSize :: Maybe Integer

The number of results returned in each page. The default is 400. You cannot specify a page size greater than 400 or less than 1.

dapiPageSize :: Lens' DescribeLoadBalancers (Maybe Integer)

dapiMarker :: Lens' DescribeLoadBalancers (Maybe Text)

dapiLoadBalancerNames :: Lens' DescribeLoadBalancers [Text]

data DescribeLoadBalancersResponse

Constructors

DescribeLoadBalancersResponse

Fields

_dapoLoadBalancerDescriptions :: [LoadBalancerDescription]

A list of load balancer description structures.

_dapoNextMarker :: Maybe Text

Specifies the value of next marker if the request returned more than one page of results.

dapoNextMarker :: Lens' DescribeLoadBalancersResponse (Maybe Text)

dapoLoadBalancerDescriptions :: Lens' DescribeLoadBalancersResponse [LoadBalancerDescription]