Network.AWS.ELB.V2012_06_01.RemoveTags

Description

Removes one or more tags from the specified load balancer. Remove Two Tag Keys from the Load Balancer https:/elasticloadbalancing.amazonaws.com?LoadBalancerName=my-test-loadbalancer &Tags.member.1.Key=owner &Tags.member.2.Key=project &Action=RemoveTags &Version=2012-06-01 &AUTHPARAMS 83c88b9d-12b7-11e3-8b82-87b12EXAMPLE.

Documentation

data RemoveTags

Constructors

RemoveTags

Fields

_rtiLoadBalancerNames :: [Text]

The name of the load balancer. You can specify a maximum of one load balancer name.

_rtiTags :: [TagKeyOnly]

A list of tag keys to remove.

rtiTags :: Lens' RemoveTags [TagKeyOnly]

rtiLoadBalancerNames :: Lens' RemoveTags [Text]

data RemoveTagsResponse

Constructors

RemoveTagsResponse