Network.AWS.ELB.V2012_06_01.SetLoadBalancerListenerSSLCertificate

Description

Sets the certificate that terminates the specified listener's SSL connections. The specified certificate replaces any prior certificate that was used on the same load balancer and port. For more information on updating your SSL certificate, see Updating an SSL Certificate for a Load Balancer in the Elastic Load Balancing Developer Guide. https:/elasticloadbalancing.amazonaws.com?LoadBalancerName=MyInternalLoadBalancer &SSLCertificateId=arn:aws:iam::123456789012:server-certificate/testcert &LoadBalancerPort=443 &Version=2012-06-01 &Action=SetLoadBalancerListenerSSLCertificate &AUTHPARAMS 83c88b9d-12b7-11e3-8b82-87b12EXAMPLE.

Documentation

data SetLoadBalancerListenerSSLCertificate

Constructors

SetLoadBalancerListenerSSLCertificate

Fields

_slblsslciLoadBalancerName :: Text

The name of the load balancer.

_slblsslciLoadBalancerPort :: Integer

The port that uses the specified SSL certificate.

_slblsslciSSLCertificateId :: Text

The Amazon Resource Number (ARN) of the SSL certificate chain to use. For more information on SSL certificates, see Managing Server Certificates in the AWS Identity and Access Management User Guide.

slblsslciSSLCertificateId :: Lens' SetLoadBalancerListenerSSLCertificate Text

slblsslciLoadBalancerPort :: Lens' SetLoadBalancerListenerSSLCertificate Integer

slblsslciLoadBalancerName :: Lens' SetLoadBalancerListenerSSLCertificate Text

data SetLoadBalancerListenerSSLCertificateResponse

Constructors

SetLoadBalancerListenerSSLCertificateResponse