Network.AWS.ELB.V2012_06_01.SetLoadBalancerPoliciesForBackendServer

Description

Replaces the current set of policies associated with a port on which the back-end server is listening with a new set of policies. After the policies have been created using CreateLoadBalancerPolicy, they can be applied here as a list. At this time, only the back-end server authentication policy type can be applied to the back-end ports; this policy type is composed of multiple public key policies. The SetLoadBalancerPoliciesForBackendServer replaces the current set of policies associated with the specified instance port. Every time you use this action to enable the policies, use the PolicyNames parameter to list all the policies you want to enable. https:/elasticloadbalancing.amazonaws.com?InstancePort=80 &PolicyNames.member.1=EnableProxyProtocol &PolicyNames.member.2=MyPolicyName2 &PolicyNames.member.3=MyPolicyName3 &LoadBalancerName=my-test-loadbalancer &Version=2012-06-01 &Action=SetLoadBalancerPoliciesForBackendServer &AUTHPARAMS 0eb9b381-dde0-11e2-8d78-6ddbaEXAMPLE You can use DescribeLoadBalancers or DescribeLoadBalancerPolicies action to verify that the policy has been associated with the back-end server.

Request

data SetLoadBalancerPoliciesForBackendServer

The input for the SetLoadBalancerPoliciesForBackendServer action.

Request constructor

mkSetLoadBalancerPoliciesForBackendServer

Arguments

:: Text
slbpfbsLoadBalancerName

-> Integer
slbpfbsInstancePort

-> [Text]
slbpfbsPolicyNames

-> SetLoadBalancerPoliciesForBackendServer

Smart constructor for the minimum required parameters to construct a valid SetLoadBalancerPoliciesForBackendServer request.

Request lenses

slbpfbsLoadBalancerName :: Lens' SetLoadBalancerPoliciesForBackendServer Text

The mnemonic name associated with the load balancer. This name must be unique within the set of your load balancers.

slbpfbsInstancePort :: Lens' SetLoadBalancerPoliciesForBackendServer Integer

The port number associated with the back-end server.

slbpfbsPolicyNames :: Lens' SetLoadBalancerPoliciesForBackendServer [Text]

List of policy names to be set. If the list is empty, then all current polices are removed from the back-end server.

Response

data SetLoadBalancerPoliciesForBackendServerResponse

The output for the SetLoadBalancerPoliciesForBackendServer action.