Network.AWS.ELB.V2012_06_01.SetLoadBalancerPoliciesOfListener

Description

Associates, updates, or disables a policy with a listener on the load balancer. You can associate multiple policies with a listener. Associate MySSLNegotiationPolicy with the load balancer port 443 on the MyInternalLoadbalancer load balancer. https:/elasticloadbalancing.amazonaws.com?PolicyNames.member.1=MySSLNegotiationPolicy &LoadBalancerName=MyInternalLoadBalancer &LoadBalancerPort=443 &Version=2012-06-01 &Action=SetLoadBalancerPoliciesOfListener &AUTHPARAMS azonaws.comdoc2012-06-01/"> 07b1ecbc-1100-11e3-acaf-dd7edEXAMPLE.

Documentation

data SetLoadBalancerPoliciesOfListener

Constructors

SetLoadBalancerPoliciesOfListener

Fields

_slbpoliLoadBalancerName :: Text

The name of the load balancer.

_slbpoliLoadBalancerPort :: Integer

The external port of the load balancer to associate the policy.

_slbpoliPolicyNames :: [Text]

List of policies to be associated with the listener. If the list is empty, the current policy is removed from the listener.

slbpoliPolicyNames :: Lens' SetLoadBalancerPoliciesOfListener [Text]

slbpoliLoadBalancerPort :: Lens' SetLoadBalancerPoliciesOfListener Integer

slbpoliLoadBalancerName :: Lens' SetLoadBalancerPoliciesOfListener Text

data SetLoadBalancerPoliciesOfListenerResponse

Constructors

SetLoadBalancerPoliciesOfListenerResponse