Network.AWS.ELB.V2012_06_01.Types

Description

Elastic Load Balancing automatically distributes incoming application traffic across multiple Amazon EC2 instances. It enables you to achieve greater levels of fault tolerance in your applications, seamlessly providing the required amount of load balancing capacity needed to distribute application traffic.

Service

data ELB

Supported version (2012-06-01) of the Elastic Load Balancing service.

XML

xmlOptions :: Tagged a XMLOptions

ConnectionSettings

data ConnectionSettings

The name of the load balancer attribute. By default, Elastic Load Balancing maintains a 60-second idle connection timeout for both front-end and back-end connections of your load balancer. If the ConnectionSettings attribute is set, Elastic Load Balancing will allow the connections to remain idle (no data is sent over the connection) for the specified duration. For more information, see Configure Idle Connection Timeout.

mkConnectionSettings

Arguments

:: Integer
csIdleTimeout

-> ConnectionSettings

Smart constructor for the minimum required fields to construct a valid ConnectionSettings data type to populate a request.

csIdleTimeout :: Lens' ConnectionSettings Integer

Specifies the time (in seconds) the connection is allowed to be idle (no data has been sent over the connection) before it is closed by the load balancer.

CrossZoneLoadBalancing

data CrossZoneLoadBalancing

The name of the load balancer attribute. If enabled, the load balancer routes the request traffic evenly across all back-end instances regardless of the Availability Zones. For more information, see Enable Cross-Zone Load Balancing.

mkCrossZoneLoadBalancing

Arguments

:: Bool
czlbEnabled

-> CrossZoneLoadBalancing

Smart constructor for the minimum required fields to construct a valid CrossZoneLoadBalancing data type to populate a request.

czlbEnabled :: Lens' CrossZoneLoadBalancing Bool

Specifies whether cross-zone load balancing is enabled for the load balancer.

Instance

data Instance

The Instance data type.

mkInstance :: Instance

Smart constructor for the minimum required fields to construct a valid Instance data type to populate a request.

iInstanceId :: Lens' Instance (Maybe Text)

Provides an EC2 instance ID.

TagKeyOnly

data TagKeyOnly

The key of a tag to be removed.

mkTagKeyOnly :: TagKeyOnly

Smart constructor for the minimum required fields to construct a valid TagKeyOnly data type to populate a request.

tkoKey :: Lens' TagKeyOnly (Maybe Text)

The name of the key.

AccessLog

data AccessLog

The name of the load balancer attribute. If enabled, the load balancer captures detailed information of all the requests and delivers the information to the Amazon S3 bucket that you specify. For more information, see Enable Access Logs.

mkAccessLog

Arguments

:: Bool
alEnabled

-> AccessLog

Smart constructor for the minimum required fields to construct a valid AccessLog data type to populate a request.

alEnabled :: Lens' AccessLog Bool

Specifies whether access log is enabled for the load balancer.

alS3BucketName :: Lens' AccessLog (Maybe Text)

The name of the Amazon S3 bucket where the access logs are stored.

alEmitInterval :: Lens' AccessLog (Maybe Integer)

The interval for publishing the access logs. You can specify an interval of either 5 minutes or 60 minutes. Default: 60 minutes.

alS3BucketPrefix :: Lens' AccessLog (Maybe Text)

The logical hierarchy you created for your Amazon S3 bucket, for example my-bucket-prefix/prod. If the prefix is not provided, the log is placed at the root level of the bucket.

AppCookieStickinessPolicy

data AppCookieStickinessPolicy

The AppCookieStickinessPolicy data type.

mkAppCookieStickinessPolicy :: AppCookieStickinessPolicy

Smart constructor for the minimum required fields to construct a valid AppCookieStickinessPolicy data type to populate a request.

acspPolicyName :: Lens' AppCookieStickinessPolicy (Maybe Text)

The mnemonic name for the policy being created. The name must be unique within a set of policies for this load balancer.

acspCookieName :: Lens' AppCookieStickinessPolicy (Maybe Text)

The name of the application cookie used for stickiness.

BackendServerDescription

data BackendServerDescription

This data type is used as a response element in the DescribeLoadBalancers action to describe the configuration of the back-end server.

mkBackendServerDescription :: BackendServerDescription

Smart constructor for the minimum required fields to construct a valid BackendServerDescription data type to populate a request.

bsdInstancePort :: Lens' BackendServerDescription (Maybe Integer)

Provides the port on which the back-end server is listening.

bsdPolicyNames :: Lens' BackendServerDescription [Text]

Provides a list of policy names enabled for the back-end server.

ConnectionDraining

data ConnectionDraining

The name of the load balancer attribute. If enabled, the load balancer allows existing requests to complete before the load balancer shifts traffic away from a deregistered or unhealthy back-end instance. For more information, see Enable Connection Draining.

mkConnectionDraining

Arguments

:: Bool
cdEnabled

-> ConnectionDraining

Smart constructor for the minimum required fields to construct a valid ConnectionDraining data type to populate a request.

cdEnabled :: Lens' ConnectionDraining Bool

Specifies whether connection draining is enabled for the load balancer.

cdTimeout :: Lens' ConnectionDraining (Maybe Integer)

Specifies the maximum time (in seconds) to keep the existing connections open before deregistering the instances.

HealthCheck

data HealthCheck

A structure containing the configuration information for the new healthcheck.

mkHealthCheck

Arguments

:: Text
hcTarget

-> Integer
hcInterval

-> Integer
hcTimeout

-> Integer
hcUnhealthyThreshold

-> Integer
hcHealthyThreshold

-> HealthCheck

Smart constructor for the minimum required fields to construct a valid HealthCheck data type to populate a request.

hcTarget :: Lens' HealthCheck Text

Specifies the instance being checked. The protocol is either TCP, HTTP, HTTPS, or SSL. The range of valid ports is one (1) through 65535. TCP is the default, specified as a TCP: port pair, for example TCP:5000. In this case a healthcheck simply attempts to open a TCP connection to the instance on the specified port. Failure to connect within the configured timeout is considered unhealthy. SSL is also specified as SSL: port pair, for example, SSL:5000. For HTTP or HTTPS protocol, the situation is different. You have to include a ping path in the string. HTTP is specified as a HTTP:port;/;PathToPing; grouping, for example "HTTP:80weatheruswaseattle". In this case, a HTTP GET request is issued to the instance on the given port and path. Any answer other than "200 OK" within the timeout period is considered unhealthy. The total length of the HTTP ping target needs to be 1024 16-bit Unicode characters or less.

hcInterval :: Lens' HealthCheck Integer

Specifies the approximate interval, in seconds, between health checks of an individual instance.

hcTimeout :: Lens' HealthCheck Integer

Specifies the amount of time, in seconds, during which no response means a failed health probe. This value must be less than the Interval value.

hcUnhealthyThreshold :: Lens' HealthCheck Integer

Specifies the number of consecutive health probe failures required before moving the instance to the Unhealthy state.

hcHealthyThreshold :: Lens' HealthCheck Integer

Specifies the number of consecutive health probe successes required before moving the instance to the Healthy state.

InstanceState

data InstanceState

The InstanceState data type.

mkInstanceState :: InstanceState

Smart constructor for the minimum required fields to construct a valid InstanceState data type to populate a request.

isInstanceId :: Lens' InstanceState (Maybe Text)

Provides an EC2 instance ID.

isState :: Lens' InstanceState (Maybe Text)

Specifies the current state of the instance. Valid value: InService|OutOfService|Unknown.

isReasonCode :: Lens' InstanceState (Maybe Text)

Provides information about the cause of OutOfService instances. Specifically, it indicates whether the cause is Elastic Load Balancing or the instance behind the load balancer. Valid value: ELB|Instance|N/A.

isDescription :: Lens' InstanceState (Maybe Text)

Provides a description of the instance state.

LBCookieStickinessPolicy

data LBCookieStickinessPolicy

The LBCookieStickinessPolicy data type.

mkLBCookieStickinessPolicy :: LBCookieStickinessPolicy

Smart constructor for the minimum required fields to construct a valid LBCookieStickinessPolicy data type to populate a request.

lbcspPolicyName :: Lens' LBCookieStickinessPolicy (Maybe Text)

The name for the policy being created. The name must be unique within the set of policies for this load balancer.

lbcspCookieExpirationPeriod :: Lens' LBCookieStickinessPolicy (Maybe Integer)

The time period in seconds after which the cookie should be considered stale. Not specifying this parameter indicates that the stickiness session will last for the duration of the browser session.

Listener

data Listener

The Listener data type.

mkListener

Arguments

:: Text
lProtocol

-> Integer
lLoadBalancerPort

-> Integer
lInstancePort

-> Listener

Smart constructor for the minimum required fields to construct a valid Listener data type to populate a request.

lProtocol :: Lens' Listener Text

Specifies the load balancer transport protocol to use for routing - HTTP, HTTPS, TCP or SSL. This property cannot be modified for the life of the load balancer.

lLoadBalancerPort :: Lens' Listener Integer

Specifies the external load balancer port number. This property cannot be modified for the life of the load balancer.

lInstanceProtocol :: Lens' Listener (Maybe Text)

Specifies the protocol to use for routing traffic to back-end instances - HTTP, HTTPS, TCP, or SSL. This property cannot be modified for the life of the load balancer. If the front-end protocol is HTTP or HTTPS, InstanceProtocol has to be at the same protocol layer, i.e., HTTP or HTTPS. Likewise, if the front-end protocol is TCP or SSL, InstanceProtocol has to be TCP or SSL. If there is another listener with the same InstancePort whose InstanceProtocol is secure, i.e., HTTPS or SSL, the listener's InstanceProtocol has to be secure, i.e., HTTPS or SSL. If there is another listener with the same InstancePort whose InstanceProtocol is HTTP or TCP, the listener's InstanceProtocol must be either HTTP or TCP.

lInstancePort :: Lens' Listener Integer

Specifies the TCP port on which the instance server is listening. This property cannot be modified for the life of the load balancer.

lSSLCertificateId :: Lens' Listener (Maybe Text)

The ARN string of the server certificate. To get the ARN of the server certificate, call the AWS Identity and Access Management UploadServerCertificate API.

ListenerDescription

data ListenerDescription

The ListenerDescription data type.

mkListenerDescription :: ListenerDescription

Smart constructor for the minimum required fields to construct a valid ListenerDescription data type to populate a request.

ldListener :: Lens' ListenerDescription (Maybe Listener)

The Listener data type.

ldPolicyNames :: Lens' ListenerDescription [Text]

A list of policies enabled for this listener. An empty list indicates that no policies are enabled.

LoadBalancerAttributes

data LoadBalancerAttributes

The load balancer attributes structure.

mkLoadBalancerAttributes :: LoadBalancerAttributes

Smart constructor for the minimum required fields to construct a valid LoadBalancerAttributes data type to populate a request.

lbaCrossZoneLoadBalancing :: Lens' LoadBalancerAttributes (Maybe CrossZoneLoadBalancing)

The name of the load balancer attribute. If enabled, the load balancer routes the request traffic evenly across all back-end instances regardless of the Availability Zones. For more information, see Enable Cross-Zone Load Balancing.

lbaAccessLog :: Lens' LoadBalancerAttributes (Maybe AccessLog)

The name of the load balancer attribute. If enabled, the load balancer captures detailed information of all the requests and delivers the information to the Amazon S3 bucket that you specify. For more information, see Enable Access Logs.

lbaConnectionDraining :: Lens' LoadBalancerAttributes (Maybe ConnectionDraining)

The name of the load balancer attribute. If enabled, the load balancer allows existing requests to complete before the load balancer shifts traffic away from a deregistered or unhealthy back-end instance. For more information, see Enable Connection Draining.

lbaConnectionSettings :: Lens' LoadBalancerAttributes (Maybe ConnectionSettings)

The name of the load balancer attribute. By default, Elastic Load Balancing maintains a 60-second idle connection timeout for both front-end and back-end connections of your load balancer. If the ConnectionSettings attribute is set, Elastic Load Balancing will allow the connections to remain idle (no data is sent over the connection) for the specified duration. For more information, see Configure Idle Connection Timeout.

LoadBalancerDescription

data LoadBalancerDescription

Contains the result of a successful invocation of DescribeLoadBalancers.

mkLoadBalancerDescription :: LoadBalancerDescription

Smart constructor for the minimum required fields to construct a valid LoadBalancerDescription data type to populate a request.

lbdLoadBalancerName :: Lens' LoadBalancerDescription (Maybe Text)

Specifies the name associated with the load balancer.

lbdDNSName :: Lens' LoadBalancerDescription (Maybe Text)

Specifies the external DNS name associated with the load balancer.

lbdCanonicalHostedZoneName :: Lens' LoadBalancerDescription (Maybe Text)

Provides the name of the Amazon Route 53 hosted zone that is associated with the load balancer. For information on how to associate your load balancer with a hosted zone, go to Using Domain Names With Elastic Load Balancing in the Elastic Load Balancing Developer Guide.

lbdCanonicalHostedZoneNameID :: Lens' LoadBalancerDescription (Maybe Text)

Provides the ID of the Amazon Route 53 hosted zone name that is associated with the load balancer. For information on how to associate or disassociate your load balancer with a hosted zone, go to Using Domain Names With Elastic Load Balancing in the Elastic Load Balancing Developer Guide.

lbdListenerDescriptions :: Lens' LoadBalancerDescription [ListenerDescription]

LoadBalancerPort, InstancePort, Protocol, InstanceProtocol, and PolicyNames are returned in a list of tuples in the ListenerDescriptions element.

lbdPolicies :: Lens' LoadBalancerDescription (Maybe Policies)

Provides a list of policies defined for the load balancer.

lbdBackendServerDescriptions :: Lens' LoadBalancerDescription [BackendServerDescription]

Contains a list of back-end server descriptions.

lbdAvailabilityZones :: Lens' LoadBalancerDescription [Text]

Specifies a list of Availability Zones.

lbdSubnets :: Lens' LoadBalancerDescription [Text]

Provides a list of VPC subnet IDs for the load balancer.

lbdVPCId :: Lens' LoadBalancerDescription (Maybe Text)

Provides the ID of the VPC attached to the load balancer.

lbdInstances :: Lens' LoadBalancerDescription [Instance]

Provides a list of EC2 instance IDs for the load balancer.

lbdHealthCheck :: Lens' LoadBalancerDescription (Maybe HealthCheck)

Specifies information regarding the various health probes conducted on the load balancer.

lbdSourceSecurityGroup :: Lens' LoadBalancerDescription (Maybe SourceSecurityGroup)

The security group that you can use as part of your inbound rules for your load balancer's back-end Amazon EC2 application instances. To only allow traffic from load balancers, add a security group rule to your back end instance that specifies this source security group as the inbound source.

lbdSecurityGroups :: Lens' LoadBalancerDescription [Text]

The security groups the load balancer is a member of (VPC only).

lbdCreatedTime :: Lens' LoadBalancerDescription (Maybe ISO8601)

Provides the date and time the load balancer was created.

lbdScheme :: Lens' LoadBalancerDescription (Maybe Text)

Specifies the type of load balancer. If the Scheme is internet-facing, the load balancer has a publicly resolvable DNS name that resolves to public IP addresses. If the Scheme is internal, the load balancer has a publicly resolvable DNS name that resolves to private IP addresses. This option is only available for load balancers attached to an Amazon VPC.

Policies

data Policies

Provides a list of policies defined for the load balancer.

mkPolicies :: Policies

Smart constructor for the minimum required fields to construct a valid Policies data type to populate a request.

pAppCookieStickinessPolicies :: Lens' Policies [AppCookieStickinessPolicy]

A list of the AppCookieStickinessPolicy objects created with CreateAppCookieStickinessPolicy.

pLBCookieStickinessPolicies :: Lens' Policies [LBCookieStickinessPolicy]

A list of LBCookieStickinessPolicy objects created with CreateAppCookieStickinessPolicy.

pOtherPolicies :: Lens' Policies [Text]

A list of policy names other than the stickiness policies.

PolicyAttribute

data PolicyAttribute

The PolicyAttribute data type. This data type contains a key/value pair that defines properties of a specific policy.

mkPolicyAttribute :: PolicyAttribute

Smart constructor for the minimum required fields to construct a valid PolicyAttribute data type to populate a request.

paAttributeName :: Lens' PolicyAttribute (Maybe Text)

The name of the attribute associated with the policy.

paAttributeValue :: Lens' PolicyAttribute (Maybe Text)

The value of the attribute associated with the policy.

PolicyAttributeDescription

data PolicyAttributeDescription

The PolicyAttributeDescription data type. This data type is used to describe the attributes and values associated with a policy.

mkPolicyAttributeDescription :: PolicyAttributeDescription

Smart constructor for the minimum required fields to construct a valid PolicyAttributeDescription data type to populate a request.

padAttributeName :: Lens' PolicyAttributeDescription (Maybe Text)

The name of the attribute associated with the policy.

padAttributeValue :: Lens' PolicyAttributeDescription (Maybe Text)

The value of the attribute associated with the policy.

PolicyAttributeTypeDescription

data PolicyAttributeTypeDescription

The PolicyAttributeTypeDescription data type. This data type is used to describe values that are acceptable for the policy attribute.

mkPolicyAttributeTypeDescription :: PolicyAttributeTypeDescription

Smart constructor for the minimum required fields to construct a valid PolicyAttributeTypeDescription data type to populate a request.

patdAttributeName :: Lens' PolicyAttributeTypeDescription (Maybe Text)

The name of the attribute associated with the policy type.

patdAttributeType :: Lens' PolicyAttributeTypeDescription (Maybe Text)

The type of attribute. For example, Boolean, Integer, etc.

patdDescription :: Lens' PolicyAttributeTypeDescription (Maybe Text)

A human-readable description of the attribute.

patdDefaultValue :: Lens' PolicyAttributeTypeDescription (Maybe Text)

The default value of the attribute, if applicable.

patdCardinality :: Lens' PolicyAttributeTypeDescription (Maybe Text)

The cardinality of the attribute. Valid Values: ONE(1) : Single value required ZERO_OR_ONE(0..1) : Up to one value can be supplied ZERO_OR_MORE(0..*) : Optional. Multiple values are allowed ONE_OR_MORE(1..*0) : Required. Multiple values are allowed.

PolicyDescription

data PolicyDescription

The PolicyDescription data type.

mkPolicyDescription :: PolicyDescription

Smart constructor for the minimum required fields to construct a valid PolicyDescription data type to populate a request.

pdPolicyName :: Lens' PolicyDescription (Maybe Text)

The name of the policy associated with the load balancer.

pdPolicyTypeName :: Lens' PolicyDescription (Maybe Text)

The name of the policy type associated with the load balancer.

pdPolicyAttributeDescriptions :: Lens' PolicyDescription [PolicyAttributeDescription]

A list of policy attribute description structures.

PolicyTypeDescription

data PolicyTypeDescription

The PolicyTypeDescription data type.

mkPolicyTypeDescription :: PolicyTypeDescription

Smart constructor for the minimum required fields to construct a valid PolicyTypeDescription data type to populate a request.

ptdPolicyTypeName :: Lens' PolicyTypeDescription (Maybe Text)

The name of the policy type.

ptdDescription :: Lens' PolicyTypeDescription (Maybe Text)

A human-readable description of the policy type.

ptdPolicyAttributeTypeDescriptions :: Lens' PolicyTypeDescription [PolicyAttributeTypeDescription]

The description of the policy attributes associated with the load balancer policies defined by the Elastic Load Balancing service.

SourceSecurityGroup

data SourceSecurityGroup

The security group that you can use as part of your inbound rules for your load balancer's back-end Amazon EC2 application instances. To only allow traffic from load balancers, add a security group rule to your back end instance that specifies this source security group as the inbound source.

mkSourceSecurityGroup :: SourceSecurityGroup

Smart constructor for the minimum required fields to construct a valid SourceSecurityGroup data type to populate a request.

ssgOwnerAlias :: Lens' SourceSecurityGroup (Maybe Text)

Owner of the source security group. Use this value for the --source-group-user parameter of the ec2-authorize command in the Amazon EC2 command line tool.

ssgGroupName :: Lens' SourceSecurityGroup (Maybe Text)

Name of the source security group. Use this value for the --source-group parameter of the ec2-authorize command in the Amazon EC2 command line tool.

Tag

data Tag

Metadata assigned to a load balancer consisting of key-value pair. For more information, see Tagging in the Elastic Load Balancing Developer Guide.

mkTag

Arguments

:: Text
tKey

-> Tag

Smart constructor for the minimum required fields to construct a valid Tag data type to populate a request.

tKey :: Lens' Tag Text

The key of the tag.

tValue :: Lens' Tag (Maybe Text)

The value of the tag.

TagDescription

data TagDescription

The descriptions of all the tags associated with load balancer.

mkTagDescription :: TagDescription

Smart constructor for the minimum required fields to construct a valid TagDescription data type to populate a request.

tdLoadBalancerName :: Lens' TagDescription (Maybe Text)

The name of the load balancer.

tdTags :: Lens' TagDescription (Maybe (List1 Tag))

List of tags associated with the load balancer.