Network.AWS.ELB.AddTags

Description

Adds the specified tags to the specified load balancer. Each load balancer can have a maximum of 10 tags.

Each tag consists of a key and an optional value. If a tag with the same key is already associated with the load balancer, AddTags updates its value.

For more information, see Tag Your Classic Load Balancer in the Classic Load Balancer Guide .

Creating a Request

addTags

Arguments

:: NonEmpty Tag
atTags

-> AddTags

Creates a value of AddTags with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

atLoadBalancerNames - The name of the load balancer. You can specify one load balancer only.

atTags - The tags.

data AddTags

Contains the parameters for AddTags.

See: addTags smart constructor.

Request Lenses

atLoadBalancerNames :: Lens' AddTags [Text]

The name of the load balancer. You can specify one load balancer only.

atTags :: Lens' AddTags (NonEmpty Tag)

The tags.

Destructuring the Response

addTagsResponse

Arguments

:: Int
atrsResponseStatus

-> AddTagsResponse

Creates a value of AddTagsResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

atrsResponseStatus - -- | The response status code.

data AddTagsResponse

Contains the output of AddTags.

See: addTagsResponse smart constructor.

Response Lenses

atrsResponseStatus :: Lens' AddTagsResponse Int

- | The response status code.