Network.AWS.ELB.ApplySecurityGroupsToLoadBalancer

Description

Associates one or more security groups with your load balancer in a virtual private cloud (VPC). The specified security groups override the previously associated security groups.

For more information, see Security Groups for Load Balancers in a VPC in the Classic Load Balancer Guide .

Creating a Request

applySecurityGroupsToLoadBalancer

Arguments

:: Text
asgtlbLoadBalancerName

-> ApplySecurityGroupsToLoadBalancer

Creates a value of ApplySecurityGroupsToLoadBalancer with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

asgtlbLoadBalancerName - The name of the load balancer.

asgtlbSecurityGroups - The IDs of the security groups to associate with the load balancer. Note that you cannot specify the name of the security group.

data ApplySecurityGroupsToLoadBalancer

Contains the parameters for ApplySecurityGroupsToLoadBalancer.

See: applySecurityGroupsToLoadBalancer smart constructor.

Request Lenses

asgtlbLoadBalancerName :: Lens' ApplySecurityGroupsToLoadBalancer Text

The name of the load balancer.

asgtlbSecurityGroups :: Lens' ApplySecurityGroupsToLoadBalancer [Text]

The IDs of the security groups to associate with the load balancer. Note that you cannot specify the name of the security group.

Destructuring the Response

applySecurityGroupsToLoadBalancerResponse

Arguments

:: Int
asgtlbrsResponseStatus

-> ApplySecurityGroupsToLoadBalancerResponse

Creates a value of ApplySecurityGroupsToLoadBalancerResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

asgtlbrsSecurityGroups - The IDs of the security groups associated with the load balancer.

asgtlbrsResponseStatus - -- | The response status code.

data ApplySecurityGroupsToLoadBalancerResponse

Contains the output of ApplySecurityGroupsToLoadBalancer.

See: applySecurityGroupsToLoadBalancerResponse smart constructor.

Response Lenses

asgtlbrsSecurityGroups :: Lens' ApplySecurityGroupsToLoadBalancerResponse [Text]

The IDs of the security groups associated with the load balancer.

asgtlbrsResponseStatus :: Lens' ApplySecurityGroupsToLoadBalancerResponse Int

- | The response status code.