Network.AWS.ELB.ApplySecurityGroupsToLoadBalancer

Description

Associates one or more security groups with your load balancer in Amazon Virtual Private Cloud (Amazon VPC). The provided security group IDs will override any currently applied security groups.

For more information, see Manage Security Groups in Amazon VPC in the Elastic Load Balancing Developer Guide.

http://docs.aws.amazon.com/ElasticLoadBalancing/latest/APIReference/API_ApplySecurityGroupsToLoadBalancer.html

Request

data ApplySecurityGroupsToLoadBalancer

Request constructor

applySecurityGroupsToLoadBalancer

Arguments

:: Text
asgtlbLoadBalancerName

-> ApplySecurityGroupsToLoadBalancer

ApplySecurityGroupsToLoadBalancer constructor.

The fields accessible through corresponding lenses are:

asgtlbLoadBalancerName :: Text

asgtlbSecurityGroups :: [Text]

Request lenses

asgtlbLoadBalancerName :: Lens' ApplySecurityGroupsToLoadBalancer Text

The name associated with the load balancer. The name must be unique within the set of load balancers associated with your AWS account.

asgtlbSecurityGroups :: Lens' ApplySecurityGroupsToLoadBalancer [Text]

A list of security group IDs to associate with your load balancer in VPC. The security group IDs must be provided as the ID and not the security group name (For example, sg-1234).

Response

data ApplySecurityGroupsToLoadBalancerResponse

Response constructor

applySecurityGroupsToLoadBalancerResponse :: ApplySecurityGroupsToLoadBalancerResponse

ApplySecurityGroupsToLoadBalancerResponse constructor.

The fields accessible through corresponding lenses are:

asgtlbrSecurityGroups :: [Text]

Response lenses

asgtlbrSecurityGroups :: Lens' ApplySecurityGroupsToLoadBalancerResponse [Text]

A list of security group IDs associated with your load balancer.