Network.AWS.ELB.AttachLoadBalancerToSubnets

Description

Adds one or more subnets to the set of configured subnets for the specified load balancer.

The load balancer evenly distributes requests across all registered subnets. For more information, see Add or Remove Subnets for Your Load Balancer in a VPC in the Classic Load Balancer Guide .

Creating a Request

attachLoadBalancerToSubnets

Arguments

:: Text
albtsLoadBalancerName

-> AttachLoadBalancerToSubnets

Creates a value of AttachLoadBalancerToSubnets with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

albtsLoadBalancerName - The name of the load balancer.

albtsSubnets - The IDs of the subnets to add. You can add only one subnet per Availability Zone.

data AttachLoadBalancerToSubnets

Contains the parameters for AttachLoaBalancerToSubnets.

See: attachLoadBalancerToSubnets smart constructor.

Request Lenses

albtsLoadBalancerName :: Lens' AttachLoadBalancerToSubnets Text

The name of the load balancer.

albtsSubnets :: Lens' AttachLoadBalancerToSubnets [Text]

The IDs of the subnets to add. You can add only one subnet per Availability Zone.

Destructuring the Response

attachLoadBalancerToSubnetsResponse

Arguments

:: Int
albtsrsResponseStatus

-> AttachLoadBalancerToSubnetsResponse

Creates a value of AttachLoadBalancerToSubnetsResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

albtsrsSubnets - The IDs of the subnets attached to the load balancer.

albtsrsResponseStatus - -- | The response status code.

data AttachLoadBalancerToSubnetsResponse

Contains the output of AttachLoadBalancerToSubnets.

See: attachLoadBalancerToSubnetsResponse smart constructor.

Response Lenses

albtsrsSubnets :: Lens' AttachLoadBalancerToSubnetsResponse [Text]

The IDs of the subnets attached to the load balancer.

albtsrsResponseStatus :: Lens' AttachLoadBalancerToSubnetsResponse Int

- | The response status code.