Network.AWS.ELB.AttachLoadBalancerToSubnets

Description

Adds one or more subnets to the set of configured subnets in the Amazon Virtual Private Cloud (Amazon VPC) for the load balancer.

The load balancers evenly distribute requests across all of the registered subnets. For more information, see Deploy Elastic Load Balancing in Amazon VPC in the Elastic Load Balancing Developer Guide.

http://docs.aws.amazon.com/ElasticLoadBalancing/latest/APIReference/API_AttachLoadBalancerToSubnets.html

Request

data AttachLoadBalancerToSubnets

Request constructor

attachLoadBalancerToSubnets

Arguments

:: Text
albtsLoadBalancerName

-> AttachLoadBalancerToSubnets

AttachLoadBalancerToSubnets constructor.

The fields accessible through corresponding lenses are:

albtsLoadBalancerName :: Text

albtsSubnets :: [Text]

Request lenses

albtsLoadBalancerName :: Lens' AttachLoadBalancerToSubnets Text

The name associated with the load balancer. The name must be unique within the set of load balancers associated with your AWS account.

albtsSubnets :: Lens' AttachLoadBalancerToSubnets [Text]

A list of subnet IDs to add for the load balancer. You can add only one subnet per Availability Zone.

Response

data AttachLoadBalancerToSubnetsResponse

Response constructor

attachLoadBalancerToSubnetsResponse :: AttachLoadBalancerToSubnetsResponse

AttachLoadBalancerToSubnetsResponse constructor.

The fields accessible through corresponding lenses are:

albtsrSubnets :: [Text]

Response lenses

albtsrSubnets :: Lens' AttachLoadBalancerToSubnetsResponse [Text]

A list of subnet IDs attached to the load balancer.