Network.AWS.ELB.CreateLBCookieStickinessPolicy

Description

Generates a stickiness policy with sticky session lifetimes controlled by the lifetime of the browser (user-agent) or a specified expiration period. This policy can be associated only with HTTP/HTTPS listeners.

When a load balancer implements this policy, the load balancer uses a special cookie to track the instance for each request. When the load balancer receives a request, it first checks to see if this cookie is present in the request. If so, the load balancer sends the request to the application server specified in the cookie. If not, the load balancer sends the request to a server that is chosen based on the existing load-balancing algorithm.

A cookie is inserted into the response for binding subsequent requests from the same user to that server. The validity of the cookie is based on the cookie expiration time, which is specified in the policy configuration.

For more information, see Duration-Based Session Stickiness in the Classic Load Balancer Guide .

Creating a Request

createLBCookieStickinessPolicy

Arguments

:: Text
clbcspLoadBalancerName

-> Text
clbcspPolicyName

-> CreateLBCookieStickinessPolicy

Creates a value of CreateLBCookieStickinessPolicy with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

clbcspCookieExpirationPeriod - The time period, in seconds, after which the cookie should be considered stale. If you do not specify this parameter, the default value is 0, which indicates that the sticky session should last for the duration of the browser session.

clbcspLoadBalancerName - The name of the load balancer.

clbcspPolicyName - The name of the policy being created. Policy names must consist of alphanumeric characters and dashes (-). This name must be unique within the set of policies for this load balancer.

data CreateLBCookieStickinessPolicy

Contains the parameters for CreateLBCookieStickinessPolicy.

See: createLBCookieStickinessPolicy smart constructor.

Request Lenses

clbcspCookieExpirationPeriod :: Lens' CreateLBCookieStickinessPolicy (Maybe Integer)

The time period, in seconds, after which the cookie should be considered stale. If you do not specify this parameter, the default value is 0, which indicates that the sticky session should last for the duration of the browser session.

clbcspLoadBalancerName :: Lens' CreateLBCookieStickinessPolicy Text

The name of the load balancer.

clbcspPolicyName :: Lens' CreateLBCookieStickinessPolicy Text

The name of the policy being created. Policy names must consist of alphanumeric characters and dashes (-). This name must be unique within the set of policies for this load balancer.

Destructuring the Response

createLBCookieStickinessPolicyResponse

Arguments

:: Int
clbcsprsResponseStatus

-> CreateLBCookieStickinessPolicyResponse

Creates a value of CreateLBCookieStickinessPolicyResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

clbcsprsResponseStatus - -- | The response status code.

data CreateLBCookieStickinessPolicyResponse

Contains the output for CreateLBCookieStickinessPolicy.

See: createLBCookieStickinessPolicyResponse smart constructor.

Response Lenses

clbcsprsResponseStatus :: Lens' CreateLBCookieStickinessPolicyResponse Int

- | The response status code.