Network.AWS.ELB.CreateLoadBalancer

Description

Creates a Classic Load Balancer.

You can add listeners, security groups, subnets, and tags when you create your load balancer, or you can add them later using CreateLoadBalancerListeners , ApplySecurityGroupsToLoadBalancer , AttachLoadBalancerToSubnets , and AddTags .

To describe your current load balancers, see DescribeLoadBalancers . When you are finished with a load balancer, you can delete it using DeleteLoadBalancer .

You can create up to 20 load balancers per region per account. You can request an increase for the number of load balancers for your account. For more information, see Limits for Your Classic Load Balancer in the Classic Load Balancer Guide .

Creating a Request

createLoadBalancer

Arguments

:: Text
clbLoadBalancerName

-> CreateLoadBalancer

Creates a value of CreateLoadBalancer with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

clbSecurityGroups - The IDs of the security groups to assign to the load balancer.

clbSubnets - The IDs of the subnets in your VPC to attach to the load balancer. Specify one subnet per Availability Zone specified in AvailabilityZones .

clbAvailabilityZones - One or more Availability Zones from the same region as the load balancer. You must specify at least one Availability Zone. You can add more Availability Zones after you create the load balancer using EnableAvailabilityZonesForLoadBalancer .

clbScheme - The type of a load balancer. Valid only for load balancers in a VPC. By default, Elastic Load Balancing creates an Internet-facing load balancer with a DNS name that resolves to public IP addresses. For more information about Internet-facing and Internal load balancers, see Load Balancer Scheme in the Elastic Load Balancing User Guide . Specify internal to create a load balancer with a DNS name that resolves to private IP addresses.

clbTags - A list of tags to assign to the load balancer. For more information about tagging your load balancer, see Tag Your Classic Load Balancer in the Classic Load Balancer Guide .

clbLoadBalancerName - The name of the load balancer. This name must be unique within your set of load balancers for the region, must have a maximum of 32 characters, must contain only alphanumeric characters or hyphens, and cannot begin or end with a hyphen.

clbListeners - The listeners. For more information, see Listeners for Your Classic Load Balancer in the Classic Load Balancer Guide .

data CreateLoadBalancer

Contains the parameters for CreateLoadBalancer.

See: createLoadBalancer smart constructor.

Request Lenses

clbSecurityGroups :: Lens' CreateLoadBalancer [Text]

The IDs of the security groups to assign to the load balancer.

clbSubnets :: Lens' CreateLoadBalancer [Text]

The IDs of the subnets in your VPC to attach to the load balancer. Specify one subnet per Availability Zone specified in AvailabilityZones .

clbAvailabilityZones :: Lens' CreateLoadBalancer [Text]

One or more Availability Zones from the same region as the load balancer. You must specify at least one Availability Zone. You can add more Availability Zones after you create the load balancer using EnableAvailabilityZonesForLoadBalancer .

clbScheme :: Lens' CreateLoadBalancer (Maybe Text)

The type of a load balancer. Valid only for load balancers in a VPC. By default, Elastic Load Balancing creates an Internet-facing load balancer with a DNS name that resolves to public IP addresses. For more information about Internet-facing and Internal load balancers, see Load Balancer Scheme in the Elastic Load Balancing User Guide . Specify internal to create a load balancer with a DNS name that resolves to private IP addresses.

clbTags :: Lens' CreateLoadBalancer (Maybe (NonEmpty Tag))

A list of tags to assign to the load balancer. For more information about tagging your load balancer, see Tag Your Classic Load Balancer in the Classic Load Balancer Guide .

clbLoadBalancerName :: Lens' CreateLoadBalancer Text

The name of the load balancer. This name must be unique within your set of load balancers for the region, must have a maximum of 32 characters, must contain only alphanumeric characters or hyphens, and cannot begin or end with a hyphen.

clbListeners :: Lens' CreateLoadBalancer [Listener]

The listeners. For more information, see Listeners for Your Classic Load Balancer in the Classic Load Balancer Guide .

Destructuring the Response

createLoadBalancerResponse

Arguments

:: Int
clbrsResponseStatus

-> CreateLoadBalancerResponse

Creates a value of CreateLoadBalancerResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

clbrsDNSName - The DNS name of the load balancer.

clbrsResponseStatus - -- | The response status code.

data CreateLoadBalancerResponse

Contains the output for CreateLoadBalancer.

See: createLoadBalancerResponse smart constructor.

Response Lenses

clbrsDNSName :: Lens' CreateLoadBalancerResponse (Maybe Text)

The DNS name of the load balancer.

clbrsResponseStatus :: Lens' CreateLoadBalancerResponse Int

- | The response status code.