Network.AWS.ELB.CreateLoadBalancer

Description

Creates a new load balancer.

After the call has completed successfully, a new load balancer is created with a unique Domain Name Service (DNS) name. The DNS name includes the name of the AWS region in which the load balance was created. For example, if your load balancer was created in the United States, the DNS name might end with either of the following:

us-east-1.elb.amazonaws.com (for the Northern Virginia region) us-west-1.elb.amazonaws.com (for the Northern California region) For information about the AWS regions supported by Elastic Load Balancing, see Regions and Endpoints.

You can create up to 20 load balancers per region per account.

Elastic Load Balancing supports load balancing your Amazon EC2 instances launched within any one of the following platforms:

EC2-Classic For information on creating and managing your load balancers in EC2-Classic, see Deploy Elastic Load Balancing in Amazon EC2-Classic.

EC2-VPC For information on creating and managing your load balancers in EC2-VPC, see Deploy Elastic Load Balancing in Amazon VPC.

http://docs.aws.amazon.com/ElasticLoadBalancing/latest/APIReference/API_CreateLoadBalancer.html

Request

data CreateLoadBalancer

Request constructor

createLoadBalancer

Arguments

:: Text
clbLoadBalancerName

-> NonEmpty Tag
clbTags

-> CreateLoadBalancer

CreateLoadBalancer constructor.

The fields accessible through corresponding lenses are:

clbAvailabilityZones :: [Text]

clbListeners :: [Listener]

clbLoadBalancerName :: Text

clbScheme :: Maybe Text

clbSecurityGroups :: [Text]

clbSubnets :: [Text]

clbTags :: NonEmpty Tag

Request lenses

clbAvailabilityZones :: Lens' CreateLoadBalancer [Text]

A list of Availability Zones.

At least one Availability Zone must be specified. Specified Availability Zones must be in the same EC2 Region as the load balancer. Traffic will be equally distributed across all zones.

You can later add more Availability Zones after the creation of the load balancer by calling EnableAvailabilityZonesForLoadBalancer action.

clbListeners :: Lens' CreateLoadBalancer [Listener]

A list of the following tuples: Protocol, LoadBalancerPort, InstanceProtocol, InstancePort, and SSLCertificateId.

clbLoadBalancerName :: Lens' CreateLoadBalancer Text

The name associated with the load balancer. The name must be unique within your set of load balancers, must have a maximum of 32 characters, and must only contain alphanumeric characters or hyphens.

clbScheme :: Lens' CreateLoadBalancer (Maybe Text)

The type of a load balancer.

By default, Elastic Load Balancing creates an Internet-facing load balancer with a publicly resolvable DNS name, which resolves to public IP addresses. For more information about Internet-facing and Internal load balancers, see Internet-facing and Internal Load Balancers.

Specify the value internal for this option to create an internal load balancer with a DNS name that resolves to private IP addresses.

This option is only available for load balancers created within EC2-VPC.

clbSecurityGroups :: Lens' CreateLoadBalancer [Text]

The security groups to assign to your load balancer within your VPC.

clbSubnets :: Lens' CreateLoadBalancer [Text]

A list of subnet IDs in your VPC to attach to your load balancer. Specify one subnet per Availability Zone.

clbTags :: Lens' CreateLoadBalancer (NonEmpty Tag)

A list of tags to assign to the load balancer.

For more information about setting tags for your load balancer, see Tagging.

Response

data CreateLoadBalancerResponse

Response constructor

createLoadBalancerResponse :: CreateLoadBalancerResponse

CreateLoadBalancerResponse constructor.

The fields accessible through corresponding lenses are:

clbrDNSName :: Maybe Text

Response lenses

clbrDNSName :: Lens' CreateLoadBalancerResponse (Maybe Text)

The DNS name for the load balancer.