Network.AWS.ELB.CreateLoadBalancerListeners

Description

Creates one or more listeners on a load balancer for the specified port. If a listener with the given port does not already exist, it will be created; otherwise, the properties of the new listener must match the properties of the existing listener.

For more information, see Add a Listener to Your Load Balancer in the Elastic Load Balancing Developer Guide.

http://docs.aws.amazon.com/ElasticLoadBalancing/latest/APIReference/API_CreateLoadBalancerListeners.html

Request

data CreateLoadBalancerListeners

Request constructor

createLoadBalancerListeners

Arguments

:: Text
clblLoadBalancerName

-> CreateLoadBalancerListeners

CreateLoadBalancerListeners constructor.

The fields accessible through corresponding lenses are:

clblListeners :: [Listener]

clblLoadBalancerName :: Text

Request lenses

clblListeners :: Lens' CreateLoadBalancerListeners [Listener]

A list of LoadBalancerPort, InstancePort, Protocol, InstanceProtocol, and SSLCertificateId items.

clblLoadBalancerName :: Lens' CreateLoadBalancerListeners Text

The name of the load balancer.

Response

data CreateLoadBalancerListenersResponse

Response constructor

createLoadBalancerListenersResponse :: CreateLoadBalancerListenersResponse

CreateLoadBalancerListenersResponse constructor.