Network.AWS.ELB.CreateLoadBalancerPolicy

Description

Creates a policy with the specified attributes for the specified load balancer.

Policies are settings that are saved for your load balancer and that can be applied to the listener or the application server, depending on the policy type.

Creating a Request

createLoadBalancerPolicy

Arguments

:: Text
clbpLoadBalancerName

-> Text
clbpPolicyName

-> Text
clbpPolicyTypeName

-> CreateLoadBalancerPolicy

Creates a value of CreateLoadBalancerPolicy with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

clbpPolicyAttributes - The policy attributes.

clbpLoadBalancerName - The name of the load balancer.

clbpPolicyName - The name of the load balancer policy to be created. This name must be unique within the set of policies for this load balancer.

clbpPolicyTypeName - The name of the base policy type. To get the list of policy types, use DescribeLoadBalancerPolicyTypes .

data CreateLoadBalancerPolicy

Contains the parameters for CreateLoadBalancerPolicy.

See: createLoadBalancerPolicy smart constructor.

Request Lenses

clbpPolicyAttributes :: Lens' CreateLoadBalancerPolicy [PolicyAttribute]

The policy attributes.

clbpLoadBalancerName :: Lens' CreateLoadBalancerPolicy Text

The name of the load balancer.

clbpPolicyName :: Lens' CreateLoadBalancerPolicy Text

The name of the load balancer policy to be created. This name must be unique within the set of policies for this load balancer.

clbpPolicyTypeName :: Lens' CreateLoadBalancerPolicy Text

The name of the base policy type. To get the list of policy types, use DescribeLoadBalancerPolicyTypes .

Destructuring the Response

createLoadBalancerPolicyResponse

Arguments

:: Int
clbprsResponseStatus

-> CreateLoadBalancerPolicyResponse

Creates a value of CreateLoadBalancerPolicyResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

clbprsResponseStatus - -- | The response status code.

data CreateLoadBalancerPolicyResponse

Contains the output of CreateLoadBalancerPolicy.

See: createLoadBalancerPolicyResponse smart constructor.

Response Lenses

clbprsResponseStatus :: Lens' CreateLoadBalancerPolicyResponse Int

- | The response status code.