Network.AWS.ELB.DeregisterInstancesFromLoadBalancer

Description

Deregisters the specified instances from the specified load balancer. After the instance is deregistered, it no longer receives traffic from the load balancer.

You can use DescribeLoadBalancers to verify that the instance is deregistered from the load balancer.

For more information, see Register or De-Register EC2 Instances in the Classic Load Balancer Guide .

Creating a Request

deregisterInstancesFromLoadBalancer

Arguments

:: Text
diflbLoadBalancerName

-> DeregisterInstancesFromLoadBalancer

Creates a value of DeregisterInstancesFromLoadBalancer with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

diflbLoadBalancerName - The name of the load balancer.

diflbInstances - The IDs of the instances.

data DeregisterInstancesFromLoadBalancer

Contains the parameters for DeregisterInstancesFromLoadBalancer.

See: deregisterInstancesFromLoadBalancer smart constructor.

Request Lenses

diflbLoadBalancerName :: Lens' DeregisterInstancesFromLoadBalancer Text

The name of the load balancer.

diflbInstances :: Lens' DeregisterInstancesFromLoadBalancer [Instance]

The IDs of the instances.

Destructuring the Response

deregisterInstancesFromLoadBalancerResponse

Arguments

:: Int
diflbrsResponseStatus

-> DeregisterInstancesFromLoadBalancerResponse

Creates a value of DeregisterInstancesFromLoadBalancerResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

diflbrsInstances - The remaining instances registered with the load balancer.

diflbrsResponseStatus - -- | The response status code.

data DeregisterInstancesFromLoadBalancerResponse

Contains the output of DeregisterInstancesFromLoadBalancer.

See: deregisterInstancesFromLoadBalancerResponse smart constructor.

Response Lenses

diflbrsInstances :: Lens' DeregisterInstancesFromLoadBalancerResponse [Instance]

The remaining instances registered with the load balancer.

diflbrsResponseStatus :: Lens' DeregisterInstancesFromLoadBalancerResponse Int

- | The response status code.