Network.AWS.ELB.DescribeInstanceHealth

Description

Describes the state of the specified instances with respect to the specified load balancer. If no instances are specified, the call describes the state of all instances that are currently registered with the load balancer. If instances are specified, their state is returned even if they are no longer registered with the load balancer. The state of terminated instances is not returned.

Creating a Request

describeInstanceHealth

Arguments

:: Text
dihLoadBalancerName

-> DescribeInstanceHealth

Creates a value of DescribeInstanceHealth with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dihInstances - The IDs of the instances.

dihLoadBalancerName - The name of the load balancer.

data DescribeInstanceHealth

Contains the parameters for DescribeInstanceHealth.

See: describeInstanceHealth smart constructor.

Request Lenses

dihInstances :: Lens' DescribeInstanceHealth [Instance]

The IDs of the instances.

dihLoadBalancerName :: Lens' DescribeInstanceHealth Text

The name of the load balancer.

Destructuring the Response

describeInstanceHealthResponse

Arguments

:: Int
dihrsResponseStatus

-> DescribeInstanceHealthResponse

Creates a value of DescribeInstanceHealthResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dihrsInstanceStates - Information about the health of the instances.

dihrsResponseStatus - -- | The response status code.

data DescribeInstanceHealthResponse

Contains the output for DescribeInstanceHealth.

See: describeInstanceHealthResponse smart constructor.

Response Lenses

dihrsInstanceStates :: Lens' DescribeInstanceHealthResponse [InstanceState]

Information about the health of the instances.

dihrsResponseStatus :: Lens' DescribeInstanceHealthResponse Int

- | The response status code.