Network.AWS.ELB.DescribeInstanceHealth

Description

Returns the current state of the specified instances registered with the specified load balancer. If no instances are specified, the state of all the instances registered with the load balancer is returned.

You must provide the same account credentials as those that were used to create the load balancer.

http://docs.aws.amazon.com/ElasticLoadBalancing/latest/APIReference/API_DescribeInstanceHealth.html

Request

data DescribeInstanceHealth

Request constructor

describeInstanceHealth

Arguments

:: Text
dihLoadBalancerName

-> DescribeInstanceHealth

DescribeInstanceHealth constructor.

The fields accessible through corresponding lenses are:

dihInstances :: [Instance]

dihLoadBalancerName :: Text

Request lenses

dihInstances :: Lens' DescribeInstanceHealth [Instance]

A list of instance IDs whose states are being queried.

dihLoadBalancerName :: Lens' DescribeInstanceHealth Text

The name of the load balancer.

Response

data DescribeInstanceHealthResponse

Response constructor

describeInstanceHealthResponse :: DescribeInstanceHealthResponse

DescribeInstanceHealthResponse constructor.

The fields accessible through corresponding lenses are:

dihrInstanceStates :: [InstanceState]

Response lenses

dihrInstanceStates :: Lens' DescribeInstanceHealthResponse [InstanceState]

A list containing health information for the specified instances.