Network.AWS.ELB.DescribeLoadBalancerPolicies

Description

Describes the specified policies.

If you specify a load balancer name, the action returns the descriptions of all policies created for the load balancer. If you specify a policy name associated with your load balancer, the action returns the description of that policy. If you don't specify a load balancer name, the action returns descriptions of the specified sample policies, or descriptions of all sample policies. The names of the sample policies have the ELBSample- prefix.

Creating a Request

describeLoadBalancerPolicies :: DescribeLoadBalancerPolicies

Creates a value of DescribeLoadBalancerPolicies with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dlbpPolicyNames - The names of the policies.

dlbpLoadBalancerName - The name of the load balancer.

data DescribeLoadBalancerPolicies

Contains the parameters for DescribeLoadBalancerPolicies.

See: describeLoadBalancerPolicies smart constructor.

Request Lenses

dlbpPolicyNames :: Lens' DescribeLoadBalancerPolicies [Text]

The names of the policies.

dlbpLoadBalancerName :: Lens' DescribeLoadBalancerPolicies (Maybe Text)

The name of the load balancer.

Destructuring the Response

describeLoadBalancerPoliciesResponse

Arguments

:: Int
dlbprsResponseStatus

-> DescribeLoadBalancerPoliciesResponse

Creates a value of DescribeLoadBalancerPoliciesResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dlbprsPolicyDescriptions - Information about the policies.

dlbprsResponseStatus - -- | The response status code.

data DescribeLoadBalancerPoliciesResponse

Contains the output of DescribeLoadBalancerPolicies.

See: describeLoadBalancerPoliciesResponse smart constructor.

Response Lenses

dlbprsPolicyDescriptions :: Lens' DescribeLoadBalancerPoliciesResponse [PolicyDescription]

Information about the policies.

dlbprsResponseStatus :: Lens' DescribeLoadBalancerPoliciesResponse Int

- | The response status code.