Network.AWS.ELB.DescribeLoadBalancerPolicyTypes

Description

Describes the specified load balancer policy types or all load balancer policy types.

The description of each type indicates how it can be used. For example, some policies can be used only with layer 7 listeners, some policies can be used only with layer 4 listeners, and some policies can be used only with your EC2 instances.

You can use CreateLoadBalancerPolicy to create a policy configuration for any of these policy types. Then, depending on the policy type, use either SetLoadBalancerPoliciesOfListener or SetLoadBalancerPoliciesForBackendServer to set the policy.

Creating a Request

describeLoadBalancerPolicyTypes :: DescribeLoadBalancerPolicyTypes

Creates a value of DescribeLoadBalancerPolicyTypes with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dlbptPolicyTypeNames - The names of the policy types. If no names are specified, describes all policy types defined by Elastic Load Balancing.

data DescribeLoadBalancerPolicyTypes

Contains the parameters for DescribeLoadBalancerPolicyTypes.

See: describeLoadBalancerPolicyTypes smart constructor.

Request Lenses

dlbptPolicyTypeNames :: Lens' DescribeLoadBalancerPolicyTypes [Text]

The names of the policy types. If no names are specified, describes all policy types defined by Elastic Load Balancing.

Destructuring the Response

describeLoadBalancerPolicyTypesResponse

Arguments

:: Int
dlbptrsResponseStatus

-> DescribeLoadBalancerPolicyTypesResponse

Creates a value of DescribeLoadBalancerPolicyTypesResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dlbptrsPolicyTypeDescriptions - Information about the policy types.

dlbptrsResponseStatus - -- | The response status code.

data DescribeLoadBalancerPolicyTypesResponse

Contains the output of DescribeLoadBalancerPolicyTypes.

See: describeLoadBalancerPolicyTypesResponse smart constructor.

Response Lenses

dlbptrsPolicyTypeDescriptions :: Lens' DescribeLoadBalancerPolicyTypesResponse [PolicyTypeDescription]

Information about the policy types.

dlbptrsResponseStatus :: Lens' DescribeLoadBalancerPolicyTypesResponse Int

- | The response status code.