Network.AWS.ELB.DescribeLoadBalancers

Description

Describes the specified the load balancers. If no load balancers are specified, the call describes all of your load balancers.

This operation returns paginated results.

Creating a Request

describeLoadBalancers :: DescribeLoadBalancers

Creates a value of DescribeLoadBalancers with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dlbMarker - The marker for the next set of results. (You received this marker from a previous call.)

dlbPageSize - The maximum number of results to return with this call (a number from 1 to 400). The default is 400.

dlbLoadBalancerNames - The names of the load balancers.

data DescribeLoadBalancers

Contains the parameters for DescribeLoadBalancers.

See: describeLoadBalancers smart constructor.

Request Lenses

dlbMarker :: Lens' DescribeLoadBalancers (Maybe Text)

The marker for the next set of results. (You received this marker from a previous call.)

dlbPageSize :: Lens' DescribeLoadBalancers (Maybe Natural)

The maximum number of results to return with this call (a number from 1 to 400). The default is 400.

dlbLoadBalancerNames :: Lens' DescribeLoadBalancers [Text]

The names of the load balancers.

Destructuring the Response

describeLoadBalancersResponse

Arguments

:: Int
dlbrsResponseStatus

-> DescribeLoadBalancersResponse

Creates a value of DescribeLoadBalancersResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dlbrsLoadBalancerDescriptions - Information about the load balancers.

dlbrsNextMarker - The marker to use when requesting the next set of results. If there are no additional results, the string is empty.

dlbrsResponseStatus - -- | The response status code.

data DescribeLoadBalancersResponse

Contains the parameters for DescribeLoadBalancers.

See: describeLoadBalancersResponse smart constructor.

Response Lenses

dlbrsLoadBalancerDescriptions :: Lens' DescribeLoadBalancersResponse [LoadBalancerDescription]

Information about the load balancers.

dlbrsNextMarker :: Lens' DescribeLoadBalancersResponse (Maybe Text)

The marker to use when requesting the next set of results. If there are no additional results, the string is empty.

dlbrsResponseStatus :: Lens' DescribeLoadBalancersResponse Int

- | The response status code.