Network.AWS.ELB.DescribeLoadBalancers

Description

Returns detailed configuration information for all the load balancers created for the account. If you specify load balancer names, the action returns configuration information of the specified load balancers.

In order to retrieve this information, you must provide the same account credentials that was used to create the load balancer.

http://docs.aws.amazon.com/ElasticLoadBalancing/latest/APIReference/API_DescribeLoadBalancers.html

Request

data DescribeLoadBalancers

Request constructor

describeLoadBalancers :: DescribeLoadBalancers

DescribeLoadBalancers constructor.

The fields accessible through corresponding lenses are:

dlbLoadBalancerNames :: [Text]

dlbMarker :: Maybe Text

dlbPageSize :: Maybe Natural

Request lenses

dlbLoadBalancerNames :: Lens' DescribeLoadBalancers [Text]

A list of load balancer names associated with the account.

dlbMarker :: Lens' DescribeLoadBalancers (Maybe Text)

An optional parameter used for pagination of results from this call. If specified, the response includes only records beyond the marker.

dlbPageSize :: Lens' DescribeLoadBalancers (Maybe Natural)

The number of results returned in each page. The default is 400. You cannot specify a page size greater than 400 or less than 1.

Response

data DescribeLoadBalancersResponse

Response constructor

describeLoadBalancersResponse :: DescribeLoadBalancersResponse

DescribeLoadBalancersResponse constructor.

The fields accessible through corresponding lenses are:

dlbrLoadBalancerDescriptions :: [LoadBalancerDescription]

dlbrNextMarker :: Maybe Text

Response lenses

dlbrLoadBalancerDescriptions :: Lens' DescribeLoadBalancersResponse [LoadBalancerDescription]

A list of load balancer description structures.

dlbrNextMarker :: Lens' DescribeLoadBalancersResponse (Maybe Text)

Specifies the value of next marker if the request returned more than one page of results.