Network.AWS.ELB.DescribeTags

Description

Describes the tags associated with the specified load balancers.

Creating a Request

describeTags

Arguments

:: NonEmpty Text
dtLoadBalancerNames

-> DescribeTags

Creates a value of DescribeTags with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dtLoadBalancerNames - The names of the load balancers.

data DescribeTags

Contains the parameters for DescribeTags.

See: describeTags smart constructor.

Request Lenses

dtLoadBalancerNames :: Lens' DescribeTags (NonEmpty Text)

The names of the load balancers.

Destructuring the Response

describeTagsResponse

Arguments

:: Int
dtrsResponseStatus

-> DescribeTagsResponse

Creates a value of DescribeTagsResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dtrsTagDescriptions - Information about the tags.

dtrsResponseStatus - -- | The response status code.

data DescribeTagsResponse

Contains the output for DescribeTags.

See: describeTagsResponse smart constructor.

Response Lenses

dtrsTagDescriptions :: Lens' DescribeTagsResponse [TagDescription]

Information about the tags.

dtrsResponseStatus :: Lens' DescribeTagsResponse Int

- | The response status code.