Network.AWS.ELB.DetachLoadBalancerFromSubnets

Description

Removes the specified subnets from the set of configured subnets for the load balancer.

After a subnet is removed, all EC2 instances registered with the load balancer in the removed subnet go into the OutOfService state. Then, the load balancer balances the traffic among the remaining routable subnets.

Creating a Request

detachLoadBalancerFromSubnets

Arguments

:: Text
dlbfsLoadBalancerName

-> DetachLoadBalancerFromSubnets

Creates a value of DetachLoadBalancerFromSubnets with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dlbfsLoadBalancerName - The name of the load balancer.

dlbfsSubnets - The IDs of the subnets.

data DetachLoadBalancerFromSubnets

Contains the parameters for DetachLoadBalancerFromSubnets.

See: detachLoadBalancerFromSubnets smart constructor.

Request Lenses

dlbfsLoadBalancerName :: Lens' DetachLoadBalancerFromSubnets Text

The name of the load balancer.

dlbfsSubnets :: Lens' DetachLoadBalancerFromSubnets [Text]

The IDs of the subnets.

Destructuring the Response

detachLoadBalancerFromSubnetsResponse

Arguments

:: Int
dlbfsrsResponseStatus

-> DetachLoadBalancerFromSubnetsResponse

Creates a value of DetachLoadBalancerFromSubnetsResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dlbfsrsSubnets - The IDs of the remaining subnets for the load balancer.

dlbfsrsResponseStatus - -- | The response status code.

data DetachLoadBalancerFromSubnetsResponse

Contains the output of DetachLoadBalancerFromSubnets.

See: detachLoadBalancerFromSubnetsResponse smart constructor.

Response Lenses

dlbfsrsSubnets :: Lens' DetachLoadBalancerFromSubnetsResponse [Text]

The IDs of the remaining subnets for the load balancer.

dlbfsrsResponseStatus :: Lens' DetachLoadBalancerFromSubnetsResponse Int

- | The response status code.