Network.AWS.ELB.DisableAvailabilityZonesForLoadBalancer

Description

Removes the specified Availability Zones from the set of Availability Zones for the specified load balancer.

There must be at least one Availability Zone registered with a load balancer at all times. After an Availability Zone is removed, all instances registered with the load balancer that are in the removed Availability Zone go into the OutOfService state. Then, the load balancer attempts to equally balance the traffic among its remaining Availability Zones.

For more information, see Add or Remove Availability Zones in the Classic Load Balancer Guide .

Creating a Request

disableAvailabilityZonesForLoadBalancer

Arguments

:: Text
dazflbLoadBalancerName

-> DisableAvailabilityZonesForLoadBalancer

Creates a value of DisableAvailabilityZonesForLoadBalancer with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dazflbLoadBalancerName - The name of the load balancer.

dazflbAvailabilityZones - The Availability Zones.

data DisableAvailabilityZonesForLoadBalancer

Contains the parameters for DisableAvailabilityZonesForLoadBalancer.

See: disableAvailabilityZonesForLoadBalancer smart constructor.

Request Lenses

dazflbLoadBalancerName :: Lens' DisableAvailabilityZonesForLoadBalancer Text

The name of the load balancer.

dazflbAvailabilityZones :: Lens' DisableAvailabilityZonesForLoadBalancer [Text]

The Availability Zones.

Destructuring the Response

disableAvailabilityZonesForLoadBalancerResponse

Arguments

:: Int
dazflbrsResponseStatus

-> DisableAvailabilityZonesForLoadBalancerResponse

Creates a value of DisableAvailabilityZonesForLoadBalancerResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dazflbrsAvailabilityZones - The remaining Availability Zones for the load balancer.

dazflbrsResponseStatus - -- | The response status code.

data DisableAvailabilityZonesForLoadBalancerResponse

Contains the output for DisableAvailabilityZonesForLoadBalancer.

See: disableAvailabilityZonesForLoadBalancerResponse smart constructor.

Response Lenses

dazflbrsAvailabilityZones :: Lens' DisableAvailabilityZonesForLoadBalancerResponse [Text]

The remaining Availability Zones for the load balancer.

dazflbrsResponseStatus :: Lens' DisableAvailabilityZonesForLoadBalancerResponse Int

- | The response status code.