Network.AWS.ELB.DisableAvailabilityZonesForLoadBalancer

Description

Removes the specified EC2 Availability Zones from the set of configured Availability Zones for the load balancer.

There must be at least one Availability Zone registered with a load balancer at all times. Once an Availability Zone is removed, all the instances registered with the load balancer that are in the removed Availability Zone go into the OutOfService state. Upon Availability Zone removal, the load balancer attempts to equally balance the traffic among its remaining usable Availability Zones. Trying to remove an Availability Zone that was not associated with the load balancer does nothing.

For more information, see Disable an Availability Zone from a Load-BalancedApplication in the Elastic Load Balancing Developer Guide.

http://docs.aws.amazon.com/ElasticLoadBalancing/latest/APIReference/API_DisableAvailabilityZonesForLoadBalancer.html

Request

data DisableAvailabilityZonesForLoadBalancer

Request constructor

disableAvailabilityZonesForLoadBalancer

Arguments

:: Text
dazflbLoadBalancerName

-> DisableAvailabilityZonesForLoadBalancer

DisableAvailabilityZonesForLoadBalancer constructor.

The fields accessible through corresponding lenses are:

dazflbAvailabilityZones :: [Text]

dazflbLoadBalancerName :: Text

Request lenses

dazflbAvailabilityZones :: Lens' DisableAvailabilityZonesForLoadBalancer [Text]

A list of Availability Zones to be removed from the load balancer.

There must be at least one Availability Zone registered with a load balancer at all times. Specified Availability Zones must be in the same region.

dazflbLoadBalancerName :: Lens' DisableAvailabilityZonesForLoadBalancer Text

The name associated with the load balancer.

Response

data DisableAvailabilityZonesForLoadBalancerResponse

Response constructor

disableAvailabilityZonesForLoadBalancerResponse :: DisableAvailabilityZonesForLoadBalancerResponse

DisableAvailabilityZonesForLoadBalancerResponse constructor.

The fields accessible through corresponding lenses are:

dazflbrAvailabilityZones :: [Text]

Response lenses

dazflbrAvailabilityZones :: Lens' DisableAvailabilityZonesForLoadBalancerResponse [Text]

A list of updated Availability Zones for the load balancer.