Network.AWS.ELB.EnableAvailabilityZonesForLoadBalancer

Description

Adds the specified Availability Zones to the set of Availability Zones for the specified load balancer.

The load balancer evenly distributes requests across all its registered Availability Zones that contain instances.

For more information, see Add or Remove Availability Zones in the Classic Load Balancer Guide .

Creating a Request

enableAvailabilityZonesForLoadBalancer

Arguments

:: Text
eazflbLoadBalancerName

-> EnableAvailabilityZonesForLoadBalancer

Creates a value of EnableAvailabilityZonesForLoadBalancer with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

eazflbLoadBalancerName - The name of the load balancer.

eazflbAvailabilityZones - The Availability Zones. These must be in the same region as the load balancer.

data EnableAvailabilityZonesForLoadBalancer

Contains the parameters for EnableAvailabilityZonesForLoadBalancer.

See: enableAvailabilityZonesForLoadBalancer smart constructor.

Request Lenses

eazflbLoadBalancerName :: Lens' EnableAvailabilityZonesForLoadBalancer Text

The name of the load balancer.

eazflbAvailabilityZones :: Lens' EnableAvailabilityZonesForLoadBalancer [Text]

The Availability Zones. These must be in the same region as the load balancer.

Destructuring the Response

enableAvailabilityZonesForLoadBalancerResponse

Arguments

:: Int
eazflbrsResponseStatus

-> EnableAvailabilityZonesForLoadBalancerResponse

Creates a value of EnableAvailabilityZonesForLoadBalancerResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

eazflbrsAvailabilityZones - The updated list of Availability Zones for the load balancer.

eazflbrsResponseStatus - -- | The response status code.

data EnableAvailabilityZonesForLoadBalancerResponse

Contains the output of EnableAvailabilityZonesForLoadBalancer.

See: enableAvailabilityZonesForLoadBalancerResponse smart constructor.

Response Lenses

eazflbrsAvailabilityZones :: Lens' EnableAvailabilityZonesForLoadBalancerResponse [Text]

The updated list of Availability Zones for the load balancer.

eazflbrsResponseStatus :: Lens' EnableAvailabilityZonesForLoadBalancerResponse Int

- | The response status code.