Network.AWS.ELB.RegisterInstancesWithLoadBalancer

Description

Adds the specified instances to the specified load balancer.

The instance must be a running instance in the same network as the load balancer (EC2-Classic or the same VPC). If you have EC2-Classic instances and a load balancer in a VPC with ClassicLink enabled, you can link the EC2-Classic instances to that VPC and then register the linked EC2-Classic instances with the load balancer in the VPC.

Note that RegisterInstanceWithLoadBalancer completes when the request has been registered. Instance registration takes a little time to complete. To check the state of the registered instances, use DescribeLoadBalancers or DescribeInstanceHealth .

After the instance is registered, it starts receiving traffic and requests from the load balancer. Any instance that is not in one of the Availability Zones registered for the load balancer is moved to the OutOfService state. If an Availability Zone is added to the load balancer later, any instances registered with the load balancer move to the InService state.

To deregister instances from a load balancer, use DeregisterInstancesFromLoadBalancer .

For more information, see Register or De-Register EC2 Instances in the Classic Load Balancer Guide .

Creating a Request

registerInstancesWithLoadBalancer

Arguments

:: Text
riwlbLoadBalancerName

-> RegisterInstancesWithLoadBalancer

Creates a value of RegisterInstancesWithLoadBalancer with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

riwlbLoadBalancerName - The name of the load balancer.

riwlbInstances - The IDs of the instances.

data RegisterInstancesWithLoadBalancer

Contains the parameters for RegisterInstancesWithLoadBalancer.

See: registerInstancesWithLoadBalancer smart constructor.

Request Lenses

riwlbLoadBalancerName :: Lens' RegisterInstancesWithLoadBalancer Text

The name of the load balancer.

riwlbInstances :: Lens' RegisterInstancesWithLoadBalancer [Instance]

The IDs of the instances.

Destructuring the Response

registerInstancesWithLoadBalancerResponse

Arguments

:: Int
riwlbrsResponseStatus

-> RegisterInstancesWithLoadBalancerResponse

Creates a value of RegisterInstancesWithLoadBalancerResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

riwlbrsInstances - The updated list of instances for the load balancer.

riwlbrsResponseStatus - -- | The response status code.

data RegisterInstancesWithLoadBalancerResponse

Contains the output of RegisterInstancesWithLoadBalancer.

See: registerInstancesWithLoadBalancerResponse smart constructor.

Response Lenses

riwlbrsInstances :: Lens' RegisterInstancesWithLoadBalancerResponse [Instance]

The updated list of instances for the load balancer.

riwlbrsResponseStatus :: Lens' RegisterInstancesWithLoadBalancerResponse Int

- | The response status code.