Network.AWS.ELB.RemoveTags

Description

Removes one or more tags from the specified load balancer.

http://docs.aws.amazon.com/ElasticLoadBalancing/latest/APIReference/API_RemoveTags.html

Request

data RemoveTags

Request constructor

removeTags

Arguments

:: NonEmpty TagKeyOnly
rtTags

-> RemoveTags

RemoveTags constructor.

The fields accessible through corresponding lenses are:

rtLoadBalancerNames :: [Text]

rtTags :: NonEmpty TagKeyOnly

Request lenses

rtLoadBalancerNames :: Lens' RemoveTags [Text]

The name of the load balancer. You can specify a maximum of one load balancer name.

rtTags :: Lens' RemoveTags (NonEmpty TagKeyOnly)

A list of tag keys to remove.

Response

data RemoveTagsResponse

Response constructor

removeTagsResponse :: RemoveTagsResponse

RemoveTagsResponse constructor.