Network.AWS.ELB.SetLoadBalancerPoliciesForBackendServer

Description

Replaces the set of policies associated with the specified port on which the EC2 instance is listening with a new set of policies. At this time, only the back-end server authentication policy type can be applied to the instance ports; this policy type is composed of multiple public key policies.

Each time you use SetLoadBalancerPoliciesForBackendServer to enable the policies, use the PolicyNames parameter to list the policies that you want to enable.

You can use DescribeLoadBalancers or DescribeLoadBalancerPolicies to verify that the policy is associated with the EC2 instance.

For more information about enabling back-end instance authentication, see Configure Back-end Instance Authentication in the Classic Load Balancer Guide . For more information about Proxy Protocol, see Configure Proxy Protocol Support in the Classic Load Balancer Guide .

Creating a Request

setLoadBalancerPoliciesForBackendServer

Arguments

:: Text
slbpfbsLoadBalancerName

-> Int
slbpfbsInstancePort

-> SetLoadBalancerPoliciesForBackendServer

Creates a value of SetLoadBalancerPoliciesForBackendServer with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

slbpfbsLoadBalancerName - The name of the load balancer.

slbpfbsInstancePort - The port number associated with the EC2 instance.

slbpfbsPolicyNames - The names of the policies. If the list is empty, then all current polices are removed from the EC2 instance.

data SetLoadBalancerPoliciesForBackendServer

Contains the parameters for SetLoadBalancerPoliciesForBackendServer.

See: setLoadBalancerPoliciesForBackendServer smart constructor.

Request Lenses

slbpfbsLoadBalancerName :: Lens' SetLoadBalancerPoliciesForBackendServer Text

The name of the load balancer.

slbpfbsInstancePort :: Lens' SetLoadBalancerPoliciesForBackendServer Int

The port number associated with the EC2 instance.

slbpfbsPolicyNames :: Lens' SetLoadBalancerPoliciesForBackendServer [Text]

The names of the policies. If the list is empty, then all current polices are removed from the EC2 instance.

Destructuring the Response

setLoadBalancerPoliciesForBackendServerResponse

Arguments

:: Int
slbpfbsrsResponseStatus

-> SetLoadBalancerPoliciesForBackendServerResponse

Creates a value of SetLoadBalancerPoliciesForBackendServerResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

slbpfbsrsResponseStatus - -- | The response status code.

data SetLoadBalancerPoliciesForBackendServerResponse

Contains the output of SetLoadBalancerPoliciesForBackendServer.

See: setLoadBalancerPoliciesForBackendServerResponse smart constructor.

Response Lenses

slbpfbsrsResponseStatus :: Lens' SetLoadBalancerPoliciesForBackendServerResponse Int

- | The response status code.