Network.AWS.ELB.SetLoadBalancerPoliciesOfListener

Description

Replaces the current set of policies for the specified load balancer port with the specified set of policies.

To enable back-end server authentication, use SetLoadBalancerPoliciesForBackendServer .

For more information about setting policies, see Update the SSL Negotiation Configuration , Duration-Based Session Stickiness , and Application-Controlled Session Stickiness in the Classic Load Balancer Guide .

Creating a Request

setLoadBalancerPoliciesOfListener

Arguments

:: Text
slbpolLoadBalancerName

-> Int
slbpolLoadBalancerPort

-> SetLoadBalancerPoliciesOfListener

Creates a value of SetLoadBalancerPoliciesOfListener with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

slbpolLoadBalancerName - The name of the load balancer.

slbpolLoadBalancerPort - The external port of the load balancer.

slbpolPolicyNames - The names of the policies. This list must include all policies to be enabled. If you omit a policy that is currently enabled, it is disabled. If the list is empty, all current policies are disabled.

data SetLoadBalancerPoliciesOfListener

Contains the parameters for SetLoadBalancePoliciesOfListener.

See: setLoadBalancerPoliciesOfListener smart constructor.

Request Lenses

slbpolLoadBalancerName :: Lens' SetLoadBalancerPoliciesOfListener Text

The name of the load balancer.

slbpolLoadBalancerPort :: Lens' SetLoadBalancerPoliciesOfListener Int

The external port of the load balancer.

slbpolPolicyNames :: Lens' SetLoadBalancerPoliciesOfListener [Text]

The names of the policies. This list must include all policies to be enabled. If you omit a policy that is currently enabled, it is disabled. If the list is empty, all current policies are disabled.

Destructuring the Response

setLoadBalancerPoliciesOfListenerResponse

Arguments

:: Int
slbpolrsResponseStatus

-> SetLoadBalancerPoliciesOfListenerResponse

Creates a value of SetLoadBalancerPoliciesOfListenerResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

slbpolrsResponseStatus - -- | The response status code.

data SetLoadBalancerPoliciesOfListenerResponse

Contains the output of SetLoadBalancePoliciesOfListener.

See: setLoadBalancerPoliciesOfListenerResponse smart constructor.

Response Lenses

slbpolrsResponseStatus :: Lens' SetLoadBalancerPoliciesOfListenerResponse Int

- | The response status code.