Network.AWS.ELB.Types

Description

Service Configuration

elb :: Service

API version 2012-06-01 of the Amazon Elastic Load Balancing SDK configuration.

Errors

_PolicyNotFoundException :: AsError a => Getting (First ServiceError) a ServiceError

One or more of the specified policies do not exist.

_AccessPointNotFoundException :: AsError a => Getting (First ServiceError) a ServiceError

The specified load balancer does not exist.

_DuplicatePolicyNameException :: AsError a => Getting (First ServiceError) a ServiceError

A policy with the specified name already exists for this load balancer.

_InvalidConfigurationRequestException :: AsError a => Getting (First ServiceError) a ServiceError

The requested configuration change is not valid.

_SubnetNotFoundException :: AsError a => Getting (First ServiceError) a ServiceError

One or more of the specified subnets do not exist.

_LoadBalancerAttributeNotFoundException :: AsError a => Getting (First ServiceError) a ServiceError

The specified load balancer attribute does not exist.

_InvalidSubnetException :: AsError a => Getting (First ServiceError) a ServiceError

The specified VPC has no associated Internet gateway.

_DuplicateTagKeysException :: AsError a => Getting (First ServiceError) a ServiceError

A tag key was specified more than once.

_DuplicateListenerException :: AsError a => Getting (First ServiceError) a ServiceError

A listener already exists for the specified load balancer name and port, but with a different instance port, protocol, or SSL certificate.

_TooManyTagsException :: AsError a => Getting (First ServiceError) a ServiceError

The quota for the number of tags that can be assigned to a load balancer has been reached.

_PolicyTypeNotFoundException :: AsError a => Getting (First ServiceError) a ServiceError

One or more of the specified policy types do not exist.

_UnsupportedProtocolException :: AsError a => Getting (First ServiceError) a ServiceError

The specified protocol or signature version is not supported.

_DuplicateAccessPointNameException :: AsError a => Getting (First ServiceError) a ServiceError

The specified load balancer name already exists for this account.

_InvalidSecurityGroupException :: AsError a => Getting (First ServiceError) a ServiceError

One or more of the specified security groups do not exist.

_ListenerNotFoundException :: AsError a => Getting (First ServiceError) a ServiceError

The load balancer does not have a listener configured at the specified port.

_InvalidEndPointException :: AsError a => Getting (First ServiceError) a ServiceError

The specified endpoint is not valid.

_DependencyThrottleException :: AsError a => Getting (First ServiceError) a ServiceError

Prism for DependencyThrottleException' errors.

_InvalidSchemeException :: AsError a => Getting (First ServiceError) a ServiceError

The specified value for the schema is not valid. You can only specify a scheme for load balancers in a VPC.

_TooManyAccessPointsException :: AsError a => Getting (First ServiceError) a ServiceError

The quota for the number of load balancers has been reached.

_TooManyPoliciesException :: AsError a => Getting (First ServiceError) a ServiceError

The quota for the number of policies for this load balancer has been reached.

_CertificateNotFoundException :: AsError a => Getting (First ServiceError) a ServiceError

The specified ARN does not refer to a valid SSL certificate in AWS Identity and Access Management (IAM) or AWS Certificate Manager (ACM). Note that if you recently uploaded the certificate to IAM, this error might indicate that the certificate is not fully available yet.

Re-exported Types

getAccountId :: Region -> Text

This account identifier is used when attaching a policy to your S3 bucket allowing ELB to upload and write access logs.

See: Attach a Policy to Your S3 Bucket.

AccessLog

data AccessLog

Information about the AccessLog attribute.

See: accessLog smart constructor.

accessLog

Arguments

:: Bool
alEnabled

-> AccessLog

Creates a value of AccessLog with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

alEmitInterval - The interval for publishing the access logs. You can specify an interval of either 5 minutes or 60 minutes. Default: 60 minutes

alS3BucketPrefix - The logical hierarchy you created for your Amazon S3 bucket, for example my-bucket-prefix/prod . If the prefix is not provided, the log is placed at the root level of the bucket.

alS3BucketName - The name of the Amazon S3 bucket where the access logs are stored.

alEnabled - Specifies whether access logs are enabled for the load balancer.

alEmitInterval :: Lens' AccessLog (Maybe Int)

The interval for publishing the access logs. You can specify an interval of either 5 minutes or 60 minutes. Default: 60 minutes

alS3BucketPrefix :: Lens' AccessLog (Maybe Text)

The logical hierarchy you created for your Amazon S3 bucket, for example my-bucket-prefix/prod . If the prefix is not provided, the log is placed at the root level of the bucket.

alS3BucketName :: Lens' AccessLog (Maybe Text)

The name of the Amazon S3 bucket where the access logs are stored.

alEnabled :: Lens' AccessLog Bool

Specifies whether access logs are enabled for the load balancer.

AdditionalAttribute

data AdditionalAttribute

This data type is reserved.

See: additionalAttribute smart constructor.

additionalAttribute :: AdditionalAttribute

Creates a value of AdditionalAttribute with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

aaValue - This parameter is reserved.

aaKey - This parameter is reserved.

aaValue :: Lens' AdditionalAttribute (Maybe Text)

This parameter is reserved.

aaKey :: Lens' AdditionalAttribute (Maybe Text)

This parameter is reserved.

AppCookieStickinessPolicy

data AppCookieStickinessPolicy

Information about a policy for application-controlled session stickiness.

See: appCookieStickinessPolicy smart constructor.

appCookieStickinessPolicy :: AppCookieStickinessPolicy

Creates a value of AppCookieStickinessPolicy with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

acspPolicyName - The mnemonic name for the policy being created. The name must be unique within a set of policies for this load balancer.

acspCookieName - The name of the application cookie used for stickiness.

acspPolicyName :: Lens' AppCookieStickinessPolicy (Maybe Text)

The mnemonic name for the policy being created. The name must be unique within a set of policies for this load balancer.

acspCookieName :: Lens' AppCookieStickinessPolicy (Maybe Text)

The name of the application cookie used for stickiness.

BackendServerDescription

data BackendServerDescription

Information about the configuration of an EC2 instance.

See: backendServerDescription smart constructor.

backendServerDescription :: BackendServerDescription

Creates a value of BackendServerDescription with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

bsdPolicyNames - The names of the policies enabled for the EC2 instance.

bsdInstancePort - The port on which the EC2 instance is listening.

bsdPolicyNames :: Lens' BackendServerDescription [Text]

The names of the policies enabled for the EC2 instance.

bsdInstancePort :: Lens' BackendServerDescription (Maybe Natural)

The port on which the EC2 instance is listening.

ConnectionDraining

data ConnectionDraining

Information about the ConnectionDraining attribute.

See: connectionDraining smart constructor.

connectionDraining

Arguments

:: Bool
cdEnabled

-> ConnectionDraining

Creates a value of ConnectionDraining with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

cdTimeout - The maximum time, in seconds, to keep the existing connections open before deregistering the instances.

cdEnabled - Specifies whether connection draining is enabled for the load balancer.

cdTimeout :: Lens' ConnectionDraining (Maybe Int)

The maximum time, in seconds, to keep the existing connections open before deregistering the instances.

cdEnabled :: Lens' ConnectionDraining Bool

Specifies whether connection draining is enabled for the load balancer.

ConnectionSettings

data ConnectionSettings

Information about the ConnectionSettings attribute.

See: connectionSettings smart constructor.

connectionSettings

Arguments

:: Natural
csIdleTimeout

-> ConnectionSettings

Creates a value of ConnectionSettings with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

csIdleTimeout - The time, in seconds, that the connection is allowed to be idle (no data has been sent over the connection) before it is closed by the load balancer.

csIdleTimeout :: Lens' ConnectionSettings Natural

The time, in seconds, that the connection is allowed to be idle (no data has been sent over the connection) before it is closed by the load balancer.

CrossZoneLoadBalancing

data CrossZoneLoadBalancing

Information about the CrossZoneLoadBalancing attribute.

See: crossZoneLoadBalancing smart constructor.

crossZoneLoadBalancing

Arguments

:: Bool
czlbEnabled

-> CrossZoneLoadBalancing

Creates a value of CrossZoneLoadBalancing with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

czlbEnabled - Specifies whether cross-zone load balancing is enabled for the load balancer.

czlbEnabled :: Lens' CrossZoneLoadBalancing Bool

Specifies whether cross-zone load balancing is enabled for the load balancer.

HealthCheck

data HealthCheck

Information about a health check.

See: healthCheck smart constructor.

healthCheck

Arguments

:: Text
hcTarget

-> Natural
hcInterval

-> Natural
hcTimeout

-> Natural
hcUnhealthyThreshold

-> Natural
hcHealthyThreshold

-> HealthCheck

Creates a value of HealthCheck with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

hcTarget - The instance being checked. The protocol is either TCP, HTTP, HTTPS, or SSL. The range of valid ports is one (1) through 65535. TCP is the default, specified as a TCP: port pair, for example TCP:5000. In this case, a health check simply attempts to open a TCP connection to the instance on the specified port. Failure to connect within the configured timeout is considered unhealthy. SSL is also specified as SSL: port pair, for example, SSL:5000. For HTTPHTTPS, you must include a ping path in the string. HTTP is specified as a HTTP:port;;PathToPing; grouping, for example "HTTP:80weatheruswaseattle". In this case, a HTTP GET request is issued to the instance on the given port and path. Any answer other than "200 OK" within the timeout period is considered unhealthy. The total length of the HTTP ping target must be 1024 16-bit Unicode characters or less.

hcInterval - The approximate interval, in seconds, between health checks of an individual instance.

hcTimeout - The amount of time, in seconds, during which no response means a failed health check. This value must be less than the Interval value.

hcUnhealthyThreshold - The number of consecutive health check failures required before moving the instance to the Unhealthy state.

hcHealthyThreshold - The number of consecutive health checks successes required before moving the instance to the Healthy state.

hcTarget :: Lens' HealthCheck Text

The instance being checked. The protocol is either TCP, HTTP, HTTPS, or SSL. The range of valid ports is one (1) through 65535. TCP is the default, specified as a TCP: port pair, for example TCP:5000. In this case, a health check simply attempts to open a TCP connection to the instance on the specified port. Failure to connect within the configured timeout is considered unhealthy. SSL is also specified as SSL: port pair, for example, SSL:5000. For HTTPHTTPS, you must include a ping path in the string. HTTP is specified as a HTTP:port;;PathToPing; grouping, for example "HTTP:80weatheruswaseattle". In this case, a HTTP GET request is issued to the instance on the given port and path. Any answer other than "200 OK" within the timeout period is considered unhealthy. The total length of the HTTP ping target must be 1024 16-bit Unicode characters or less.

hcInterval :: Lens' HealthCheck Natural

The approximate interval, in seconds, between health checks of an individual instance.

hcTimeout :: Lens' HealthCheck Natural

The amount of time, in seconds, during which no response means a failed health check. This value must be less than the Interval value.

hcUnhealthyThreshold :: Lens' HealthCheck Natural

The number of consecutive health check failures required before moving the instance to the Unhealthy state.

hcHealthyThreshold :: Lens' HealthCheck Natural

The number of consecutive health checks successes required before moving the instance to the Healthy state.

Instance

data Instance

The ID of an EC2 instance.

See: instance' smart constructor.

instance' :: Instance

Creates a value of Instance with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

iInstanceId - The instance ID.

iInstanceId :: Lens' Instance (Maybe Text)

The instance ID.

InstanceState

data InstanceState

Information about the state of an EC2 instance.

See: instanceState smart constructor.

instanceState :: InstanceState

Creates a value of InstanceState with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

isInstanceId - The ID of the instance.

isState - The current state of the instance. Valid values: InService | OutOfService | Unknown

isReasonCode - Information about the cause of OutOfService instances. Specifically, whether the cause is Elastic Load Balancing or the instance. Valid values: ELB | Instance | N/A

isDescription - A description of the instance state. This string can contain one or more of the following messages. * N/A * A transient error occurred. Please try again later. * Instance has failed at least the UnhealthyThreshold number of health checks consecutively. * Instance has not passed the configured HealthyThreshold number of health checks consecutively. * Instance registration is still in progress. * Instance is in the EC2 Availability Zone for which LoadBalancer is not configured to route traffic to. * Instance is not currently registered with the LoadBalancer. * Instance deregistration currently in progress. * Disable Availability Zone is currently in progress. * Instance is in pending state. * Instance is in stopped state. * Instance is in terminated state.

isInstanceId :: Lens' InstanceState (Maybe Text)

The ID of the instance.

isState :: Lens' InstanceState (Maybe Text)

The current state of the instance. Valid values: InService | OutOfService | Unknown

isReasonCode :: Lens' InstanceState (Maybe Text)

Information about the cause of OutOfService instances. Specifically, whether the cause is Elastic Load Balancing or the instance. Valid values: ELB | Instance | N/A

isDescription :: Lens' InstanceState (Maybe Text)

A description of the instance state. This string can contain one or more of the following messages. * N/A * A transient error occurred. Please try again later. * Instance has failed at least the UnhealthyThreshold number of health checks consecutively. * Instance has not passed the configured HealthyThreshold number of health checks consecutively. * Instance registration is still in progress. * Instance is in the EC2 Availability Zone for which LoadBalancer is not configured to route traffic to. * Instance is not currently registered with the LoadBalancer. * Instance deregistration currently in progress. * Disable Availability Zone is currently in progress. * Instance is in pending state. * Instance is in stopped state. * Instance is in terminated state.

LBCookieStickinessPolicy

data LBCookieStickinessPolicy

Information about a policy for duration-based session stickiness.

See: lBCookieStickinessPolicy smart constructor.

lBCookieStickinessPolicy :: LBCookieStickinessPolicy

Creates a value of LBCookieStickinessPolicy with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

lbcspPolicyName - The name of the policy. This name must be unique within the set of policies for this load balancer.

lbcspCookieExpirationPeriod - The time period, in seconds, after which the cookie should be considered stale. If this parameter is not specified, the stickiness session lasts for the duration of the browser session.

lbcspPolicyName :: Lens' LBCookieStickinessPolicy (Maybe Text)

The name of the policy. This name must be unique within the set of policies for this load balancer.

lbcspCookieExpirationPeriod :: Lens' LBCookieStickinessPolicy (Maybe Integer)

The time period, in seconds, after which the cookie should be considered stale. If this parameter is not specified, the stickiness session lasts for the duration of the browser session.

Limit

data Limit

Information about an Elastic Load Balancing resource limit for your AWS account.

See: limit smart constructor.

limit :: Limit

Creates a value of Limit with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

lMax - The maximum value of the limit.

lName - The name of the limit. The possible values are: * classic-listeners * classic-load-balancers

lMax :: Lens' Limit (Maybe Text)

The maximum value of the limit.

lName :: Lens' Limit (Maybe Text)

The name of the limit. The possible values are: * classic-listeners * classic-load-balancers

Listener

data Listener

Information about a listener.

For information about the protocols and the ports supported by Elastic Load Balancing, see Listeners for Your Classic Load Balancer in the Classic Load Balancer Guide .

See: listener smart constructor.

listener

Arguments

:: Text
lProtocol

-> Int
lLoadBalancerPort

-> Natural
lInstancePort

-> Listener

Creates a value of Listener with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

lInstanceProtocol - The protocol to use for routing traffic to instances: HTTP, HTTPS, TCP, or SSL. If the front-end protocol is HTTP, HTTPS, TCP, or SSL, InstanceProtocol must be at the same protocol. If there is another listener with the same InstancePort whose InstanceProtocol is secure, (HTTPS or SSL), the listener's InstanceProtocol must also be secure. If there is another listener with the same InstancePort whose InstanceProtocol is HTTP or TCP, the listener's InstanceProtocol must be HTTP or TCP.

lSSLCertificateId - The Amazon Resource Name (ARN) of the server certificate.

lProtocol - The load balancer transport protocol to use for routing: HTTP, HTTPS, TCP, or SSL.

lLoadBalancerPort - The port on which the load balancer is listening. On EC2-VPC, you can specify any port from the range 1-65535. On EC2-Classic, you can specify any port from the following list: 25, 80, 443, 465, 587, 1024-65535.

lInstancePort - The port on which the instance is listening.

lInstanceProtocol :: Lens' Listener (Maybe Text)

The protocol to use for routing traffic to instances: HTTP, HTTPS, TCP, or SSL. If the front-end protocol is HTTP, HTTPS, TCP, or SSL, InstanceProtocol must be at the same protocol. If there is another listener with the same InstancePort whose InstanceProtocol is secure, (HTTPS or SSL), the listener's InstanceProtocol must also be secure. If there is another listener with the same InstancePort whose InstanceProtocol is HTTP or TCP, the listener's InstanceProtocol must be HTTP or TCP.

lSSLCertificateId :: Lens' Listener (Maybe Text)

The Amazon Resource Name (ARN) of the server certificate.

lProtocol :: Lens' Listener Text

The load balancer transport protocol to use for routing: HTTP, HTTPS, TCP, or SSL.

lLoadBalancerPort :: Lens' Listener Int

The port on which the load balancer is listening. On EC2-VPC, you can specify any port from the range 1-65535. On EC2-Classic, you can specify any port from the following list: 25, 80, 443, 465, 587, 1024-65535.

lInstancePort :: Lens' Listener Natural

The port on which the instance is listening.

ListenerDescription

data ListenerDescription

The policies enabled for a listener.

See: listenerDescription smart constructor.

listenerDescription :: ListenerDescription

Creates a value of ListenerDescription with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

ldPolicyNames - The policies. If there are no policies enabled, the list is empty.

ldListener - The listener.

ldPolicyNames :: Lens' ListenerDescription [Text]

The policies. If there are no policies enabled, the list is empty.

ldListener :: Lens' ListenerDescription (Maybe Listener)

The listener.

LoadBalancerAttributes

data LoadBalancerAttributes

The attributes for a load balancer.

See: loadBalancerAttributes smart constructor.

loadBalancerAttributes :: LoadBalancerAttributes

Creates a value of LoadBalancerAttributes with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

lbaCrossZoneLoadBalancing - If enabled, the load balancer routes the request traffic evenly across all instances regardless of the Availability Zones. For more information, see Configure Cross-Zone Load Balancing in the Classic Load Balancer Guide .

lbaAccessLog - If enabled, the load balancer captures detailed information of all requests and delivers the information to the Amazon S3 bucket that you specify. For more information, see Enable Access Logs in the Classic Load Balancer Guide .

lbaAdditionalAttributes - This parameter is reserved.

lbaConnectionSettings - If enabled, the load balancer allows the connections to remain idle (no data is sent over the connection) for the specified duration. By default, Elastic Load Balancing maintains a 60-second idle connection timeout for both front-end and back-end connections of your load balancer. For more information, see Configure Idle Connection Timeout in the Classic Load Balancer Guide .

lbaConnectionDraining - If enabled, the load balancer allows existing requests to complete before the load balancer shifts traffic away from a deregistered or unhealthy instance. For more information, see Configure Connection Draining in the Classic Load Balancer Guide .

lbaCrossZoneLoadBalancing :: Lens' LoadBalancerAttributes (Maybe CrossZoneLoadBalancing)

If enabled, the load balancer routes the request traffic evenly across all instances regardless of the Availability Zones. For more information, see Configure Cross-Zone Load Balancing in the Classic Load Balancer Guide .

lbaAccessLog :: Lens' LoadBalancerAttributes (Maybe AccessLog)

If enabled, the load balancer captures detailed information of all requests and delivers the information to the Amazon S3 bucket that you specify. For more information, see Enable Access Logs in the Classic Load Balancer Guide .

lbaAdditionalAttributes :: Lens' LoadBalancerAttributes [AdditionalAttribute]

This parameter is reserved.

lbaConnectionSettings :: Lens' LoadBalancerAttributes (Maybe ConnectionSettings)

If enabled, the load balancer allows the connections to remain idle (no data is sent over the connection) for the specified duration. By default, Elastic Load Balancing maintains a 60-second idle connection timeout for both front-end and back-end connections of your load balancer. For more information, see Configure Idle Connection Timeout in the Classic Load Balancer Guide .

lbaConnectionDraining :: Lens' LoadBalancerAttributes (Maybe ConnectionDraining)

If enabled, the load balancer allows existing requests to complete before the load balancer shifts traffic away from a deregistered or unhealthy instance. For more information, see Configure Connection Draining in the Classic Load Balancer Guide .

LoadBalancerDescription

data LoadBalancerDescription

Information about a load balancer.

See: loadBalancerDescription smart constructor.

loadBalancerDescription :: LoadBalancerDescription

Creates a value of LoadBalancerDescription with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

lbdSourceSecurityGroup - The security group for the load balancer, which you can use as part of your inbound rules for your registered instances. To only allow traffic from load balancers, add a security group rule that specifies this source security group as the inbound source.

lbdCanonicalHostedZoneName - The DNS name of the load balancer. For more information, see Configure a Custom Domain Name in the Classic Load Balancer Guide .

lbdSecurityGroups - The security groups for the load balancer. Valid only for load balancers in a VPC.

lbdHealthCheck - Information about the health checks conducted on the load balancer.

lbdLoadBalancerName - The name of the load balancer.

lbdCreatedTime - The date and time the load balancer was created.

lbdVPCId - The ID of the VPC for the load balancer.

lbdSubnets - The IDs of the subnets for the load balancer.

lbdAvailabilityZones - The Availability Zones for the load balancer.

lbdBackendServerDescriptions - Information about your EC2 instances.

lbdCanonicalHostedZoneNameId - The ID of the Amazon Route 53 hosted zone for the load balancer.

lbdInstances - The IDs of the instances for the load balancer.

lbdScheme - The type of load balancer. Valid only for load balancers in a VPC. If Scheme is internet-facing , the load balancer has a public DNS name that resolves to a public IP address. If Scheme is internal , the load balancer has a public DNS name that resolves to a private IP address.

lbdListenerDescriptions - The listeners for the load balancer.

lbdDNSName - The DNS name of the load balancer.

lbdPolicies - The policies defined for the load balancer.

lbdSourceSecurityGroup :: Lens' LoadBalancerDescription (Maybe SourceSecurityGroup)

The security group for the load balancer, which you can use as part of your inbound rules for your registered instances. To only allow traffic from load balancers, add a security group rule that specifies this source security group as the inbound source.

lbdCanonicalHostedZoneName :: Lens' LoadBalancerDescription (Maybe Text)

The DNS name of the load balancer. For more information, see Configure a Custom Domain Name in the Classic Load Balancer Guide .

lbdSecurityGroups :: Lens' LoadBalancerDescription [Text]

The security groups for the load balancer. Valid only for load balancers in a VPC.

lbdHealthCheck :: Lens' LoadBalancerDescription (Maybe HealthCheck)

Information about the health checks conducted on the load balancer.

lbdLoadBalancerName :: Lens' LoadBalancerDescription (Maybe Text)

The name of the load balancer.

lbdCreatedTime :: Lens' LoadBalancerDescription (Maybe UTCTime)

The date and time the load balancer was created.

lbdVPCId :: Lens' LoadBalancerDescription (Maybe Text)

The ID of the VPC for the load balancer.

lbdSubnets :: Lens' LoadBalancerDescription [Text]

The IDs of the subnets for the load balancer.

lbdAvailabilityZones :: Lens' LoadBalancerDescription [Text]

The Availability Zones for the load balancer.

lbdBackendServerDescriptions :: Lens' LoadBalancerDescription [BackendServerDescription]

Information about your EC2 instances.

lbdCanonicalHostedZoneNameId :: Lens' LoadBalancerDescription (Maybe Text)

The ID of the Amazon Route 53 hosted zone for the load balancer.

lbdInstances :: Lens' LoadBalancerDescription [Instance]

The IDs of the instances for the load balancer.

lbdScheme :: Lens' LoadBalancerDescription (Maybe Text)

The type of load balancer. Valid only for load balancers in a VPC. If Scheme is internet-facing , the load balancer has a public DNS name that resolves to a public IP address. If Scheme is internal , the load balancer has a public DNS name that resolves to a private IP address.

lbdListenerDescriptions :: Lens' LoadBalancerDescription [ListenerDescription]

The listeners for the load balancer.

lbdDNSName :: Lens' LoadBalancerDescription (Maybe Text)

The DNS name of the load balancer.

lbdPolicies :: Lens' LoadBalancerDescription (Maybe Policies)

The policies defined for the load balancer.

Policies

data Policies

The policies for a load balancer.

See: policies smart constructor.

policies :: Policies

Creates a value of Policies with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

pOtherPolicies - The policies other than the stickiness policies.

pLBCookieStickinessPolicies - The stickiness policies created using CreateLBCookieStickinessPolicy .

pAppCookieStickinessPolicies - The stickiness policies created using CreateAppCookieStickinessPolicy .

pOtherPolicies :: Lens' Policies [Text]

The policies other than the stickiness policies.

pLBCookieStickinessPolicies :: Lens' Policies [LBCookieStickinessPolicy]

The stickiness policies created using CreateLBCookieStickinessPolicy .

pAppCookieStickinessPolicies :: Lens' Policies [AppCookieStickinessPolicy]

The stickiness policies created using CreateAppCookieStickinessPolicy .

PolicyAttribute

data PolicyAttribute

Information about a policy attribute.

See: policyAttribute smart constructor.

policyAttribute :: PolicyAttribute

Creates a value of PolicyAttribute with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

paAttributeValue - The value of the attribute.

paAttributeName - The name of the attribute.

paAttributeValue :: Lens' PolicyAttribute (Maybe Text)

The value of the attribute.

paAttributeName :: Lens' PolicyAttribute (Maybe Text)

The name of the attribute.

PolicyAttributeDescription

data PolicyAttributeDescription

Information about a policy attribute.

See: policyAttributeDescription smart constructor.

policyAttributeDescription :: PolicyAttributeDescription

Creates a value of PolicyAttributeDescription with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

padAttributeValue - The value of the attribute.

padAttributeName - The name of the attribute.

padAttributeValue :: Lens' PolicyAttributeDescription (Maybe Text)

The value of the attribute.

padAttributeName :: Lens' PolicyAttributeDescription (Maybe Text)

The name of the attribute.

PolicyAttributeTypeDescription

data PolicyAttributeTypeDescription

Information about a policy attribute type.

See: policyAttributeTypeDescription smart constructor.

policyAttributeTypeDescription :: PolicyAttributeTypeDescription

Creates a value of PolicyAttributeTypeDescription with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

patdAttributeType - The type of the attribute. For example, Boolean or Integer .

patdCardinality - The cardinality of the attribute. Valid values: * ONE(1) : Single value required * ZERO_OR_ONE(0..1) : Up to one value is allowed * ZERO_OR_MORE(0..*) : Optional. Multiple values are allowed * ONE_OR_MORE(1..*0) : Required. Multiple values are allowed

patdDefaultValue - The default value of the attribute, if applicable.

patdAttributeName - The name of the attribute.

patdDescription - A description of the attribute.

patdAttributeType :: Lens' PolicyAttributeTypeDescription (Maybe Text)

The type of the attribute. For example, Boolean or Integer .

patdCardinality :: Lens' PolicyAttributeTypeDescription (Maybe Text)

The cardinality of the attribute. Valid values: * ONE(1) : Single value required * ZERO_OR_ONE(0..1) : Up to one value is allowed * ZERO_OR_MORE(0..*) : Optional. Multiple values are allowed * ONE_OR_MORE(1..*0) : Required. Multiple values are allowed

patdDefaultValue :: Lens' PolicyAttributeTypeDescription (Maybe Text)

The default value of the attribute, if applicable.

patdAttributeName :: Lens' PolicyAttributeTypeDescription (Maybe Text)

The name of the attribute.

patdDescription :: Lens' PolicyAttributeTypeDescription (Maybe Text)

A description of the attribute.

PolicyDescription

data PolicyDescription

Information about a policy.

See: policyDescription smart constructor.

policyDescription :: PolicyDescription

Creates a value of PolicyDescription with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

pdPolicyName - The name of the policy.

pdPolicyAttributeDescriptions - The policy attributes.

pdPolicyTypeName - The name of the policy type.

pdPolicyName :: Lens' PolicyDescription (Maybe Text)

The name of the policy.

pdPolicyAttributeDescriptions :: Lens' PolicyDescription [PolicyAttributeDescription]

The policy attributes.

pdPolicyTypeName :: Lens' PolicyDescription (Maybe Text)

The name of the policy type.

PolicyTypeDescription

data PolicyTypeDescription

Information about a policy type.

See: policyTypeDescription smart constructor.

policyTypeDescription :: PolicyTypeDescription

Creates a value of PolicyTypeDescription with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

ptdPolicyTypeName - The name of the policy type.

ptdDescription - A description of the policy type.

ptdPolicyAttributeTypeDescriptions - The description of the policy attributes associated with the policies defined by Elastic Load Balancing.

ptdPolicyTypeName :: Lens' PolicyTypeDescription (Maybe Text)

The name of the policy type.

ptdDescription :: Lens' PolicyTypeDescription (Maybe Text)

A description of the policy type.

ptdPolicyAttributeTypeDescriptions :: Lens' PolicyTypeDescription [PolicyAttributeTypeDescription]

The description of the policy attributes associated with the policies defined by Elastic Load Balancing.

SourceSecurityGroup

data SourceSecurityGroup

Information about a source security group.

See: sourceSecurityGroup smart constructor.

sourceSecurityGroup :: SourceSecurityGroup

Creates a value of SourceSecurityGroup with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

ssgOwnerAlias - The owner of the security group.

ssgGroupName - The name of the security group.

ssgOwnerAlias :: Lens' SourceSecurityGroup (Maybe Text)

The owner of the security group.

ssgGroupName :: Lens' SourceSecurityGroup (Maybe Text)

The name of the security group.

Tag

data Tag

Information about a tag.

See: tag smart constructor.

tag

Arguments

:: Text
tagKey

-> Tag

Creates a value of Tag with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

tagValue - The value of the tag.

tagKey - The key of the tag.

tagValue :: Lens' Tag (Maybe Text)

The value of the tag.

tagKey :: Lens' Tag Text

The key of the tag.

TagDescription

data TagDescription

The tags associated with a load balancer.

See: tagDescription smart constructor.

tagDescription :: TagDescription

Creates a value of TagDescription with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

tdLoadBalancerName - The name of the load balancer.

tdTags - The tags.

tdLoadBalancerName :: Lens' TagDescription (Maybe Text)

The name of the load balancer.

tdTags :: Lens' TagDescription (Maybe (NonEmpty Tag))

The tags.

TagKeyOnly

data TagKeyOnly

The key of a tag.

See: tagKeyOnly smart constructor.

tagKeyOnly :: TagKeyOnly

Creates a value of TagKeyOnly with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

tkoKey - The name of the key.

tkoKey :: Lens' TagKeyOnly (Maybe Text)

The name of the key.