Network.AWS.ELB.Types

Service

data ELB

Version 2012-06-01 of the Amazon Elastic Load Balancing service.

Error

data RESTError :: *

XML

ns :: Text

SourceSecurityGroup

data SourceSecurityGroup

sourceSecurityGroup :: SourceSecurityGroup

SourceSecurityGroup constructor.

The fields accessible through corresponding lenses are:

ssgGroupName :: Maybe Text

ssgOwnerAlias :: Maybe Text

ssgGroupName :: Lens' SourceSecurityGroup (Maybe Text)

Name of the source security group. Use this value for the '--source-group' parameter of the 'ec2-authorize' command in the Amazon EC2 command line tool.

ssgOwnerAlias :: Lens' SourceSecurityGroup (Maybe Text)

Owner of the source security group. Use this value for the '--source-group-user' parameter of the 'ec2-authorize' command in the Amazon EC2 command line tool.

TagDescription

data TagDescription

tagDescription

Arguments

:: NonEmpty Tag
tdTags

-> TagDescription

TagDescription constructor.

The fields accessible through corresponding lenses are:

tdLoadBalancerName :: Maybe Text

tdTags :: NonEmpty Tag

tdLoadBalancerName :: Lens' TagDescription (Maybe Text)

The name of the load balancer.

tdTags :: Lens' TagDescription (NonEmpty Tag)

List of tags associated with the load balancer.

Tag

data Tag

tag

Arguments

:: Text
tagKey

-> Tag

Tag constructor.

The fields accessible through corresponding lenses are:

tagKey :: Text

tagValue :: Maybe Text

tagKey :: Lens' Tag Text

The key of the tag.

tagValue :: Lens' Tag (Maybe Text)

The value of the tag.

PolicyAttributeTypeDescription

data PolicyAttributeTypeDescription

policyAttributeTypeDescription :: PolicyAttributeTypeDescription

PolicyAttributeTypeDescription constructor.

The fields accessible through corresponding lenses are:

patdAttributeName :: Maybe Text

patdAttributeType :: Maybe Text

patdCardinality :: Maybe Text

patdDefaultValue :: Maybe Text

patdDescription :: Maybe Text

patdAttributeName :: Lens' PolicyAttributeTypeDescription (Maybe Text)

The name of the attribute associated with the policy type.

patdAttributeType :: Lens' PolicyAttributeTypeDescription (Maybe Text)

The type of attribute. For example, Boolean, Integer, etc.

patdCardinality :: Lens' PolicyAttributeTypeDescription (Maybe Text)

The cardinality of the attribute. Valid Values: ONE(1) : Single value required ZERO_OR_ONE(0..1) : Up to one value can be supplied ZERO_OR_MORE(0..*) : Optional. Multiple values are allowed ONE_OR_MORE(1..*0) : Required. Multiple values are allowed

patdDefaultValue :: Lens' PolicyAttributeTypeDescription (Maybe Text)

The default value of the attribute, if applicable.

patdDescription :: Lens' PolicyAttributeTypeDescription (Maybe Text)

A human-readable description of the attribute.

HealthCheck

data HealthCheck

healthCheck

Arguments

:: Text
hcTarget

-> Natural
hcInterval

-> Natural
hcTimeout

-> Natural
hcUnhealthyThreshold

-> Natural
hcHealthyThreshold

-> HealthCheck

HealthCheck constructor.

The fields accessible through corresponding lenses are:

hcHealthyThreshold :: Natural

hcInterval :: Natural

hcTarget :: Text

hcTimeout :: Natural

hcUnhealthyThreshold :: Natural

hcHealthyThreshold :: Lens' HealthCheck Natural

Specifies the number of consecutive health probe successes required before moving the instance to the Healthy state.

hcInterval :: Lens' HealthCheck Natural

Specifies the approximate interval, in seconds, between health checks of an individual instance.

hcTarget :: Lens' HealthCheck Text

Specifies the instance being checked. The protocol is either TCP, HTTP, HTTPS, or SSL. The range of valid ports is one (1) through 65535.

TCP is the default, specified as a TCP: port pair, for example TCP:5000. In this case a healthcheck simply attempts to open a TCP connection to the instance on the specified port. Failure to connect within the configured timeout is considered unhealthy.

SSL is also specified as SSL: port pair, for example, SSL:5000.

For HTTP or HTTPS protocol, the situation is different. You have to include a ping path in the string. HTTP is specified as a HTTP:port;/;PathToPing; grouping, for example "HTTP:80weatheruswaseattle". In this case, a HTTP GET request is issued to the instance on the given port and path. Any answer other than "200 OK" within the timeout period is considered unhealthy.

The total length of the HTTP ping target needs to be 1024 16-bit Unicode characters or less.

hcTimeout :: Lens' HealthCheck Natural

Specifies the amount of time, in seconds, during which no response means a failed health probe.

This value must be less than the Interval value.

hcUnhealthyThreshold :: Lens' HealthCheck Natural

Specifies the number of consecutive health probe failures required before moving the instance to the Unhealthy state.

CrossZoneLoadBalancing

data CrossZoneLoadBalancing

crossZoneLoadBalancing

Arguments

:: Bool
czlbEnabled

-> CrossZoneLoadBalancing

CrossZoneLoadBalancing constructor.

The fields accessible through corresponding lenses are:

czlbEnabled :: Bool

czlbEnabled :: Lens' CrossZoneLoadBalancing Bool

Specifies whether cross-zone load balancing is enabled for the load balancer.

LoadBalancerAttributes

data LoadBalancerAttributes

loadBalancerAttributes :: LoadBalancerAttributes

LoadBalancerAttributes constructor.

The fields accessible through corresponding lenses are:

lbaAccessLog :: Maybe AccessLog

lbaAdditionalAttributes :: [AdditionalAttribute]

lbaConnectionDraining :: Maybe ConnectionDraining

lbaConnectionSettings :: Maybe ConnectionSettings

lbaCrossZoneLoadBalancing :: Maybe CrossZoneLoadBalancing

lbaAccessLog :: Lens' LoadBalancerAttributes (Maybe AccessLog)

The name of the load balancer attribute. If enabled, the load balancer captures detailed information of all the requests and delivers the information to the Amazon S3 bucket that you specify.

For more information, see Enable Access Logs.

lbaAdditionalAttributes :: Lens' LoadBalancerAttributes [AdditionalAttribute]

lbaConnectionDraining :: Lens' LoadBalancerAttributes (Maybe ConnectionDraining)

The name of the load balancer attribute. If enabled, the load balancer allows existing requests to complete before the load balancer shifts traffic away from a deregistered or unhealthy back-end instance.

For more information, see Enable Connection Draining.

lbaConnectionSettings :: Lens' LoadBalancerAttributes (Maybe ConnectionSettings)

The name of the load balancer attribute.

By default, Elastic Load Balancing maintains a 60-second idle connection timeout for both front-end and back-end connections of your load balancer. If the ConnectionSettings attribute is set, Elastic Load Balancing will allow the connections to remain idle (no data is sent over the connection) for the specified duration.

For more information, see Configure Idle Connection Timeout.

lbaCrossZoneLoadBalancing :: Lens' LoadBalancerAttributes (Maybe CrossZoneLoadBalancing)

The name of the load balancer attribute. If enabled, the load balancer routes the request traffic evenly across all back-end instances regardless of the Availability Zones.

For more information, see Enable Cross-Zone Load Balancing.

AccessLog

data AccessLog

accessLog

Arguments

:: Bool
alEnabled

-> AccessLog

AccessLog constructor.

The fields accessible through corresponding lenses are:

alEmitInterval :: Maybe Int

alEnabled :: Bool

alS3BucketName :: Maybe Text

alS3BucketPrefix :: Maybe Text

alEmitInterval :: Lens' AccessLog (Maybe Int)

The interval for publishing the access logs. You can specify an interval of either 5 minutes or 60 minutes.

Default: 60 minutes

alEnabled :: Lens' AccessLog Bool

Specifies whether access log is enabled for the load balancer.

alS3BucketName :: Lens' AccessLog (Maybe Text)

The name of the Amazon S3 bucket where the access logs are stored.

alS3BucketPrefix :: Lens' AccessLog (Maybe Text)

The logical hierarchy you created for your Amazon S3 bucket, for example 'my-bucket-prefix/prod'. If the prefix is not provided, the log is placed at the root level of the bucket.

ListenerDescription

data ListenerDescription

listenerDescription :: ListenerDescription

ListenerDescription constructor.

The fields accessible through corresponding lenses are:

ldListener :: Maybe Listener

ldPolicyNames :: [Text]

ldListener :: Lens' ListenerDescription (Maybe Listener)

ldPolicyNames :: Lens' ListenerDescription [Text]

A list of policies enabled for this listener. An empty list indicates that no policies are enabled.

LBCookieStickinessPolicy

data LBCookieStickinessPolicy

lbcookieStickinessPolicy :: LBCookieStickinessPolicy

LBCookieStickinessPolicy constructor.

The fields accessible through corresponding lenses are:

lbcspCookieExpirationPeriod :: Maybe Integer

lbcspPolicyName :: Maybe Text

lbcspCookieExpirationPeriod :: Lens' LBCookieStickinessPolicy (Maybe Integer)

The time period in seconds after which the cookie should be considered stale. Not specifying this parameter indicates that the stickiness session will last for the duration of the browser session.

lbcspPolicyName :: Lens' LBCookieStickinessPolicy (Maybe Text)

The name for the policy being created. The name must be unique within the set of policies for this load balancer.

PolicyDescription

data PolicyDescription

policyDescription :: PolicyDescription

PolicyDescription constructor.

The fields accessible through corresponding lenses are:

pdPolicyAttributeDescriptions :: [PolicyAttributeDescription]

pdPolicyName :: Maybe Text

pdPolicyTypeName :: Maybe Text

pdPolicyAttributeDescriptions :: Lens' PolicyDescription [PolicyAttributeDescription]

A list of policy attribute description structures.

pdPolicyName :: Lens' PolicyDescription (Maybe Text)

The name of the policy associated with the load balancer.

pdPolicyTypeName :: Lens' PolicyDescription (Maybe Text)

The name of the policy type associated with the load balancer.

AppCookieStickinessPolicy

data AppCookieStickinessPolicy

appCookieStickinessPolicy :: AppCookieStickinessPolicy

AppCookieStickinessPolicy constructor.

The fields accessible through corresponding lenses are:

acspCookieName :: Maybe Text

acspPolicyName :: Maybe Text

acspCookieName :: Lens' AppCookieStickinessPolicy (Maybe Text)

The name of the application cookie used for stickiness.

acspPolicyName :: Lens' AppCookieStickinessPolicy (Maybe Text)

The mnemonic name for the policy being created. The name must be unique within a set of policies for this load balancer.

PolicyAttribute

data PolicyAttribute

policyAttribute :: PolicyAttribute

PolicyAttribute constructor.

The fields accessible through corresponding lenses are:

paAttributeName :: Maybe Text

paAttributeValue :: Maybe Text

paAttributeName :: Lens' PolicyAttribute (Maybe Text)

The name of the attribute associated with the policy.

paAttributeValue :: Lens' PolicyAttribute (Maybe Text)

The value of the attribute associated with the policy.

LoadBalancerDescription

data LoadBalancerDescription

loadBalancerDescription :: LoadBalancerDescription

LoadBalancerDescription constructor.

The fields accessible through corresponding lenses are:

lbdAvailabilityZones :: [Text]

lbdBackendServerDescriptions :: [BackendServerDescription]

lbdCanonicalHostedZoneName :: Maybe Text

lbdCanonicalHostedZoneNameID :: Maybe Text

lbdCreatedTime :: Maybe UTCTime

lbdDNSName :: Maybe Text

lbdHealthCheck :: Maybe HealthCheck

lbdInstances :: [Instance]

lbdListenerDescriptions :: [ListenerDescription]

lbdLoadBalancerName :: Maybe Text

lbdPolicies :: Maybe Policies

lbdScheme :: Maybe Text

lbdSecurityGroups :: [Text]

lbdSourceSecurityGroup :: Maybe SourceSecurityGroup

lbdSubnets :: [Text]

lbdVPCId :: Maybe Text

lbdAvailabilityZones :: Lens' LoadBalancerDescription [Text]

Specifies a list of Availability Zones.

lbdBackendServerDescriptions :: Lens' LoadBalancerDescription [BackendServerDescription]

Contains a list of back-end server descriptions.

lbdCanonicalHostedZoneName :: Lens' LoadBalancerDescription (Maybe Text)

Provides the name of the Amazon Route 53 hosted zone that is associated with the load balancer. For information on how to associate your load balancer with a hosted zone, go to Using Domain Names With Elastic Load Balancing in the Elastic Load Balancing Developer Guide.

lbdCanonicalHostedZoneNameID :: Lens' LoadBalancerDescription (Maybe Text)

Provides the ID of the Amazon Route 53 hosted zone name that is associated with the load balancer. For information on how to associate or disassociate your load balancer with a hosted zone, go to Using Domain Names With ElasticLoad Balancing in the Elastic Load Balancing Developer Guide.

lbdCreatedTime :: Lens' LoadBalancerDescription (Maybe UTCTime)

Provides the date and time the load balancer was created.

lbdDNSName :: Lens' LoadBalancerDescription (Maybe Text)

Specifies the external DNS name associated with the load balancer.

lbdHealthCheck :: Lens' LoadBalancerDescription (Maybe HealthCheck)

Specifies information regarding the various health probes conducted on the load balancer.

lbdInstances :: Lens' LoadBalancerDescription [Instance]

Provides a list of EC2 instance IDs for the load balancer.

lbdListenerDescriptions :: Lens' LoadBalancerDescription [ListenerDescription]

LoadBalancerPort, InstancePort, Protocol, InstanceProtocol, and PolicyNames are returned in a list of tuples in the ListenerDescriptions element.

lbdLoadBalancerName :: Lens' LoadBalancerDescription (Maybe Text)

Specifies the name associated with the load balancer.

lbdPolicies :: Lens' LoadBalancerDescription (Maybe Policies)

Provides a list of policies defined for the load balancer.

lbdScheme :: Lens' LoadBalancerDescription (Maybe Text)

Specifies the type of load balancer.

If the Scheme is 'internet-facing', the load balancer has a publicly resolvable DNS name that resolves to public IP addresses.

If the Scheme is internal, the load balancer has a publicly resolvable DNS name that resolves to private IP addresses.

This option is only available for load balancers attached to an Amazon VPC.

lbdSecurityGroups :: Lens' LoadBalancerDescription [Text]

The security groups the load balancer is a member of (VPC only).

lbdSourceSecurityGroup :: Lens' LoadBalancerDescription (Maybe SourceSecurityGroup)

The security group that you can use as part of your inbound rules for your load balancer's back-end Amazon EC2 application instances. To only allow traffic from load balancers, add a security group rule to your back end instance that specifies this source security group as the inbound source.

lbdSubnets :: Lens' LoadBalancerDescription [Text]

Provides a list of VPC subnet IDs for the load balancer.

lbdVPCId :: Lens' LoadBalancerDescription (Maybe Text)

Provides the ID of the VPC attached to the load balancer.

BackendServerDescription

data BackendServerDescription

backendServerDescription :: BackendServerDescription

BackendServerDescription constructor.

The fields accessible through corresponding lenses are:

bsdInstancePort :: Maybe Natural

bsdPolicyNames :: [Text]

bsdInstancePort :: Lens' BackendServerDescription (Maybe Natural)

Provides the port on which the back-end server is listening.

bsdPolicyNames :: Lens' BackendServerDescription [Text]

Provides a list of policy names enabled for the back-end server.

PolicyAttributeDescription

data PolicyAttributeDescription

policyAttributeDescription :: PolicyAttributeDescription

PolicyAttributeDescription constructor.

The fields accessible through corresponding lenses are:

padAttributeName :: Maybe Text

padAttributeValue :: Maybe Text

padAttributeName :: Lens' PolicyAttributeDescription (Maybe Text)

The name of the attribute associated with the policy.

padAttributeValue :: Lens' PolicyAttributeDescription (Maybe Text)

The value of the attribute associated with the policy.

AdditionalAttribute

data AdditionalAttribute

additionalAttribute :: AdditionalAttribute

AdditionalAttribute constructor.

The fields accessible through corresponding lenses are:

aaKey :: Maybe Text

aaValue :: Maybe Text

aaKey :: Lens' AdditionalAttribute (Maybe Text)

aaValue :: Lens' AdditionalAttribute (Maybe Text)

ConnectionSettings

data ConnectionSettings

connectionSettings

Arguments

:: Natural
csIdleTimeout

-> ConnectionSettings

ConnectionSettings constructor.

The fields accessible through corresponding lenses are:

csIdleTimeout :: Natural

csIdleTimeout :: Lens' ConnectionSettings Natural

Specifies the time (in seconds) the connection is allowed to be idle (no data has been sent over the connection) before it is closed by the load balancer.

PolicyTypeDescription

data PolicyTypeDescription

policyTypeDescription :: PolicyTypeDescription

PolicyTypeDescription constructor.

The fields accessible through corresponding lenses are:

ptdDescription :: Maybe Text

ptdPolicyAttributeTypeDescriptions :: [PolicyAttributeTypeDescription]

ptdPolicyTypeName :: Maybe Text

ptdDescription :: Lens' PolicyTypeDescription (Maybe Text)

A human-readable description of the policy type.

ptdPolicyAttributeTypeDescriptions :: Lens' PolicyTypeDescription [PolicyAttributeTypeDescription]

The description of the policy attributes associated with the load balancer policies defined by the Elastic Load Balancing service.

ptdPolicyTypeName :: Lens' PolicyTypeDescription (Maybe Text)

The name of the policy type.

Policies

data Policies

policies :: Policies

Policies constructor.

The fields accessible through corresponding lenses are:

pAppCookieStickinessPolicies :: [AppCookieStickinessPolicy]

pLBCookieStickinessPolicies :: [LBCookieStickinessPolicy]

pOtherPolicies :: [Text]

pAppCookieStickinessPolicies :: Lens' Policies [AppCookieStickinessPolicy]

A list of the AppCookieStickinessPolicy objects created with CreateAppCookieStickinessPolicy.

pLBCookieStickinessPolicies :: Lens' Policies [LBCookieStickinessPolicy]

A list of LBCookieStickinessPolicy objects created with CreateAppCookieStickinessPolicy.

pOtherPolicies :: Lens' Policies [Text]

A list of policy names other than the stickiness policies.

Listener

data Listener

listener

Arguments

:: Text
lProtocol

-> Int
lLoadBalancerPort

-> Natural
lInstancePort

-> Listener

Listener constructor.

The fields accessible through corresponding lenses are:

lInstancePort :: Natural

lInstanceProtocol :: Maybe Text

lLoadBalancerPort :: Int

lProtocol :: Text

lSSLCertificateId :: Maybe Text

lInstancePort :: Lens' Listener Natural

Specifies the TCP port on which the instance server is listening. This property cannot be modified for the life of the load balancer.

lInstanceProtocol :: Lens' Listener (Maybe Text)

Specifies the protocol to use for routing traffic to back-end instances - HTTP, HTTPS, TCP, or SSL. This property cannot be modified for the life of the load balancer.

If the front-end protocol is HTTP or HTTPS, InstanceProtocol has to be at the same protocol layer, i.e., HTTP or HTTPS. Likewise, if the front-end protocol is TCP or SSL, InstanceProtocol has to be TCP or SSL. If there is another listener with the same InstancePort whose InstanceProtocol is secure, i.e., HTTPS or SSL, the listener's InstanceProtocol has to be secure, i.e., HTTPS or SSL. If there is another listener with the same InstancePort whose InstanceProtocol is HTTP or TCP, the listener's InstanceProtocol must be either HTTP or TCP.

lLoadBalancerPort :: Lens' Listener Int

Specifies the external load balancer port number. This property cannot be modified for the life of the load balancer.

lProtocol :: Lens' Listener Text

Specifies the load balancer transport protocol to use for routing - HTTP, HTTPS, TCP or SSL. This property cannot be modified for the life of the load balancer.

lSSLCertificateId :: Lens' Listener (Maybe Text)

The ARN string of the server certificate. To get the ARN of the server certificate, call the AWS Identity and Access Management UploadServerCertificate API.

ConnectionDraining

data ConnectionDraining

connectionDraining

Arguments

:: Bool
cdEnabled

-> ConnectionDraining

ConnectionDraining constructor.

The fields accessible through corresponding lenses are:

cdEnabled :: Bool

cdTimeout :: Maybe Int

cdEnabled :: Lens' ConnectionDraining Bool

Specifies whether connection draining is enabled for the load balancer.

cdTimeout :: Lens' ConnectionDraining (Maybe Int)

Specifies the maximum time (in seconds) to keep the existing connections open before deregistering the instances.

InstanceState

data InstanceState

instanceState :: InstanceState

InstanceState constructor.

The fields accessible through corresponding lenses are:

isDescription :: Maybe Text

isInstanceId :: Maybe Text

isReasonCode :: Maybe Text

isState :: Maybe Text

isDescription :: Lens' InstanceState (Maybe Text)

Provides a description of the instance state.

isInstanceId :: Lens' InstanceState (Maybe Text)

Provides an EC2 instance ID.

isReasonCode :: Lens' InstanceState (Maybe Text)

Provides information about the cause of OutOfService instances. Specifically, it indicates whether the cause is Elastic Load Balancing or the instance behind the load balancer.

Valid value: 'ELB'|'Instance'|'N/A'

isState :: Lens' InstanceState (Maybe Text)

Specifies the current state of the instance.

Valid value: 'InService'|'OutOfService'|'Unknown'

TagKeyOnly

data TagKeyOnly

tagKeyOnly :: TagKeyOnly

TagKeyOnly constructor.

The fields accessible through corresponding lenses are:

tkoKey :: Maybe Text

tkoKey :: Lens' TagKeyOnly (Maybe Text)

The name of the key.

Instance

data Instance

instance' :: Instance

Instance constructor.

The fields accessible through corresponding lenses are:

iInstanceId :: Maybe Text

iInstanceId :: Lens' Instance (Maybe Text)

Provides an EC2 instance ID.