Network.AWS.ELB.V2012_06_01.ConfigureHealthCheck

Description

Specifies the health check settings to use for evaluating the health state of your back-end instances. For more information, see Health Check in the Elastic Load Balancing Developer Guide. https:/elasticloadbalancing.amazonaws.com?HealthCheck.HealthyThreshold=2 &HealthCheck.UnhealthyThreshold=2 &LoadBalancerName=MyLoadBalancer &HealthCheck.Target=HTTP:80/ping &HealthCheck.Interval=30 &HealthCheck.Timeout=3 &Version=2012-06-01 &Action=ConfigureHealthCheck &AUTHPARAMS 30 HTTP:80/ping 2 3 2 83c88b9d-12b7-11e3-8b82-87b12EXAMPLE.

Request

data ConfigureHealthCheck

Input for the ConfigureHealthCheck action.

Request constructor

mkConfigureHealthCheck

Arguments

:: Text
chcLoadBalancerName

-> HealthCheck
chcHealthCheck

-> ConfigureHealthCheck

Smart constructor for the minimum required parameters to construct a valid ConfigureHealthCheck request.

The fields accessible through corresponding lenses are:

LoadBalancerName :: Text

HealthCheck :: HealthCheck

Request lenses

chcLoadBalancerName :: Lens' ConfigureHealthCheck Text

The mnemonic name associated with the load balancer. The name must be unique within the set of load balancers associated with your AWS account.

chcHealthCheck :: Lens' ConfigureHealthCheck HealthCheck

A structure containing the configuration information for the new healthcheck.

Response

data ConfigureHealthCheckResponse

The output for the ConfigureHealthCheck action.

Response constructor

mkConfigureHealthCheckResponse :: ConfigureHealthCheckResponse

Smart constructor for the minimum required parameters to construct a valid ConfigureHealthCheckResponse response.

This constructor is provided for convenience and testing purposes.

The fields accessible through corresponding lenses are:

HealthCheck :: Maybe HealthCheck

Response lenses

chcrHealthCheck :: Lens' ConfigureHealthCheckResponse (Maybe HealthCheck)

The updated healthcheck for the instances.