Network.AWS.ELB.V2012_06_01.DescribeLoadBalancerPolicyTypes

Description

Returns meta-information on the specified load balancer policies defined by the Elastic Load Balancing service. The policy types that are returned from this action can be used in a CreateLoadBalancerPolicy action to instantiate specific policy configurations that will be applied to a load balancer. Partial description of all the policy types defined by Elastic Load Balancing for your account https:/elasticloadbalancing.amazonaws.com?Version=2012-06-01 &Action=DescribeLoadBalancerPolicyTypes &AUTHPARAMS SSLNegotiationPolicyType BackendServerAuthenticationPolicyType PublicKeyPolicyType AppCookieStickinessPolicyType LBCookieStickinessPolicyType ProxyProtocolPolicyType 83c88b9d-12b7-11e3-8b82-87b12EXAMPLE Description of ProxyProtocolPolicyType https:/elasticloadbalancing.amazonaws.com?PolicyTypeNames.member.1=ProxyProtocolPolicyType &Version=2012-06-01 &Action=DescribeLoadBalancerPolicyTypes &AUTHPARAMS ProxyProtocol Boolean ONE ProxyProtocolPolicyType Policy that controls whether to include the IP address and port of the originating request for TCP messages. This policy operates on TCP/SSL listeners only 1549581b-12b7-11e3-895e-1334aEXAMPLE.

Request

data DescribeLoadBalancerPolicyTypes

Request constructor

mkDescribeLoadBalancerPolicyTypes :: DescribeLoadBalancerPolicyTypes

Smart constructor for the minimum required parameters to construct a valid DescribeLoadBalancerPolicyTypes request.

The fields accessible through corresponding lenses are:

PolicyTypeNames :: [Text]

Request lenses

dlbptPolicyTypeNames :: Lens' DescribeLoadBalancerPolicyTypes [Text]

Specifies the name of the policy types. If no names are specified, returns the description of all the policy types defined by Elastic Load Balancing service.

Response

data DescribeLoadBalancerPolicyTypesResponse

The output for the DescribeLoadBalancerPolicyTypes action.

Response constructor

mkDescribeLoadBalancerPolicyTypesResponse :: DescribeLoadBalancerPolicyTypesResponse

Smart constructor for the minimum required parameters to construct a valid DescribeLoadBalancerPolicyTypesResponse response.

This constructor is provided for convenience and testing purposes.

The fields accessible through corresponding lenses are:

PolicyTypeDescriptions :: [PolicyTypeDescription]

Response lenses

dlbptrPolicyTypeDescriptions :: Lens' DescribeLoadBalancerPolicyTypesResponse [PolicyTypeDescription]

List of policy type description structures of the specified policy type. If no policy type names are specified, returns the description of all the policy types defined by Elastic Load Balancing service.