Network.AWS.ELB.V2012_06_01.DetachLoadBalancerFromSubnets

Description

Removes subnets from the set of configured subnets in the Amazon Virtual Private Cloud (Amazon VPC) for the load balancer. After a subnet is removed all of the EC2 instances registered with the load balancer that are in the removed subnet will go into the OutOfService state. When a subnet is removed, the load balancer will balance the traffic among the remaining routable subnets for the load balancer. https:/elasticloadbalancing.amazonaws.com?Subnets.member.1=subnet-119f0078 &LoadBalancerName=my-test-vpc-loadbalancer &Version=2012-06-01 &Action=DetachLoadBalancerFromSubnets &AUTHPARAMS subnet-159f007c subnet-3561b05e 07b1ecbc-1100-11e3-acaf-dd7edEXAMPLE.

Request

data DetachLoadBalancerFromSubnets

The input for the DetachLoadBalancerFromSubnets action.

Request constructor

mkDetachLoadBalancerFromSubnets

Arguments

:: Text
dlbfsLoadBalancerName

-> [Text]
dlbfsSubnets

-> DetachLoadBalancerFromSubnets

Smart constructor for the minimum required parameters to construct a valid DetachLoadBalancerFromSubnets request.

Request lenses

dlbfsLoadBalancerName :: Lens' DetachLoadBalancerFromSubnets Text

The name associated with the load balancer to be detached.

dlbfsSubnets :: Lens' DetachLoadBalancerFromSubnets [Text]

A list of subnet IDs to remove from the set of configured subnets for the load balancer.

Response

data DetachLoadBalancerFromSubnetsResponse

The output for the DetachLoadBalancerFromSubnets action.

Response constructor

mkDetachLoadBalancerFromSubnetsResponse :: DetachLoadBalancerFromSubnetsResponse

Smart constructor for the minimum required parameters to construct a valid DetachLoadBalancerFromSubnetsResponse response.

This constructor is provided for convenience and testing purposes.

Response lenses

dlbfsrSubnets :: Lens' DetachLoadBalancerFromSubnetsResponse [Text]

A list of subnet IDs the load balancer is now attached to.