Network.AWS.ELB.V2012_06_01.SetLoadBalancerListenerSSLCertificate

Description

Sets the certificate that terminates the specified listener's SSL connections. The specified certificate replaces any prior certificate that was used on the same load balancer and port. For more information on updating your SSL certificate, see Updating an SSL Certificate for a Load Balancer in the Elastic Load Balancing Developer Guide. https:/elasticloadbalancing.amazonaws.com?LoadBalancerName=MyInternalLoadBalancer &SSLCertificateId=arn:aws:iam::123456789012:server-certificate/testcert &LoadBalancerPort=443 &Version=2012-06-01 &Action=SetLoadBalancerListenerSSLCertificate &AUTHPARAMS 83c88b9d-12b7-11e3-8b82-87b12EXAMPLE.

Request

data SetLoadBalancerListenerSSLCertificate

The input for the SetLoadBalancerListenerSSLCertificate action.

Request constructor

mkSetLoadBalancerListenerSSLCertificate

Arguments

:: Text
slblsslcLoadBalancerName

-> Integer
slblsslcLoadBalancerPort

-> Text
slblsslcSSLCertificateId

-> SetLoadBalancerListenerSSLCertificate

Smart constructor for the minimum required parameters to construct a valid SetLoadBalancerListenerSSLCertificate request.

Request lenses

slblsslcLoadBalancerName :: Lens' SetLoadBalancerListenerSSLCertificate Text

The name of the load balancer.

slblsslcLoadBalancerPort :: Lens' SetLoadBalancerListenerSSLCertificate Integer

The port that uses the specified SSL certificate.

slblsslcSSLCertificateId :: Lens' SetLoadBalancerListenerSSLCertificate Text

The Amazon Resource Number (ARN) of the SSL certificate chain to use. For more information on SSL certificates, see Managing Server Certificates in the AWS Identity and Access Management User Guide.

Response

data SetLoadBalancerListenerSSLCertificateResponse

The output for the SetLoadBalancerListenerSSLCertificate action.

Response constructor

mkSetLoadBalancerListenerSSLCertificateResponse :: SetLoadBalancerListenerSSLCertificateResponse

Smart constructor for the minimum required parameters to construct a valid SetLoadBalancerListenerSSLCertificateResponse response.

This constructor is provided for convenience and testing purposes.